% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_associate_backup_vault_mpa_approval_team}
\alias{backup_associate_backup_vault_mpa_approval_team}
\title{Associates an MPA approval team with a backup vault}
\usage{
backup_associate_backup_vault_mpa_approval_team(
  BackupVaultName,
  MpaApprovalTeamArn,
  RequesterComment = NULL
)
}
\arguments{
\item{BackupVaultName}{[required] The name of the backup vault to associate with the MPA approval team.}

\item{MpaApprovalTeamArn}{[required] The Amazon Resource Name (ARN) of the MPA approval team to associate
with the backup vault.}

\item{RequesterComment}{A comment provided by the requester explaining the association request.}
}
\description{
Associates an MPA approval team with a backup vault.

See \url{https://www.paws-r-sdk.com/docs/backup_associate_backup_vault_mpa_approval_team/} for full documentation.
}
\keyword{internal}
