% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_cancel_legal_hold}
\alias{backup_cancel_legal_hold}
\title{This action removes the specified legal hold on a recovery point}
\usage{
backup_cancel_legal_hold(
  LegalHoldId,
  CancelDescription,
  RetainRecordInDays = NULL
)
}
\arguments{
\item{LegalHoldId}{[required] Legal hold ID required to remove the specified legal hold on a recovery
point.}

\item{CancelDescription}{[required] String describing the reason for removing the legal hold.}

\item{RetainRecordInDays}{The integer amount in days specifying amount of days after this API
operation to remove legal hold.}
}
\description{
This action removes the specified legal hold on a recovery point. This action can only be performed by a user with sufficient permissions.

See \url{https://paws-r.github.io/docs/backup/cancel_legal_hold.html} for full documentation.
}
\keyword{internal}
