% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_create_restore_testing_selection}
\alias{backup_create_restore_testing_selection}
\title{This request can be sent after CreateRestoreTestingPlan request returns
successfully}
\usage{
backup_create_restore_testing_selection(
  CreatorRequestId = NULL,
  RestoreTestingPlanName,
  RestoreTestingSelection
)
}
\arguments{
\item{CreatorRequestId}{This is an optional unique string that identifies the request and allows
failed requests to be retried without the risk of running the operation
twice. If used, this parameter must contain 1 to 50 alphanumeric or
'-_.' characters.}

\item{RestoreTestingPlanName}{[required] Input the restore testing plan name that was returned from the related
CreateRestoreTestingPlan request.}

\item{RestoreTestingSelection}{[required] This consists of \code{RestoreTestingSelectionName}, \code{ProtectedResourceType},
and one of the following:
\itemize{
\item \code{ProtectedResourceArns}
\item \code{ProtectedResourceConditions}
}

Each protected resource type can have one single value.

A restore testing selection can include a wildcard value ("*") for
\code{ProtectedResourceArns} along with \code{ProtectedResourceConditions}.
Alternatively, you can include up to 30 specific protected resource ARNs
in \code{ProtectedResourceArns}.}
}
\description{
This request can be sent after CreateRestoreTestingPlan request returns successfully. This is the second part of creating a resource testing plan, and it must be completed sequentially.

See \url{https://www.paws-r-sdk.com/docs/backup_create_restore_testing_selection/} for full documentation.
}
\keyword{internal}
