% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_create_tiering_configuration}
\alias{backup_create_tiering_configuration}
\title{Creates a tiering configuration}
\usage{
backup_create_tiering_configuration(
  TieringConfiguration,
  TieringConfigurationTags = NULL,
  CreatorRequestId = NULL
)
}
\arguments{
\item{TieringConfiguration}{[required] A tiering configuration must contain a unique \code{TieringConfigurationName}
string you create and must contain a \code{BackupVaultName} and
\code{ResourceSelection}. You may optionally include a \code{CreatorRequestId}
string.

The \code{TieringConfigurationName} is a unique string that is the name of
the tiering configuration. This cannot be changed after creation, and it
must consist of only alphanumeric characters and underscores.}

\item{TieringConfigurationTags}{The tags to assign to the tiering configuration.}

\item{CreatorRequestId}{This is a unique string that identifies the request and allows failed
requests to be retried without the risk of running the operation twice.
This parameter is optional. If used, this parameter must contain 1 to 50
alphanumeric or '-_.' characters.}
}
\description{
Creates a tiering configuration.

See \url{https://www.paws-r-sdk.com/docs/backup_create_tiering_configuration/} for full documentation.
}
\keyword{internal}
