% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_backup_access_point}
\alias{backup_delete_backup_access_point}
\title{Deletes a backup access point}
\usage{
backup_delete_backup_access_point(AccessPointArn)
}
\arguments{
\item{AccessPointArn}{[required] The Amazon Resource Name (ARN) of the backup access point to delete.}
}
\description{
Deletes a backup access point. This deletes the underlying Amazon S3 access point and, if no other backup access points remain for the recovery point, resumes lifecycle transitions for that recovery point.

See \url{https://www.paws-r-sdk.com/docs/backup_delete_backup_access_point/} for full documentation.
}
\keyword{internal}
