% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_backup_selection}
\alias{backup_delete_backup_selection}
\title{Deletes the resource selection associated with a backup plan that is
specified by the SelectionId}
\usage{
backup_delete_backup_selection(BackupPlanId, SelectionId)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}

\item{SelectionId}{[required] Uniquely identifies the body of a request to assign a set of resources
to a backup plan.}
}
\value{
An empty list.
}
\description{
Deletes the resource selection associated with a backup plan that is
specified by the \code{SelectionId}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_backup_selection(
  BackupPlanId = "string",
  SelectionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
