% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_backup_vault_access_policy}
\alias{backup_delete_backup_vault_access_policy}
\title{Deletes the policy document that manages permissions on a backup vault}
\usage{
backup_delete_backup_vault_access_policy(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the Amazon Web Services Region where they are created. They
consist of lowercase letters, numbers, and hyphens.}
}
\description{
Deletes the policy document that manages permissions on a backup vault.

See \url{https://paws-r.github.io/docs/backup/delete_backup_vault_access_policy.html} for full documentation.
}
\keyword{internal}
