% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_backup_vault_lock_configuration}
\alias{backup_delete_backup_vault_lock_configuration}
\title{Deletes Backup Vault Lock from a backup vault specified by a backup
vault name}
\usage{
backup_delete_backup_vault_lock_configuration(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of the backup vault from which to delete Backup Vault Lock.}
}
\description{
Deletes Backup Vault Lock from a backup vault specified by a backup vault name.

See \url{https://paws-r.github.io/docs/backup/delete_backup_vault_lock_configuration.html} for full documentation.
}
\keyword{internal}
