% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_backup_vault_notifications}
\alias{backup_delete_backup_vault_notifications}
\title{Deletes event notifications for the specified backup vault}
\usage{
backup_delete_backup_vault_notifications(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the Region where they are created. They consist of lowercase
letters, numbers, and hyphens.}
}
\value{
An empty list.
}
\description{
Deletes event notifications for the specified backup vault.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_backup_vault_notifications(
  BackupVaultName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
