% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_framework}
\alias{backup_delete_framework}
\title{Deletes the framework specified by a framework name}
\usage{
backup_delete_framework(FrameworkName)
}
\arguments{
\item{FrameworkName}{[required] The unique name of a framework.}
}
\description{
Deletes the framework specified by a framework name.

See \url{https://www.paws-r-sdk.com/docs/backup_delete_framework/} for full documentation.
}
\keyword{internal}
