% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_report_plan}
\alias{backup_delete_report_plan}
\title{Deletes the report plan specified by a report plan name}
\usage{
backup_delete_report_plan(ReportPlanName)
}
\arguments{
\item{ReportPlanName}{[required] The unique name of a report plan.}
}
\description{
Deletes the report plan specified by a report plan name.

See \url{https://www.paws-r-sdk.com/docs/backup_delete_report_plan/} for full documentation.
}
\keyword{internal}
