% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_restore_testing_plan}
\alias{backup_delete_restore_testing_plan}
\title{This request deletes the specified restore testing plan}
\usage{
backup_delete_restore_testing_plan(RestoreTestingPlanName)
}
\arguments{
\item{RestoreTestingPlanName}{[required] Required unique name of the restore testing plan you wish to delete.}
}
\description{
This request deletes the specified restore testing plan.

See \url{https://www.paws-r-sdk.com/docs/backup_delete_restore_testing_plan/} for full documentation.
}
\keyword{internal}
