% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_restore_testing_selection}
\alias{backup_delete_restore_testing_selection}
\title{Input the Restore Testing Plan name and Restore Testing Selection name}
\usage{
backup_delete_restore_testing_selection(
  RestoreTestingPlanName,
  RestoreTestingSelectionName
)
}
\arguments{
\item{RestoreTestingPlanName}{[required] Required unique name of the restore testing plan that contains the
restore testing selection you wish to delete.}

\item{RestoreTestingSelectionName}{[required] Required unique name of the restore testing selection you wish to
delete.}
}
\description{
Input the Restore Testing Plan name and Restore Testing Selection name.

See \url{https://www.paws-r-sdk.com/docs/backup_delete_restore_testing_selection/} for full documentation.
}
\keyword{internal}
