% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_tiering_configuration}
\alias{backup_delete_tiering_configuration}
\title{Deletes the tiering configuration specified by a tiering configuration
name}
\usage{
backup_delete_tiering_configuration(TieringConfigurationName)
}
\arguments{
\item{TieringConfigurationName}{[required] The unique name of a tiering configuration.}
}
\description{
Deletes the tiering configuration specified by a tiering configuration name.

See \url{https://www.paws-r-sdk.com/docs/backup_delete_tiering_configuration/} for full documentation.
}
\keyword{internal}
