% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_backup_access_point}
\alias{backup_describe_backup_access_point}
\title{Returns metadata about a backup access point, including its status and
the details of the underlying Amazon S3 access point}
\usage{
backup_describe_backup_access_point(AccessPointArn)
}
\arguments{
\item{AccessPointArn}{[required] The Amazon Resource Name (ARN) of the backup access point to describe.}
}
\description{
Returns metadata about a backup access point, including its status and the details of the underlying Amazon S3 access point.

See \url{https://www.paws-r-sdk.com/docs/backup_describe_backup_access_point/} for full documentation.
}
\keyword{internal}
