% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_backup_vault}
\alias{backup_describe_backup_vault}
\title{Returns metadata about a backup vault specified by its name}
\usage{
backup_describe_backup_vault(BackupVaultName, BackupVaultAccountId = NULL)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the Amazon Web Services Region where they are created. They
consist of lowercase letters, numbers, and hyphens.}

\item{BackupVaultAccountId}{This is the account ID of the specified backup vault.}
}
\description{
Returns metadata about a backup vault specified by its name.

See \url{https://www.paws-r-sdk.com/docs/backup_describe_backup_vault/} for full documentation.
}
\keyword{internal}
