% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_global_settings}
\alias{backup_describe_global_settings}
\title{The current feature settings for the AWS Account}
\usage{
backup_describe_global_settings()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GlobalSettings = list(
    "string"
  ),
  LastUpdateTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
The current feature settings for the AWS Account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_global_settings()
}\if{html}{\out{</div>}}
}

\keyword{internal}
