% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_region_settings}
\alias{backup_describe_region_settings}
\title{Returns the current service opt-in settings for the Region}
\usage{
backup_describe_region_settings()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceTypeOptInPreference = list(
    TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the current service opt-in settings for the Region. If
service-opt-in is enabled for a service, AWS Backup tries to protect
that service's resources in this Region, when the resource is included
in an on-demand backup or scheduled backup plan. Otherwise, AWS Backup
does not try to protect that service's resources in this Region, AWS
Backup does not try to protect that service's resources in this Region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_region_settings()
}\if{html}{\out{</div>}}
}

\keyword{internal}
