% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_report_job}
\alias{backup_describe_report_job}
\title{Returns the details associated with creating a report as specified by
its ReportJobId}
\usage{
backup_describe_report_job(ReportJobId)
}
\arguments{
\item{ReportJobId}{[required] The identifier of the report job. A unique, randomly generated, Unicode,
UTF-8 encoded string that is at most 1,024 bytes long. The report job ID
cannot be edited.}
}
\description{
Returns the details associated with creating a report as specified by its \code{ReportJobId}.

See \url{https://paws-r.github.io/docs/backup/describe_report_job.html} for full documentation.
}
\keyword{internal}
