% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_report_plan}
\alias{backup_describe_report_plan}
\title{Returns a list of all report plans for an Amazon Web Services account
and Amazon Web Services Region}
\usage{
backup_describe_report_plan(ReportPlanName)
}
\arguments{
\item{ReportPlanName}{[required] The unique name of a report plan.}
}
\description{
Returns a list of all report plans for an Amazon Web Services account and Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/backup_describe_report_plan/} for full documentation.
}
\keyword{internal}
