% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_scan_job}
\alias{backup_describe_scan_job}
\title{Returns scan job details for the specified ScanJobID}
\usage{
backup_describe_scan_job(ScanJobId)
}
\arguments{
\item{ScanJobId}{[required] Uniquely identifies a request to Backup to scan a resource.}
}
\description{
Returns scan job details for the specified ScanJobID.

See \url{https://www.paws-r-sdk.com/docs/backup_describe_scan_job/} for full documentation.
}
\keyword{internal}
