% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_disassociate_backup_vault_mpa_approval_team}
\alias{backup_disassociate_backup_vault_mpa_approval_team}
\title{Removes the association between an MPA approval team and a backup vault,
disabling the MPA approval workflow for restore operations}
\usage{
backup_disassociate_backup_vault_mpa_approval_team(
  BackupVaultName,
  RequesterComment = NULL
)
}
\arguments{
\item{BackupVaultName}{[required] The name of the backup vault from which to disassociate the MPA approval
team.}

\item{RequesterComment}{An optional comment explaining the reason for disassociating the MPA
approval team from the backup vault.}
}
\description{
Removes the association between an MPA approval team and a backup vault, disabling the MPA approval workflow for restore operations.

See \url{https://www.paws-r-sdk.com/docs/backup_disassociate_backup_vault_mpa_approval_team/} for full documentation.
}
\keyword{internal}
