% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_disassociate_recovery_point}
\alias{backup_disassociate_recovery_point}
\title{Deletes the specified continuous backup recovery point from AWS Backup
and releases control of that continuous backup to the source service,
such as Amazon RDS}
\usage{
backup_disassociate_recovery_point(BackupVaultName, RecoveryPointArn)
}
\arguments{
\item{BackupVaultName}{[required] The unique name of an AWS Backup vault. Required.}

\item{RecoveryPointArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies an AWS Backup
recovery point. Required.}
}
\value{
An empty list.
}
\description{
Deletes the specified continuous backup recovery point from AWS Backup
and releases control of that continuous backup to the source service,
such as Amazon RDS. The source service will continue to create and
retain continuous backups using the lifecycle that you specified in your
original backup plan.

Does not support snapshot backup recovery points.
}
\section{Request syntax}{
\preformatted{svc$disassociate_recovery_point(
  BackupVaultName = "string",
  RecoveryPointArn = "string"
)
}
}

\keyword{internal}
