% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_disassociate_recovery_point_from_parent}
\alias{backup_disassociate_recovery_point_from_parent}
\title{This action to a specific child (nested) recovery point removes the
relationship between the specified recovery point and its parent
(composite) recovery point}
\usage{
backup_disassociate_recovery_point_from_parent(
  BackupVaultName,
  RecoveryPointArn
)
}
\arguments{
\item{BackupVaultName}{[required] This is the name of a logical container where the child (nested)
recovery point is stored. Backup vaults are identified by names that are
unique to the account used to create them and the Amazon Web Services
Region where they are created. They consist of lowercase letters,
numbers, and hyphens.}

\item{RecoveryPointArn}{[required] This is the Amazon Resource Name (ARN) that uniquely identifies the
child (nested) recovery point; for example,
\verb{arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.}}
}
\description{
This action to a specific child (nested) recovery point removes the relationship between the specified recovery point and its parent (composite) recovery point.

See \url{https://www.paws-r-sdk.com/docs/backup_disassociate_recovery_point_from_parent/} for full documentation.
}
\keyword{internal}
