% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_export_backup_plan_template}
\alias{backup_export_backup_plan_template}
\title{Returns the backup plan that is specified by the plan ID as a backup
template}
\usage{
backup_export_backup_plan_template(BackupPlanId)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BackupPlanTemplateJson = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns the backup plan that is specified by the plan ID as a backup
template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$export_backup_plan_template(
  BackupPlanId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
