% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_backup_plan_from_json}
\alias{backup_get_backup_plan_from_json}
\title{Returns a valid JSON document specifying a backup plan or an error}
\usage{
backup_get_backup_plan_from_json(BackupPlanTemplateJson)
}
\arguments{
\item{BackupPlanTemplateJson}{[required] A customer-supplied backup plan document in JSON format.}
}
\description{
Returns a valid JSON document specifying a backup plan or an error.

See \url{https://www.paws-r-sdk.com/docs/backup_get_backup_plan_from_json/} for full documentation.
}
\keyword{internal}
