% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_backup_plan_from_template}
\alias{backup_get_backup_plan_from_template}
\title{Returns the template specified by its templateId as a backup plan}
\usage{
backup_get_backup_plan_from_template(BackupPlanTemplateId)
}
\arguments{
\item{BackupPlanTemplateId}{[required] Uniquely identifies a stored backup plan template.}
}
\description{
Returns the template specified by its \code{templateId} as a backup plan.

See \url{https://paws-r.github.io/docs/backup/get_backup_plan_from_template.html} for full documentation.
}
\keyword{internal}
