% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_backup_vault_access_policy}
\alias{backup_get_backup_vault_access_policy}
\title{Returns the access policy document that is associated with the named
backup vault}
\usage{
backup_get_backup_vault_access_policy(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the Amazon Web Services Region where they are created. They
consist of lowercase letters, numbers, and hyphens.}
}
\description{
Returns the access policy document that is associated with the named backup vault.

See \url{https://www.paws-r-sdk.com/docs/backup_get_backup_vault_access_policy/} for full documentation.
}
\keyword{internal}
