% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_legal_hold}
\alias{backup_get_legal_hold}
\title{This action returns details for a specified legal hold}
\usage{
backup_get_legal_hold(LegalHoldId)
}
\arguments{
\item{LegalHoldId}{[required] The ID of the legal hold.}
}
\description{
This action returns details for a specified legal hold. The details are the body of a legal hold in JSON format, in addition to metadata.

See \url{https://www.paws-r-sdk.com/docs/backup_get_legal_hold/} for full documentation.
}
\keyword{internal}
