% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_pitr_malware_scan_results}
\alias{backup_get_pitr_malware_scan_results}
\title{Returns the malware scan results for a specified point in time within a
continuous (point-in-time recovery) backup}
\usage{
backup_get_pitr_malware_scan_results(
  RecoveryPointArn,
  BackupVaultName,
  ScanEndTime,
  MalwareScanner
)
}
\arguments{
\item{RecoveryPointArn}{[required] An ARN that uniquely identifies the target recovery point for scanning; for example, \verb{arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45}.}

\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created.}

\item{ScanEndTime}{[required] The point in time within the continuous backup to examine for malware scan results.}

\item{MalwareScanner}{[required] The scanning engine used for the corresponding scan job. Currently only \code{GUARDDUTY} is supported.}
}
\description{
Returns the malware scan results for a specified point in time within a continuous (point-in-time recovery) backup.

See \url{https://www.paws-r-sdk.com/docs/backup_get_pitr_malware_scan_results/} for full documentation.
}
\keyword{internal}
