% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_recovery_point_index_details}
\alias{backup_get_recovery_point_index_details}
\title{This operation returns the metadata and details specific to the backup
index associated with the specified recovery point}
\usage{
backup_get_recovery_point_index_details(BackupVaultName, RecoveryPointArn)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the Region where they are created.

Accepted characters include lowercase letters, numbers, and hyphens.}

\item{RecoveryPointArn}{[required] An ARN that uniquely identifies a recovery point; for example,
\verb{arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45}.}
}
\description{
This operation returns the metadata and details specific to the backup index associated with the specified recovery point.

See \url{https://www.paws-r-sdk.com/docs/backup_get_recovery_point_index_details/} for full documentation.
}
\keyword{internal}
