% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_restore_job_metadata}
\alias{backup_get_restore_job_metadata}
\title{This request returns the metadata for the specified restore job}
\usage{
backup_get_restore_job_metadata(RestoreJobId)
}
\arguments{
\item{RestoreJobId}{[required] This is a unique identifier of a restore job within Backup.}
}
\description{
This request returns the metadata for the specified restore job.

See \url{https://www.paws-r-sdk.com/docs/backup_get_restore_job_metadata/} for full documentation.
}
\keyword{internal}
