% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_restore_testing_selection}
\alias{backup_get_restore_testing_selection}
\title{Returns RestoreTestingSelection, which displays resources and elements
of the restore testing plan}
\usage{
backup_get_restore_testing_selection(
  RestoreTestingPlanName,
  RestoreTestingSelectionName
)
}
\arguments{
\item{RestoreTestingPlanName}{[required] Required unique name of the restore testing plan.}

\item{RestoreTestingSelectionName}{[required] Required unique name of the restore testing selection.}
}
\description{
Returns RestoreTestingSelection, which displays resources and elements of the restore testing plan.

See \url{https://www.paws-r-sdk.com/docs/backup_get_restore_testing_selection/} for full documentation.
}
\keyword{internal}
