% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_supported_resource_types}
\alias{backup_get_supported_resource_types}
\title{Returns the Amazon Web Services resource types supported by Backup}
\usage{
backup_get_supported_resource_types()
}
\description{
Returns the Amazon Web Services resource types supported by Backup.

See \url{https://www.paws-r-sdk.com/docs/backup_get_supported_resource_types/} for full documentation.
}
\keyword{internal}
