% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_tiering_configuration}
\alias{backup_get_tiering_configuration}
\title{Returns TieringConfiguration details for the specified
TieringConfigurationName}
\usage{
backup_get_tiering_configuration(TieringConfigurationName)
}
\arguments{
\item{TieringConfigurationName}{[required] The unique name of a tiering configuration.}
}
\description{
Returns \code{TieringConfiguration} details for the specified \code{TieringConfigurationName}. The details are the body of a tiering configuration in JSON format, in addition to configuration metadata.

See \url{https://www.paws-r-sdk.com/docs/backup_get_tiering_configuration/} for full documentation.
}
\keyword{internal}
