% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_backup_access_points_by_recovery_point}
\alias{backup_list_backup_access_points_by_recovery_point}
\title{Returns the backup access points associated with the specified recovery
point}
\usage{
backup_list_backup_access_points_by_recovery_point(
  MaxResults = NULL,
  NextToken = NULL,
  RecoveryPointArn
)
}
\arguments{
\item{MaxResults}{The maximum number of items to be returned.}

\item{NextToken}{The next item following a partial list of returned items. For example, if a request is made to return \code{MaxResults} number of items, \code{NextToken} allows you to return more items in your list starting at the location pointed to by the next token.}

\item{RecoveryPointArn}{[required] The Amazon Resource Name (ARN) of the recovery point whose backup access points you want to list.}
}
\description{
Returns the backup access points associated with the specified recovery point.

See \url{https://www.paws-r-sdk.com/docs/backup_list_backup_access_points_by_recovery_point/} for full documentation.
}
\keyword{internal}
