% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_backup_plan_templates}
\alias{backup_list_backup_plan_templates}
\title{Lists the backup plan templates}
\usage{
backup_list_backup_plan_templates(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The next item following a partial list of returned items. For example,
if a request is made to return \code{MaxResults} number of items, \code{NextToken}
allows you to return more items in your list starting at the location
pointed to by the next token.}

\item{MaxResults}{The maximum number of items to return.}
}
\description{
Lists the backup plan templates.

See \url{https://www.paws-r-sdk.com/docs/backup_list_backup_plan_templates/} for full documentation.
}
\keyword{internal}
