% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_frameworks}
\alias{backup_list_frameworks}
\title{Returns a list of all frameworks for an Amazon Web Services account and
Amazon Web Services Region}
\usage{
backup_list_frameworks(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of desired results from 1 to 1000. Optional. If unspecified,
the query will return 1 MB of data.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Returns a list of all frameworks for an Amazon Web Services account and Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/backup/list_frameworks.html} for full documentation.
}
\keyword{internal}
