% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_legal_holds}
\alias{backup_list_legal_holds}
\title{This action returns metadata about active and previous legal holds}
\usage{
backup_list_legal_holds(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The next item following a partial list of returned resources. For
example, if a request is made to return \code{maxResults} number of
resources, \code{NextToken} allows you to return more items in your list
starting at the location pointed to by the next token.}

\item{MaxResults}{The maximum number of resource list items to be returned.}
}
\description{
This action returns metadata about active and previous legal holds.

See \url{https://paws-r.github.io/docs/backup/list_legal_holds.html} for full documentation.
}
\keyword{internal}
