% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_protected_resources_by_backup_vault}
\alias{backup_list_protected_resources_by_backup_vault}
\title{This request lists the protected resources corresponding to each backup
vault}
\usage{
backup_list_protected_resources_by_backup_vault(
  BackupVaultName,
  BackupVaultAccountId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{BackupVaultName}{[required] This is the list of protected resources by backup vault within the
vault(s) you specify by name.}

\item{BackupVaultAccountId}{This is the list of protected resources by backup vault within the
vault(s) you specify by account ID.}

\item{NextToken}{The next item following a partial list of returned items. For example,
if a request is made to return \code{maxResults} number of items, \code{NextToken}
allows you to return more items in your list starting at the location
pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\description{
This request lists the protected resources corresponding to each backup vault.

See \url{https://www.paws-r-sdk.com/docs/backup_list_protected_resources_by_backup_vault/} for full documentation.
}
\keyword{internal}
