% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_recovery_points_by_legal_hold}
\alias{backup_list_recovery_points_by_legal_hold}
\title{This action returns recovery point ARNs (Amazon Resource Names) of the
specified legal hold}
\usage{
backup_list_recovery_points_by_legal_hold(
  LegalHoldId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{LegalHoldId}{[required] This is the ID of the legal hold.}

\item{NextToken}{This is the next item following a partial list of returned resources.
For example, if a request is made to return \code{maxResults} number of
resources, \code{NextToken} allows you to return more items in your list
starting at the location pointed to by the next token.}

\item{MaxResults}{This is the maximum number of resource list items to be returned.}
}
\description{
This action returns recovery point ARNs (Amazon Resource Names) of the specified legal hold.

See \url{https://paws-r.github.io/docs/backup/list_recovery_points_by_legal_hold.html} for full documentation.
}
\keyword{internal}
