% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_report_jobs}
\alias{backup_list_report_jobs}
\title{Returns details about your report jobs}
\usage{
backup_list_report_jobs(
  ByReportPlanName = NULL,
  ByCreationBefore = NULL,
  ByCreationAfter = NULL,
  ByStatus = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ByReportPlanName}{Returns only report jobs with the specified report plan name.}

\item{ByCreationBefore}{Returns only report jobs that were created before the date and time
specified in Unix format and Coordinated Universal Time (UTC). For
example, the value 1516925490 represents Friday, January 26, 2018
12:11:30 AM.}

\item{ByCreationAfter}{Returns only report jobs that were created after the date and time
specified in Unix format and Coordinated Universal Time (UTC). For
example, the value 1516925490 represents Friday, January 26, 2018
12:11:30 AM.}

\item{ByStatus}{Returns only report jobs that are in the specified status. The statuses
are:

\code{CREATED | RUNNING | COMPLETED | FAILED | COMPLETED_WITH_ISSUES}

Please note that only scanning jobs finish with state completed with
issues. For backup jobs this is a console interpretation of a job that
finishes in completed state and has a status message.}

\item{MaxResults}{The number of desired results from 1 to 1000. Optional. If unspecified,
the query will return 1 MB of data.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Returns details about your report jobs.

See \url{https://www.paws-r-sdk.com/docs/backup_list_report_jobs/} for full documentation.
}
\keyword{internal}
