% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_report_plans}
\alias{backup_list_report_plans}
\title{Returns a list of your report plans}
\usage{
backup_list_report_plans(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of desired results from 1 to 1000. Optional. If unspecified,
the query will return 1 MB of data.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Returns a list of your report plans. For detailed information about a single report plan, use \code{\link[=backup_describe_report_plan]{describe_report_plan}}.

See \url{https://www.paws-r-sdk.com/docs/backup_list_report_plans/} for full documentation.
}
\keyword{internal}
