% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_restore_access_backup_vaults}
\alias{backup_list_restore_access_backup_vaults}
\title{Returns a list of restore access backup vaults associated with a
specified backup vault}
\usage{
backup_list_restore_access_backup_vaults(
  BackupVaultName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{BackupVaultName}{[required] The name of the backup vault for which to list associated restore access
backup vaults.}

\item{NextToken}{The pagination token from a previous request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of items to return in the response.}
}
\description{
Returns a list of restore access backup vaults associated with a specified backup vault.

See \url{https://www.paws-r-sdk.com/docs/backup_list_restore_access_backup_vaults/} for full documentation.
}
\keyword{internal}
