% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_scan_job_summaries}
\alias{backup_list_scan_job_summaries}
\title{This is a request for a summary of scan jobs created or running within
the most recent 14 days}
\usage{
backup_list_scan_job_summaries(
  AccountId = NULL,
  ResourceType = NULL,
  MalwareScanner = NULL,
  ScanResultStatus = NULL,
  State = NULL,
  AggregationPeriod = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountId}{Returns the job count for the specified account.

If the request is sent from a member account or an account not part of Amazon Web Services Organizations, jobs within requestor's account will be returned.

Root, admin, and delegated administrator accounts can use the value \code{ANY} to return job counts from every account in the organization.

\code{AGGREGATE_ALL} aggregates job counts from all accounts within the authenticated organization, then returns the sum.}

\item{ResourceType}{Returns the job count for the specified resource type. Use request \code{\link[=backup_get_supported_resource_types]{get_supported_resource_types}} to obtain strings for supported resource types.

The the value \code{ANY} returns count of all resource types.

\code{AGGREGATE_ALL} aggregates job counts for all resource types and returns the sum.}

\item{MalwareScanner}{Returns only the scan jobs for the specified malware scanner. Currently the only MalwareScanner is \code{GUARDDUTY}. But the field also supports \code{ANY}, and \code{AGGREGATE_ALL}.}

\item{ScanResultStatus}{Returns only the scan jobs for the specified scan results.}

\item{State}{Returns only the scan jobs for the specified scanning job state.}

\item{AggregationPeriod}{The period for the returned results.
\itemize{
\item \code{ONE_DAY}The daily job count for the prior 1 day.
\item \code{SEVEN_DAYS}The daily job count for the prior 7 days.
\item \code{FOURTEEN_DAYS}The daily job count for the prior 14 days.
}}

\item{MaxResults}{The maximum number of items to be returned.

The value is an integer. Range of accepted values is from 1 to 500.}

\item{NextToken}{The next item following a partial list of returned items. For example, if a request is made to return \code{MaxResults} number of items, \code{NextToken} allows you to return more items in your list starting at the location pointed to by the next token.}
}
\description{
This is a request for a summary of scan jobs created or running within the most recent 14 days.

See \url{https://www.paws-r-sdk.com/docs/backup_list_scan_job_summaries/} for full documentation.
}
\keyword{internal}
