% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_put_restore_validation_result}
\alias{backup_put_restore_validation_result}
\title{This request allows you to send your independent self-run restore test
validation results}
\usage{
backup_put_restore_validation_result(
  RestoreJobId,
  ValidationStatus,
  ValidationStatusMessage = NULL
)
}
\arguments{
\item{RestoreJobId}{[required] This is a unique identifier of a restore job within Backup.}

\item{ValidationStatus}{[required] The status of your restore validation.}

\item{ValidationStatusMessage}{This is an optional message string you can input to describe the
validation status for the restore test validation.}
}
\description{
This request allows you to send your independent self-run restore test validation results. \code{RestoreJobId} and \code{ValidationStatus} are required. Optionally, you can input a \code{ValidationStatusMessage}.

See \url{https://www.paws-r-sdk.com/docs/backup_put_restore_validation_result/} for full documentation.
}
\keyword{internal}
