% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_revoke_restore_access_backup_vault}
\alias{backup_revoke_restore_access_backup_vault}
\title{Revokes access to a restore access backup vault, removing the ability to
restore from its recovery points and permanently deleting the vault}
\usage{
backup_revoke_restore_access_backup_vault(
  BackupVaultName,
  RestoreAccessBackupVaultArn,
  RequesterComment = NULL
)
}
\arguments{
\item{BackupVaultName}{[required] The name of the source backup vault associated with the restore access
backup vault to be revoked.}

\item{RestoreAccessBackupVaultArn}{[required] The ARN of the restore access backup vault to revoke.}

\item{RequesterComment}{A comment explaining the reason for revoking access to the restore
access backup vault.}
}
\description{
Revokes access to a restore access backup vault, removing the ability to restore from its recovery points and permanently deleting the vault.

See \url{https://www.paws-r-sdk.com/docs/backup_revoke_restore_access_backup_vault/} for full documentation.
}
\keyword{internal}
