% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_start_report_job}
\alias{backup_start_report_job}
\title{Starts an on-demand report job for the specified report plan}
\usage{
backup_start_report_job(ReportPlanName, IdempotencyToken = NULL)
}
\arguments{
\item{ReportPlanName}{[required] The unique name of a report plan.}

\item{IdempotencyToken}{A customer-chosen string that you can use to distinguish between
otherwise identical calls to \code{StartReportJobInput}. Retrying a
successful request with the same idempotency token results in a success
message with no action taken.}
}
\description{
Starts an on-demand report job for the specified report plan.

See \url{https://www.paws-r-sdk.com/docs/backup_start_report_job/} for full documentation.
}
\keyword{internal}
