% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_start_scan_job}
\alias{backup_start_scan_job}
\title{Starts scanning jobs for specific resources}
\usage{
backup_start_scan_job(
  BackupVaultName,
  IamRoleArn,
  IdempotencyToken = NULL,
  MalwareScanner,
  RecoveryPointArn,
  ScanBaseRecoveryPointArn = NULL,
  ScanMode,
  ScannerRoleArn
)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the Amazon Web Services Region where they are created.

Pattern: \verb{^[a-zA-Z0-9\\-_]\{2,50\}$}}

\item{IamRoleArn}{[required] Specifies the IAM role ARN used to create the target recovery point; for
example, \verb{arn:aws:iam::123456789012:role/S3Access}.}

\item{IdempotencyToken}{A customer-chosen string that you can use to distinguish between
otherwise identical calls to \code{\link[=backup_start_scan_job]{start_scan_job}}.
Retrying a successful request with the same idempotency token results in
a success message with no action taken.}

\item{MalwareScanner}{[required] Specifies the malware scanner used during the scan job. Currently only
supports \code{GUARDDUTY}.}

\item{RecoveryPointArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies a recovery point.
This is your target recovery point for a full scan. If you are running
an incremental scan, this will be your a recovery point which has been
created after your base recovery point selection.}

\item{ScanBaseRecoveryPointArn}{An ARN that uniquely identifies the base recovery point to be used for
incremental scanning.}

\item{ScanMode}{[required] Specifies the scan type use for the scan job.

Includes:
\itemize{
\item \code{FULL_SCAN} will scan the entire data lineage within the backup.
\item \code{INCREMENTAL_SCAN} will scan the data difference between the target
recovery point and base recovery point ARN.
}}

\item{ScannerRoleArn}{[required] Specified the IAM scanner role ARN.}
}
\description{
Starts scanning jobs for specific resources.

See \url{https://www.paws-r-sdk.com/docs/backup_start_scan_job/} for full documentation.
}
\keyword{internal}
