% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_stop_backup_job}
\alias{backup_stop_backup_job}
\title{Attempts to cancel a job to create a one-time backup of a resource}
\usage{
backup_stop_backup_job(BackupJobId)
}
\arguments{
\item{BackupJobId}{[required] Uniquely identifies a request to Backup to back up a resource.}
}
\description{
Attempts to cancel a job to create a one-time backup of a resource.

See \url{https://www.paws-r-sdk.com/docs/backup_stop_backup_job/} for full documentation.
}
\keyword{internal}
