% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_untag_resource}
\alias{backup_untag_resource}
\title{Removes a set of key-value pairs from a recovery point, backup plan, or
backup vault identified by an Amazon Resource Name (ARN)}
\usage{
backup_untag_resource(ResourceArn, TagKeyList)
}
\arguments{
\item{ResourceArn}{[required] An ARN that uniquely identifies a resource. The format of the ARN
depends on the type of the tagged resource.}

\item{TagKeyList}{[required] A list of keys to identify which key-value tags to remove from a
resource.}
}
\description{
Removes a set of key-value pairs from a recovery point, backup plan, or backup vault identified by an Amazon Resource Name (ARN)

See \url{https://paws-r.github.io/docs/backup/untag_resource.html} for full documentation.
}
\keyword{internal}
