% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_update_backup_plan}
\alias{backup_update_backup_plan}
\title{Updates the specified backup plan}
\usage{
backup_update_backup_plan(BackupPlanId, BackupPlan)
}
\arguments{
\item{BackupPlanId}{[required] The ID of the backup plan.}

\item{BackupPlan}{[required] The body of a backup plan. Includes a \code{BackupPlanName} and one or more
sets of \code{Rules}.}
}
\description{
Updates the specified backup plan. The new version is uniquely identified by its ID.

See \url{https://www.paws-r-sdk.com/docs/backup_update_backup_plan/} for full documentation.
}
\keyword{internal}
