% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_update_restore_testing_selection}
\alias{backup_update_restore_testing_selection}
\title{Updates the specified restore testing selection}
\usage{
backup_update_restore_testing_selection(
  RestoreTestingPlanName,
  RestoreTestingSelection,
  RestoreTestingSelectionName
)
}
\arguments{
\item{RestoreTestingPlanName}{[required] The restore testing plan name is required to update the indicated
testing plan.}

\item{RestoreTestingSelection}{[required] To update your restore testing selection, you can use either protected
resource ARNs or conditions, but not both. That is, if your selection
has \code{ProtectedResourceArns}, requesting an update with the parameter
\code{ProtectedResourceConditions} will be unsuccessful.}

\item{RestoreTestingSelectionName}{[required] The required restore testing selection name of the restore testing
selection you wish to update.}
}
\description{
Updates the specified restore testing selection.

See \url{https://www.paws-r-sdk.com/docs/backup_update_restore_testing_selection/} for full documentation.
}
\keyword{internal}
