% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_update_tiering_configuration}
\alias{backup_update_tiering_configuration}
\title{This request will send changes to your specified tiering configuration}
\usage{
backup_update_tiering_configuration(
  TieringConfigurationName,
  TieringConfiguration
)
}
\arguments{
\item{TieringConfigurationName}{[required] The name of a tiering configuration to update.}

\item{TieringConfiguration}{[required] Specifies the body of a tiering configuration.}
}
\description{
This request will send changes to your specified tiering configuration. \code{TieringConfigurationName} cannot be updated after it is created.

See \url{https://www.paws-r-sdk.com/docs/backup_update_tiering_configuration/} for full documentation.
}
\keyword{internal}
