% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_delete_object}
\alias{backupstorage_delete_object}
\title{Delete Object from the incremental base Backup}
\usage{
backupstorage_delete_object(BackupJobId, ObjectName)
}
\arguments{
\item{BackupJobId}{[required] Backup job Id for the in-progress backup.}

\item{ObjectName}{[required] The name of the Object.}
}
\description{
Delete Object from the incremental base Backup.

See \url{https://paws-r.github.io/docs/backupstorage/delete_object.html} for full documentation.
}
\keyword{internal}
