% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_get_chunk}
\alias{backupstorage_get_chunk}
\title{Gets the specified object's chunk}
\usage{
backupstorage_get_chunk(StorageJobId, ChunkToken)
}
\arguments{
\item{StorageJobId}{[required] Storage job id}

\item{ChunkToken}{[required] Chunk token}
}
\description{
Gets the specified object's chunk.

See \url{https://www.paws-r-sdk.com/docs/backupstorage_get_chunk/} for full documentation.
}
\keyword{internal}
