% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_get_object_metadata}
\alias{backupstorage_get_object_metadata}
\title{Get metadata associated with an Object}
\usage{
backupstorage_get_object_metadata(StorageJobId, ObjectToken)
}
\arguments{
\item{StorageJobId}{[required] Backup job id for the in-progress backup.}

\item{ObjectToken}{[required] Object token.}
}
\description{
Get metadata associated with an Object.

See \url{https://paws-r.github.io/docs/backupstorage/get_object_metadata.html} for full documentation.
}
\keyword{internal}
