% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_list_objects}
\alias{backupstorage_list_objects}
\title{List all Objects in a given Backup}
\usage{
backupstorage_list_objects(
  StorageJobId,
  StartingObjectName = NULL,
  StartingObjectPrefix = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  CreatedBefore = NULL,
  CreatedAfter = NULL
)
}
\arguments{
\item{StorageJobId}{[required] Storage job id}

\item{StartingObjectName}{Optional, specifies the starting Object name to list from. Ignored if
NextToken is not NULL}

\item{StartingObjectPrefix}{Optional, specifies the starting Object prefix to list from. Ignored if
NextToken is not NULL}

\item{MaxResults}{Maximum objects count}

\item{NextToken}{Pagination token}

\item{CreatedBefore}{(Optional) Created before filter}

\item{CreatedAfter}{(Optional) Created after filter}
}
\description{
List all Objects in a given Backup.

See \url{https://paws-r.github.io/docs/backupstorage/list_objects.html} for full documentation.
}
\keyword{internal}
