% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_notify_object_complete}
\alias{backupstorage_notify_object_complete}
\title{Complete upload}
\usage{
backupstorage_notify_object_complete(
  BackupJobId,
  UploadId,
  ObjectChecksum,
  ObjectChecksumAlgorithm,
  MetadataString = NULL,
  MetadataBlob = NULL,
  MetadataBlobLength = NULL,
  MetadataBlobChecksum = NULL,
  MetadataBlobChecksumAlgorithm = NULL
)
}
\arguments{
\item{BackupJobId}{[required] Backup job Id for the in-progress backup}

\item{UploadId}{[required] Upload Id for the in-progress upload}

\item{ObjectChecksum}{[required] Object checksum}

\item{ObjectChecksumAlgorithm}{[required] Checksum algorithm}

\item{MetadataString}{Optional metadata associated with an Object. Maximum string length is
256 bytes.}

\item{MetadataBlob}{Optional metadata associated with an Object. Maximum length is 4MB.}

\item{MetadataBlobLength}{The size of MetadataBlob.}

\item{MetadataBlobChecksum}{Checksum of MetadataBlob.}

\item{MetadataBlobChecksumAlgorithm}{Checksum algorithm.}
}
\description{
Complete upload

See \url{https://www.paws-r-sdk.com/docs/backupstorage_notify_object_complete/} for full documentation.
}
\keyword{internal}
