% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_put_chunk}
\alias{backupstorage_put_chunk}
\title{Upload chunk}
\usage{
backupstorage_put_chunk(
  BackupJobId,
  UploadId,
  ChunkIndex,
  Data,
  Length,
  Checksum,
  ChecksumAlgorithm
)
}
\arguments{
\item{BackupJobId}{[required] Backup job Id for the in-progress backup.}

\item{UploadId}{[required] Upload Id for the in-progress upload.}

\item{ChunkIndex}{[required] Describes this chunk's position relative to the other chunks}

\item{Data}{[required] Data to be uploaded}

\item{Length}{[required] Data length}

\item{Checksum}{[required] Data checksum}

\item{ChecksumAlgorithm}{[required] Checksum algorithm}
}
\description{
Upload chunk.

See \url{https://www.paws-r-sdk.com/docs/backupstorage_put_chunk/} for full documentation.
}
\keyword{internal}
