% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_delete_lifecycle_policy}
\alias{dlm_delete_lifecycle_policy}
\title{Deletes the specified lifecycle policy and halts the automated
operations that the policy specified}
\usage{
dlm_delete_lifecycle_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The identifier of the lifecycle policy.}
}
\description{
Deletes the specified lifecycle policy and halts the automated operations that the policy specified.

See \url{https://paws-r.github.io/docs/dlm/delete_lifecycle_policy.html} for full documentation.
}
\keyword{internal}
