% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_get_lifecycle_policy}
\alias{dlm_get_lifecycle_policy}
\title{Gets detailed information about the specified lifecycle policy}
\usage{
dlm_get_lifecycle_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The identifier of the lifecycle policy.}
}
\description{
Gets detailed information about the specified lifecycle policy.

See \url{https://paws-r.github.io/docs/dlm/get_lifecycle_policy.html} for full documentation.
}
\keyword{internal}
