% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_list_tags_for_resource}
\alias{dlm_list_tags_for_resource}
\title{Lists the tags for the specified resource}
\usage{
dlm_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Lists the tags for the specified resource.

See \url{https://paws-r.github.io/docs/dlm/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
