% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_untag_resource}
\alias{dlm_untag_resource}
\title{Removes the specified tags from the specified resource}
\usage{
dlm_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required] The tag keys.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Removes the specified tags from the specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
