% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_create_tags}
\alias{efs_create_tags}
\title{DEPRECATED - CreateTags is deprecated and not maintained}
\usage{
efs_create_tags(FileSystemId, Tags)
}
\arguments{
\item{FileSystemId}{[required] The ID of the file system whose tags you want to modify (String). This
operation modifies the tags only, not the file system.}

\item{Tags}{[required] An array of \code{Tag} objects to add. Each \code{Tag} object is a key-value pair.}
}
\description{
DEPRECATED - \code{\link[=efs_create_tags]{create_tags}} is deprecated and not maintained. To create tags for EFS resources, use the API action.

See \url{https://paws-r.github.io/docs/efs/create_tags.html} for full documentation.
}
\keyword{internal}
