% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_delete_file_system_policy}
\alias{efs_delete_file_system_policy}
\title{Deletes the FileSystemPolicy for the specified file system}
\usage{
efs_delete_file_system_policy(FileSystemId)
}
\arguments{
\item{FileSystemId}{[required] Specifies the EFS file system for which to delete the
\code{FileSystemPolicy}.}
}
\value{
An empty list.
}
\description{
Deletes the \code{FileSystemPolicy} for the specified file system. The
default \code{FileSystemPolicy} goes into effect once the existing policy is
deleted. For more information about the default file system policy, see
\href{https://docs.aws.amazon.com/efs/latest/ug/}{Using Resource-based Policies with EFS}.

This operation requires permissions for the
\code{elasticfilesystem:DeleteFileSystemPolicy} action.
}
\section{Request syntax}{
\preformatted{svc$delete_file_system_policy(
  FileSystemId = "string"
)
}
}

\keyword{internal}
