% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_delete_tags}
\alias{efs_delete_tags}
\title{DEPRECATED - DeleteTags is deprecated and not maintained}
\usage{
efs_delete_tags(FileSystemId, TagKeys)
}
\arguments{
\item{FileSystemId}{[required] The ID of the file system whose tags you want to delete (String).}

\item{TagKeys}{[required] A list of tag keys to delete.}
}
\description{
DEPRECATED - \code{\link[=efs_delete_tags]{delete_tags}} is deprecated and not maintained. To remove tags from EFS resources, use the API action.

See \url{https://www.paws-r-sdk.com/docs/efs_delete_tags/} for full documentation.
}
\keyword{internal}
