% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_describe_backup_policy}
\alias{efs_describe_backup_policy}
\title{Returns the backup policy for the specified EFS file system}
\usage{
efs_describe_backup_policy(FileSystemId)
}
\arguments{
\item{FileSystemId}{[required] Specifies which EFS file system to retrieve the \code{BackupPolicy} for.}
}
\value{
A list with the following syntax:\preformatted{list(
  BackupPolicy = list(
    Status = "ENABLED"|"ENABLING"|"DISABLED"|"DISABLING"
  )
)
}
}
\description{
Returns the backup policy for the specified EFS file system.
}
\section{Request syntax}{
\preformatted{svc$describe_backup_policy(
  FileSystemId = "string"
)
}
}

\keyword{internal}
