% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_describe_mount_target_security_groups}
\alias{efs_describe_mount_target_security_groups}
\title{Returns the security groups currently in effect for a mount target}
\usage{
efs_describe_mount_target_security_groups(MountTargetId)
}
\arguments{
\item{MountTargetId}{[required] The ID of the mount target whose security groups you want to retrieve.}
}
\description{
Returns the security groups currently in effect for a mount target. This operation requires that the network interface of the mount target has been created and the lifecycle state of the mount target is not \code{deleted}.

See \url{https://paws-r.github.io/docs/efs/describe_mount_target_security_groups.html} for full documentation.
}
\keyword{internal}
