% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_describe_replication_configurations}
\alias{efs_describe_replication_configurations}
\title{Retrieves the replication configuration for a specific file system}
\usage{
efs_describe_replication_configurations(
  FileSystemId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{FileSystemId}{You can retrieve the replication configuration for a specific file
system by providing its file system ID.}

\item{NextToken}{\code{NextToken} is present if the response is paginated. You can use
\code{NextToken} in a subsequent request to fetch the next page of output.}

\item{MaxResults}{(Optional) To limit the number of objects returned in a response, you
can specify the \code{MaxItems} parameter. The default value is 100.}
}
\description{
Retrieves the replication configuration for a specific file system. If a file system is not specified, all of the replication configurations for the Amazon Web Services account in an Amazon Web Services Region are retrieved.

See \url{https://paws-r.github.io/docs/efs/describe_replication_configurations.html} for full documentation.
}
\keyword{internal}
