% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_put_backup_policy}
\alias{efs_put_backup_policy}
\title{Updates the file system's backup policy}
\usage{
efs_put_backup_policy(FileSystemId, BackupPolicy)
}
\arguments{
\item{FileSystemId}{[required] Specifies which EFS file system to update the backup policy for.}

\item{BackupPolicy}{[required] The backup policy included in the
\code{\link[=efs_put_backup_policy]{put_backup_policy}} request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BackupPolicy = list(
    Status = "ENABLED"|"ENABLING"|"DISABLED"|"DISABLING"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the file system's backup policy. Use this action to start or
stop automatic backups of the file system.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_backup_policy(
  FileSystemId = "string",
  BackupPolicy = list(
    Status = "ENABLED"|"ENABLING"|"DISABLED"|"DISABLING"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
