% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_associate_user_to_permission_group}
\alias{finspacedata_associate_user_to_permission_group}
\title{Adds a user to a permission group to grant permissions for actions a
user can perform in FinSpace}
\usage{
finspacedata_associate_user_to_permission_group(
  permissionGroupId,
  userId,
  clientToken = NULL
)
}
\arguments{
\item{permissionGroupId}{[required] The unique identifier for the permission group.}

\item{userId}{[required] The unique identifier for the user.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Adds a user to a permission group to grant permissions for actions a user can perform in FinSpace.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_associate_user_to_permission_group/} for full documentation.
}
\keyword{internal}
