% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_delete_permission_group}
\alias{finspacedata_delete_permission_group}
\title{Deletes a permission group}
\usage{
finspacedata_delete_permission_group(permissionGroupId, clientToken = NULL)
}
\arguments{
\item{permissionGroupId}{[required] The unique identifier for the permission group that you want to delete.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Deletes a permission group. This action is irreversible.

See \url{https://paws-r.github.io/docs/finspacedata/delete_permission_group.html} for full documentation.
}
\keyword{internal}
