% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_disable_user}
\alias{finspacedata_disable_user}
\title{Denies access to the FinSpace web application and API for the specified
user}
\usage{
finspacedata_disable_user(userId, clientToken = NULL)
}
\arguments{
\item{userId}{[required] The unique identifier for the user account that you want to disable.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Denies access to the FinSpace web application and API for the specified user.

See \url{https://paws-r.github.io/docs/finspacedata/disable_user.html} for full documentation.
}
\keyword{internal}
