% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_changeset}
\alias{finspacedata_get_changeset}
\title{Get information about a Changeset}
\usage{
finspacedata_get_changeset(datasetId, changesetId)
}
\arguments{
\item{datasetId}{[required] The unique identifier for the FinSpace Dataset where the Changeset is
created.}

\item{changesetId}{[required] The unique identifier of the Changeset for which to get data.}
}
\description{
Get information about a Changeset.

See \url{https://paws-r.github.io/docs/finspacedata/get_changeset.html} for full documentation.
}
\keyword{internal}
