% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_dataset}
\alias{finspacedata_get_dataset}
\title{Returns information about a Dataset}
\usage{
finspacedata_get_dataset(datasetId)
}
\arguments{
\item{datasetId}{[required] The unique identifier for a Dataset.}
}
\description{
Returns information about a Dataset.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_get_dataset/} for full documentation.
}
\keyword{internal}
