% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_external_data_view_access_details}
\alias{finspacedata_get_external_data_view_access_details}
\title{Returns the credentials to access the external Dataview from an S3
location}
\usage{
finspacedata_get_external_data_view_access_details(dataViewId, datasetId)
}
\arguments{
\item{dataViewId}{[required] The unique identifier for the Dataview that you want to access.}

\item{datasetId}{[required] The unique identifier for the Dataset.}
}
\description{
Returns the credentials to access the external Dataview from an S3 location. To call this API:

See \url{https://www.paws-r-sdk.com/docs/finspacedata_get_external_data_view_access_details/} for full documentation.
}
\keyword{internal}
