% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_user}
\alias{finspacedata_get_user}
\title{Retrieves details for a specific user}
\usage{
finspacedata_get_user(userId)
}
\arguments{
\item{userId}{[required] The unique identifier of the user to get data for.}
}
\description{
Retrieves details for a specific user.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_get_user/} for full documentation.
}
\keyword{internal}
