% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_working_location}
\alias{finspacedata_get_working_location}
\title{A temporary Amazon S3 location, where you can copy your files from a
source location to stage or use as a scratch space in FinSpace notebook}
\usage{
finspacedata_get_working_location(locationType = NULL)
}
\arguments{
\item{locationType}{Specify the type of the working location.
\itemize{
\item \code{SAGEMAKER} – Use the Amazon S3 location as a temporary location to
store data content when working with FinSpace Notebooks that run on
SageMaker studio.
\item \code{INGESTION} – Use the Amazon S3 location as a staging location to copy
your data content and then use the location with the Changeset
creation operation.
}}
}
\description{
A temporary Amazon S3 location, where you can copy your files from a source location to stage or use as a scratch space in FinSpace notebook.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_get_working_location/} for full documentation.
}
\keyword{internal}
