% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_list_changesets}
\alias{finspacedata_list_changesets}
\title{Lists the FinSpace Changesets for a Dataset}
\usage{
finspacedata_list_changesets(datasetId, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{datasetId}{[required] The unique identifier for the FinSpace Dataset to which the Changeset
belongs.}

\item{maxResults}{The maximum number of results per page.}

\item{nextToken}{A token that indicates where a results page should begin.}
}
\description{
Lists the FinSpace Changesets for a Dataset.

See \url{https://paws-r.github.io/docs/finspacedata/list_changesets.html} for full documentation.
}
\keyword{internal}
