% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_list_data_views}
\alias{finspacedata_list_data_views}
\title{Lists all available Dataviews for a Dataset}
\usage{
finspacedata_list_data_views(datasetId, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{datasetId}{[required] The unique identifier of the Dataset for which to retrieve Dataviews.}

\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{The maximum number of results per page.}
}
\description{
Lists all available Dataviews for a Dataset.

See \url{https://paws-r.github.io/docs/finspacedata/list_data_views.html} for full documentation.
}
\keyword{internal}
