% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_list_datasets}
\alias{finspacedata_list_datasets}
\title{Lists all of the active Datasets that a user has access to}
\usage{
finspacedata_list_datasets(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{The maximum number of results per page.}
}
\description{
Lists all of the active Datasets that a user has access to.

See \url{https://paws-r.github.io/docs/finspacedata/list_datasets.html} for full documentation.
}
\keyword{internal}
