% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_list_permission_groups}
\alias{finspacedata_list_permission_groups}
\title{Lists all available permission groups in FinSpace}
\usage{
finspacedata_list_permission_groups(nextToken = NULL, maxResults)
}
\arguments{
\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{[required] The maximum number of results per page.}
}
\description{
Lists all available permission groups in FinSpace.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_list_permission_groups/} for full documentation.
}
\keyword{internal}
