% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_list_permission_groups_by_user}
\alias{finspacedata_list_permission_groups_by_user}
\title{Lists all the permission groups that are associated with a specific user}
\usage{
finspacedata_list_permission_groups_by_user(
  userId,
  nextToken = NULL,
  maxResults
)
}
\arguments{
\item{userId}{[required] The unique identifier for the user.}

\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{[required] The maximum number of results per page.}
}
\description{
Lists all the permission groups that are associated with a specific user.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_list_permission_groups_by_user/} for full documentation.
}
\keyword{internal}
