% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_list_users}
\alias{finspacedata_list_users}
\title{Lists all available user accounts in FinSpace}
\usage{
finspacedata_list_users(nextToken = NULL, maxResults)
}
\arguments{
\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{[required] The maximum number of results per page.}
}
\description{
Lists all available user accounts in FinSpace.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_list_users/} for full documentation.
}
\keyword{internal}
