% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_list_users_by_permission_group}
\alias{finspacedata_list_users_by_permission_group}
\title{Lists details of all the users in a specific permission group}
\usage{
finspacedata_list_users_by_permission_group(
  permissionGroupId,
  nextToken = NULL,
  maxResults
)
}
\arguments{
\item{permissionGroupId}{[required] The unique identifier for the permission group.}

\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{[required] The maximum number of results per page.}
}
\description{
Lists details of all the users in a specific permission group.

See \url{https://paws-r.github.io/docs/finspacedata/list_users_by_permission_group.html} for full documentation.
}
\keyword{internal}
