% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_reset_user_password}
\alias{finspacedata_reset_user_password}
\title{Resets the password for a specified user ID and generates a temporary
one}
\usage{
finspacedata_reset_user_password(userId, clientToken = NULL)
}
\arguments{
\item{userId}{[required] The unique identifier of the user that a temporary password is requested
for.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Resets the password for a specified user ID and generates a temporary one. Only a superuser can reset password for other users. Resetting the password immediately invalidates the previous password associated with the user.

See \url{https://paws-r.github.io/docs/finspacedata/reset_user_password.html} for full documentation.
}
\keyword{internal}
