% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_cancel_data_repository_task}
\alias{fsx_cancel_data_repository_task}
\title{Cancels an existing Amazon FSx for Lustre data repository task if that
task is in either the PENDING or EXECUTING state}
\usage{
fsx_cancel_data_repository_task(TaskId)
}
\arguments{
\item{TaskId}{[required] Specifies the data repository task to cancel.}
}
\description{
Cancels an existing Amazon FSx for Lustre data repository task if that task is in either the \code{PENDING} or \code{EXECUTING} state. When you cancel am export task, Amazon FSx does the following.

See \url{https://www.paws-r-sdk.com/docs/fsx_cancel_data_repository_task/} for full documentation.
}
\keyword{internal}
