% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_create_and_attach_s3_access_point}
\alias{fsx_create_and_attach_s3_access_point}
\title{Creates an S3 access point and attaches it to an Amazon FSx volume}
\usage{
fsx_create_and_attach_s3_access_point(
  ClientRequestToken = NULL,
  Name,
  Type,
  OpenZFSConfiguration = NULL,
  OntapConfiguration = NULL,
  S3AccessPoint = NULL
)
}
\arguments{
\item{ClientRequestToken}{}

\item{Name}{[required] The name you want to assign to this S3 access point.}

\item{Type}{[required] The type of S3 access point you want to create. Only \code{OpenZFS} is
supported.}

\item{OpenZFSConfiguration}{Specifies the configuration to use when creating and attaching an S3
access point to an FSx for OpenZFS volume.}

\item{OntapConfiguration}{}

\item{S3AccessPoint}{Specifies the virtual private cloud (VPC) configuration if you're
creating an access point that is restricted to a VPC. For more
information, see \href{https://docs.aws.amazon.com/fsx/latest/OpenZFSGuide/access-points-vpc.html}{Creating access points restricted to a virtual private cloud}.}
}
\description{
Creates an S3 access point and attaches it to an Amazon FSx volume. For FSx for OpenZFS file systems, the volume must be hosted on a high-availability file system, either Single-AZ or Multi-AZ. For more information, see \href{https://docs.aws.amazon.com/fsx/latest/OpenZFSGuide/s3accesspoints-for-FSx.html}{Accessing your data using Amazon S3 access points}. in the Amazon FSx for OpenZFS User Guide.

See \url{https://www.paws-r-sdk.com/docs/fsx_create_and_attach_s3_access_point/} for full documentation.
}
\keyword{internal}
