% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_create_snapshot}
\alias{fsx_create_snapshot}
\title{Creates a snapshot of an existing Amazon FSx for OpenZFS volume}
\usage{
fsx_create_snapshot(ClientRequestToken = NULL, Name, VolumeId, Tags = NULL)
}
\arguments{
\item{ClientRequestToken}{(Optional) An idempotency token for resource creation, in a string of up to 63 ASCII characters. This token is automatically filled on your behalf when you use the Command Line Interface (CLI) or an Amazon Web Services SDK.}

\item{Name}{[required] The name of the snapshot.}

\item{VolumeId}{[required] The ID of the volume that you are taking a snapshot of.}

\item{Tags}{A list of \code{Tag} values, with a maximum of 50 elements.}
}
\description{
Creates a snapshot of an existing Amazon FSx for OpenZFS volume. With snapshots, you can easily undo file changes and compare file versions by restoring the volume to a previous version.

See \url{https://www.paws-r-sdk.com/docs/fsx_create_snapshot/} for full documentation.
}
\keyword{internal}
