% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_create_volume}
\alias{fsx_create_volume}
\title{Creates an FSx for ONTAP or Amazon FSx for OpenZFS storage volume}
\usage{
fsx_create_volume(
  ClientRequestToken = NULL,
  VolumeType,
  Name,
  OntapConfiguration = NULL,
  Tags = NULL,
  OpenZFSConfiguration = NULL
)
}
\arguments{
\item{ClientRequestToken}{}

\item{VolumeType}{[required] Specifies the type of volume to create; \code{ONTAP} and \code{OPENZFS} are the
only valid volume types.}

\item{Name}{[required] Specifies the name of the volume that you're creating.}

\item{OntapConfiguration}{Specifies the configuration to use when creating the ONTAP volume.}

\item{Tags}{}

\item{OpenZFSConfiguration}{Specifies the configuration to use when creating the OpenZFS volume.}
}
\description{
Creates an FSx for ONTAP or Amazon FSx for OpenZFS storage volume.

See \url{https://paws-r.github.io/docs/fsx/create_volume.html} for full documentation.
}
\keyword{internal}
