% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_delete_backup}
\alias{fsx_delete_backup}
\title{Deletes an Amazon FSx backup, deleting its contents}
\usage{
fsx_delete_backup(BackupId, ClientRequestToken)
}
\arguments{
\item{BackupId}{[required] The ID of the backup you want to delete.}

\item{ClientRequestToken}{A string of up to 64 ASCII characters that Amazon FSx uses to ensure
idempotent deletion. This is automatically filled on your behalf when
using the AWS CLI or SDK.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BackupId = "string",
  Lifecycle = "AVAILABLE"|"CREATING"|"TRANSFERRING"|"DELETED"|"FAILED"|"PENDING"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an Amazon FSx backup, deleting its contents. After deletion, the
backup no longer exists, and its data is gone.

The \code{\link[=fsx_delete_backup]{delete_backup}} call returns instantly. The
backup will not show up in later
\code{\link[=fsx_describe_backups]{describe_backups}} calls.

The data in a deleted backup is also deleted and can't be recovered by
any means.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_backup(
  BackupId = "string",
  ClientRequestToken = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This operation deletes an Amazon FSx file system backup.
svc$delete_backup(
  BackupId = "backup-03e3c82e0183b7b6b"
)
}

}
\keyword{internal}
