% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_delete_file_cache}
\alias{fsx_delete_file_cache}
\title{Deletes an Amazon File Cache resource}
\usage{
fsx_delete_file_cache(FileCacheId, ClientRequestToken = NULL)
}
\arguments{
\item{FileCacheId}{[required] The ID of the cache that's being deleted.}

\item{ClientRequestToken}{(Optional) An idempotency token for resource creation, in a string of up to 63 ASCII characters. This token is automatically filled on your behalf when you use the Command Line Interface (CLI) or an Amazon Web Services SDK.}
}
\description{
Deletes an Amazon File Cache resource. After deletion, the cache no longer exists, and its data is gone.

See \url{https://www.paws-r-sdk.com/docs/fsx_delete_file_cache/} for full documentation.
}
\keyword{internal}
