% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_delete_snapshot}
\alias{fsx_delete_snapshot}
\title{Deletes an Amazon FSx for OpenZFS snapshot}
\usage{
fsx_delete_snapshot(ClientRequestToken = NULL, SnapshotId)
}
\arguments{
\item{ClientRequestToken}{}

\item{SnapshotId}{[required] The ID of the snapshot that you want to delete.}
}
\description{
Deletes an Amazon FSx for OpenZFS snapshot. After deletion, the snapshot no longer exists, and its data is gone. Deleting a snapshot doesn't affect snapshots stored in a file system backup.

See \url{https://www.paws-r-sdk.com/docs/fsx_delete_snapshot/} for full documentation.
}
\keyword{internal}
