% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_delete_storage_virtual_machine}
\alias{fsx_delete_storage_virtual_machine}
\title{Deletes an existing Amazon FSx for ONTAP storage virtual machine (SVM)}
\usage{
fsx_delete_storage_virtual_machine(
  ClientRequestToken = NULL,
  StorageVirtualMachineId
)
}
\arguments{
\item{ClientRequestToken}{(Optional) An idempotency token for resource creation, in a string of up to 63 ASCII characters. This token is automatically filled on your behalf when you use the Command Line Interface (CLI) or an Amazon Web Services SDK.}

\item{StorageVirtualMachineId}{[required] The ID of the SVM that you want to delete.}
}
\description{
Deletes an existing Amazon FSx for ONTAP storage virtual machine (SVM). Prior to deleting an SVM, you must delete all non-root volumes in the SVM, otherwise the operation will fail.

See \url{https://www.paws-r-sdk.com/docs/fsx_delete_storage_virtual_machine/} for full documentation.
}
\keyword{internal}
