% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_describe_file_caches}
\alias{fsx_describe_file_caches}
\title{Returns the description of a specific Amazon File Cache resource, if a
FileCacheIds value is provided for that cache}
\usage{
fsx_describe_file_caches(
  FileCacheIds = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{FileCacheIds}{IDs of the caches whose descriptions you want to retrieve (String).}

\item{MaxResults}{The maximum number of resources to return in the response. This value must be an integer greater than zero.}

\item{NextToken}{(Optional) Opaque pagination token returned from a previous operation (String). If present, this token indicates from what point you can continue processing the request, where the previous \code{NextToken} value left off.}
}
\description{
Returns the description of a specific Amazon File Cache resource, if a \code{FileCacheIds} value is provided for that cache. Otherwise, it returns descriptions of all caches owned by your Amazon Web Services account in the Amazon Web Services Region of the endpoint that you're calling.

See \url{https://www.paws-r-sdk.com/docs/fsx_describe_file_caches/} for full documentation.
}
\keyword{internal}
