% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_describe_shared_vpc_configuration}
\alias{fsx_describe_shared_vpc_configuration}
\title{Indicates whether participant accounts in your organization can create
Amazon FSx for NetApp ONTAP Multi-AZ file systems in subnets that are
shared by a virtual private cloud (VPC) owner}
\usage{
fsx_describe_shared_vpc_configuration()
}
\description{
Indicates whether participant accounts in your organization can create Amazon FSx for NetApp ONTAP Multi-AZ file systems in subnets that are shared by a virtual private cloud (VPC) owner. For more information, see the \href{https://docs.aws.amazon.com/fsx/latest/ONTAPGuide/}{Amazon FSx for NetApp ONTAP User Guide}.

See \url{https://www.paws-r-sdk.com/docs/fsx_describe_shared_vpc_configuration/} for full documentation.
}
\keyword{internal}
