% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_describe_storage_virtual_machines}
\alias{fsx_describe_storage_virtual_machines}
\title{Describes one or more Amazon FSx for NetApp ONTAP storage virtual
machines (SVMs)}
\usage{
fsx_describe_storage_virtual_machines(
  StorageVirtualMachineIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StorageVirtualMachineIds}{Enter the ID of one or more SVMs that you want to view.}

\item{Filters}{Enter a filter name:value pair to view a select set of SVMs.}

\item{MaxResults}{The maximum number of resources to return in the response. This value must be an integer greater than zero.}

\item{NextToken}{(Optional) Opaque pagination token returned from a previous operation (String). If present, this token indicates from what point you can continue processing the request, where the previous \code{NextToken} value left off.}
}
\description{
Describes one or more Amazon FSx for NetApp ONTAP storage virtual machines (SVMs).

See \url{https://www.paws-r-sdk.com/docs/fsx_describe_storage_virtual_machines/} for full documentation.
}
\keyword{internal}
