% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_describe_volumes}
\alias{fsx_describe_volumes}
\title{Describes one or more Amazon FSx for NetApp ONTAP or Amazon FSx for
OpenZFS volumes}
\usage{
fsx_describe_volumes(
  VolumeIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{VolumeIds}{The IDs of the volumes whose descriptions you want to retrieve.}

\item{Filters}{Enter a filter \code{Name} and \code{Values} pair to view a select set of volumes.}

\item{MaxResults}{The maximum number of resources to return in the response. This value must be an integer greater than zero.}

\item{NextToken}{(Optional) Opaque pagination token returned from a previous operation (String). If present, this token indicates from what point you can continue processing the request, where the previous \code{NextToken} value left off.}
}
\description{
Describes one or more Amazon FSx for NetApp ONTAP or Amazon FSx for OpenZFS volumes.

See \url{https://www.paws-r-sdk.com/docs/fsx_describe_volumes/} for full documentation.
}
\keyword{internal}
