% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_detach_and_delete_s3_access_point}
\alias{fsx_detach_and_delete_s3_access_point}
\title{Detaches an S3 access point from an Amazon FSx volume and deletes the S3
access point}
\usage{
fsx_detach_and_delete_s3_access_point(ClientRequestToken = NULL, Name)
}
\arguments{
\item{ClientRequestToken}{(Optional) An idempotency token for resource creation, in a string of up to 63 ASCII characters. This token is automatically filled on your behalf when you use the Command Line Interface (CLI) or an Amazon Web Services SDK.}

\item{Name}{[required] The name of the S3 access point attachment that you want to delete.}
}
\description{
Detaches an S3 access point from an Amazon FSx volume and deletes the S3 access point.

See \url{https://www.paws-r-sdk.com/docs/fsx_detach_and_delete_s3_access_point/} for full documentation.
}
\keyword{internal}
