% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_disassociate_file_system_aliases}
\alias{fsx_disassociate_file_system_aliases}
\title{Use this action to disassociate, or remove, one or more Domain Name
Service (DNS) aliases from an Amazon FSx for Windows File Server file
system}
\usage{
fsx_disassociate_file_system_aliases(
  ClientRequestToken = NULL,
  FileSystemId,
  Aliases
)
}
\arguments{
\item{ClientRequestToken}{}

\item{FileSystemId}{[required] Specifies the file system from which to disassociate the DNS aliases.}

\item{Aliases}{[required] An array of one or more DNS alias names to disassociate, or remove, from
the file system.}
}
\description{
Use this action to disassociate, or remove, one or more Domain Name Service (DNS) aliases from an Amazon FSx for Windows File Server file system. If you attempt to disassociate a DNS alias that is not associated with the file system, Amazon FSx responds with a 400 Bad Request. For more information, see \href{https://docs.aws.amazon.com/fsx/latest/WindowsGuide/managing-dns-aliases.html}{Working with DNS Aliases}.

See \url{https://www.paws-r-sdk.com/docs/fsx_disassociate_file_system_aliases/} for full documentation.
}
\keyword{internal}
