% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_list_tags_for_resource}
\alias{fsx_list_tags_for_resource}
\title{Lists tags for Amazon FSx resources}
\usage{
fsx_list_tags_for_resource(ResourceARN, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the Amazon FSx resource that will have its tags listed.}

\item{MaxResults}{Maximum number of tags to return in the response (integer). This
parameter value must be greater than 0. The number of items that Amazon
FSx returns is the minimum of the \code{MaxResults} parameter specified in
the request and the service's internal maximum number of items per page.}

\item{NextToken}{Opaque pagination token returned from a previous
\code{\link[=fsx_list_tags_for_resource]{list_tags_for_resource}} operation
(String). If a token present, the action continues the list from where
the returning call left off.}
}
\description{
Lists tags for Amazon FSx resources.

See \url{https://www.paws-r-sdk.com/docs/fsx_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
