% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_start_misconfigured_state_recovery}
\alias{fsx_start_misconfigured_state_recovery}
\title{After performing steps to repair the Active Directory configuration of
an FSx for Windows File Server file system, use this action to initiate
the process of Amazon FSx attempting to reconnect to the file system}
\usage{
fsx_start_misconfigured_state_recovery(ClientRequestToken = NULL, FileSystemId)
}
\arguments{
\item{ClientRequestToken}{(Optional) An idempotency token for resource creation, in a string of up to 63 ASCII characters. This token is automatically filled on your behalf when you use the Command Line Interface (CLI) or an Amazon Web Services SDK.}

\item{FileSystemId}{[required] The globally unique ID of the file system, assigned by Amazon FSx.}
}
\description{
After performing steps to repair the Active Directory configuration of an FSx for Windows File Server file system, use this action to initiate the process of Amazon FSx attempting to reconnect to the file system.

See \url{https://www.paws-r-sdk.com/docs/fsx_start_misconfigured_state_recovery/} for full documentation.
}
\keyword{internal}
