% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_tag_resource}
\alias{fsx_tag_resource}
\title{Tags an Amazon FSx resource}
\usage{
fsx_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the Amazon FSx resource that you want
to tag.}

\item{Tags}{[required] A list of tags for the resource. If a tag with a given key already
exists, the value is replaced by the one specified in this parameter.}
}
\description{
Tags an Amazon FSx resource.

See \url{https://paws-r.github.io/docs/fsx/tag_resource.html} for full documentation.
}
\keyword{internal}
