% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_update_shared_vpc_configuration}
\alias{fsx_update_shared_vpc_configuration}
\title{Configures whether participant accounts in your organization can create
Amazon FSx for NetApp ONTAP Multi-AZ file systems in subnets that are
shared by a virtual private cloud (VPC) owner}
\usage{
fsx_update_shared_vpc_configuration(
  EnableFsxRouteTableUpdatesFromParticipantAccounts = NULL,
  ClientRequestToken = NULL
)
}
\arguments{
\item{EnableFsxRouteTableUpdatesFromParticipantAccounts}{Specifies whether participant accounts can create FSx for ONTAP Multi-AZ
file systems in shared subnets. Set to \code{true} to enable or \code{false} to
disable.}

\item{ClientRequestToken}{}
}
\description{
Configures whether participant accounts in your organization can create Amazon FSx for NetApp ONTAP Multi-AZ file systems in subnets that are shared by a virtual private cloud (VPC) owner. For more information, see the \href{https://docs.aws.amazon.com/fsx/latest/ONTAPGuide/}{Amazon FSx for NetApp ONTAP User Guide}.

See \url{https://www.paws-r-sdk.com/docs/fsx_update_shared_vpc_configuration/} for full documentation.
}
\keyword{internal}
