% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_update_snapshot}
\alias{fsx_update_snapshot}
\title{Updates the name of an Amazon FSx for OpenZFS snapshot}
\usage{
fsx_update_snapshot(ClientRequestToken = NULL, Name, SnapshotId)
}
\arguments{
\item{ClientRequestToken}{}

\item{Name}{[required] The name of the snapshot to update.}

\item{SnapshotId}{[required] The ID of the snapshot that you want to update, in the format
\verb{fsvolsnap-0123456789abcdef0}.}
}
\description{
Updates the name of an Amazon FSx for OpenZFS snapshot.

See \url{https://paws-r.github.io/docs/fsx/update_snapshot.html} for full documentation.
}
\keyword{internal}
