% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_update_storage_virtual_machine}
\alias{fsx_update_storage_virtual_machine}
\title{Updates an Amazon FSx for ONTAP storage virtual machine (SVM)}
\usage{
fsx_update_storage_virtual_machine(
  ActiveDirectoryConfiguration = NULL,
  ClientRequestToken = NULL,
  StorageVirtualMachineId,
  SvmAdminPassword = NULL
)
}
\arguments{
\item{ActiveDirectoryConfiguration}{Updates the Microsoft Active Directory (AD) configuration for an SVM
that is joined to an AD.}

\item{ClientRequestToken}{}

\item{StorageVirtualMachineId}{[required] The ID of the SVM that you want to update, in the format
\verb{svm-0123456789abcdef0}.}

\item{SvmAdminPassword}{Enter a new SvmAdminPassword if you are updating it.}
}
\description{
Updates an Amazon FSx for ONTAP storage virtual machine (SVM).

See \url{https://paws-r.github.io/docs/fsx/update_storage_virtual_machine.html} for full documentation.
}
\keyword{internal}
