% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier_operations.R
\name{glacier_abort_multipart_upload}
\alias{glacier_abort_multipart_upload}
\title{This operation aborts a multipart upload identified by the upload ID}
\usage{
glacier_abort_multipart_upload(accountId, vaultName, uploadId)
}
\arguments{
\item{accountId}{[required] The \code{AccountId} value is the AWS account ID of the account that owns the
vault. You can either specify an AWS account ID or optionally a single
'\code{-}' (hyphen), in which case Amazon Glacier uses the AWS account ID
associated with the credentials used to sign the request. If you use an
account ID, do not include any hyphens ('-') in the ID.}

\item{vaultName}{[required] The name of the vault.}

\item{uploadId}{[required] The upload ID of the multipart upload to delete.}
}
\description{
This operation aborts a multipart upload identified by the upload ID.

See \url{https://www.paws-r-sdk.com/docs/glacier_abort_multipart_upload/} for full documentation.
}
\keyword{internal}
