% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier_operations.R
\name{glacier_get_vault_lock}
\alias{glacier_get_vault_lock}
\title{This operation retrieves the following attributes from the lock-policy
subresource set on the specified vault:}
\usage{
glacier_get_vault_lock(accountId, vaultName)
}
\arguments{
\item{accountId}{[required] The \code{AccountId} value is the AWS account ID of the account that owns the
vault. You can either specify an AWS account ID or optionally a single
'\code{-}' (hyphen), in which case Amazon Glacier uses the AWS account ID
associated with the credentials used to sign the request. If you use an
account ID, do not include any hyphens ('-') in the ID.}

\item{vaultName}{[required] The name of the vault.}
}
\description{
This operation retrieves the following attributes from the \code{lock-policy} subresource set on the specified vault:

See \url{https://www.paws-r-sdk.com/docs/glacier_get_vault_lock/} for full documentation.
}
\keyword{internal}
