% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier_operations.R
\name{glacier_get_vault_notifications}
\alias{glacier_get_vault_notifications}
\title{This operation retrieves the notification-configuration subresource of
the specified vault}
\usage{
glacier_get_vault_notifications(accountId, vaultName)
}
\arguments{
\item{accountId}{[required] The \code{AccountId} value is the AWS account ID of the account that owns the
vault. You can either specify an AWS account ID or optionally a single
'\code{-}' (hyphen), in which case Amazon Glacier uses the AWS account ID
associated with the credentials used to sign the request. If you use an
account ID, do not include any hyphens ('-') in the ID.}

\item{vaultName}{[required] The name of the vault.}
}
\description{
This operation retrieves the \code{notification-configuration} subresource of the specified vault.

See \url{https://www.paws-r-sdk.com/docs/glacier_get_vault_notifications/} for full documentation.
}
\keyword{internal}
