% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier_operations.R
\name{glacier_initiate_vault_lock}
\alias{glacier_initiate_vault_lock}
\title{This operation initiates the vault locking process by doing the
following:}
\usage{
glacier_initiate_vault_lock(accountId, vaultName, policy = NULL)
}
\arguments{
\item{accountId}{[required] The \code{AccountId} value is the AWS account ID. This value must match the
AWS account ID associated with the credentials used to sign the request.
You can either specify an AWS account ID or optionally a single '\code{-}'
(hyphen), in which case Amazon Glacier uses the AWS account ID
associated with the credentials used to sign the request. If you specify
your account ID, do not include any hyphens ('-') in the ID.}

\item{vaultName}{[required] The name of the vault.}

\item{policy}{The vault lock policy as a JSON string, which uses "\\" as an escape
character.}
}
\description{
This operation initiates the vault locking process by doing the following:

See \url{https://paws-r.github.io/docs/glacier/initiate_vault_lock.html} for full documentation.
}
\keyword{internal}
