% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier_operations.R
\name{glacier_purchase_provisioned_capacity}
\alias{glacier_purchase_provisioned_capacity}
\title{This operation purchases a provisioned capacity unit for an AWS account}
\usage{
glacier_purchase_provisioned_capacity(accountId)
}
\arguments{
\item{accountId}{[required] The AWS account ID of the account that owns the vault. You can either
specify an AWS account ID or optionally a single '-' (hyphen), in which
case Amazon S3 Glacier uses the AWS account ID associated with the
credentials used to sign the request. If you use an account ID, don't
include any hyphens ('-') in the ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  capacityId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
This operation purchases a provisioned capacity unit for an AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$purchase_provisioned_capacity(
  accountId = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The example purchases provisioned capacity unit for an AWS account.
svc$purchase_provisioned_capacity(
  accountId = "-"
)
}

}
\keyword{internal}
