% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_accept_share}
\alias{omics_accept_share}
\title{Accepts a share for an analytics store}
\usage{
omics_accept_share(shareId)
}
\arguments{
\item{shareId}{[required] The ID for a share offer for analytics store data.}
}
\description{
Accepts a share for an analytics store.

See \url{https://www.paws-r-sdk.com/docs/omics_accept_share/} for full documentation.
}
\keyword{internal}
