% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_batch_delete_read_set}
\alias{omics_batch_delete_read_set}
\title{Deletes one or more read sets}
\usage{
omics_batch_delete_read_set(ids, sequenceStoreId)
}
\arguments{
\item{ids}{[required] The read sets' IDs.}

\item{sequenceStoreId}{[required] The read sets' sequence store ID.}
}
\description{
Deletes one or more read sets.

See \url{https://www.paws-r-sdk.com/docs/omics_batch_delete_read_set/} for full documentation.
}
\keyword{internal}
