% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_cancel_run}
\alias{omics_cancel_run}
\title{Cancels a run}
\usage{
omics_cancel_run(id)
}
\arguments{
\item{id}{[required] The run's ID.}
}
\description{
Cancels a run.

See \url{https://www.paws-r-sdk.com/docs/omics_cancel_run/} for full documentation.
}
\keyword{internal}
