% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_annotation_store_version}
\alias{omics_create_annotation_store_version}
\title{Creates a new version of an annotation store}
\usage{
omics_create_annotation_store_version(
  name,
  versionName,
  description = NULL,
  versionOptions = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of an annotation store version from which versions are being
created.}

\item{versionName}{[required] The name given to an annotation store version to distinguish it from
other versions.}

\item{description}{The description of an annotation store version.}

\item{versionOptions}{The options for an annotation store version.}

\item{tags}{Any tags added to annotation store version.}
}
\description{
Creates a new version of an annotation store.

See \url{https://www.paws-r-sdk.com/docs/omics_create_annotation_store_version/} for full documentation.
}
\keyword{internal}
