% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_configuration}
\alias{omics_create_configuration}
\title{Create a new configuration}
\usage{
omics_create_configuration(
  name,
  description = NULL,
  runConfigurations,
  tags = NULL,
  requestId
)
}
\arguments{
\item{name}{[required] User-friendly name for the configuration.}

\item{description}{Optional description for the configuration.}

\item{runConfigurations}{[required] Required run-specific configurations.}

\item{tags}{Optional tags for the configuration.}

\item{requestId}{[required] Optional request idempotency token. If not specified, a universally
unique identifier (UUID) will be automatically generated for the
request.}
}
\description{
Create a new configuration.

See \url{https://www.paws-r-sdk.com/docs/omics_create_configuration/} for full documentation.
}
\keyword{internal}
