% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_multipart_read_set_upload}
\alias{omics_create_multipart_read_set_upload}
\title{Begins a multipart read set upload}
\usage{
omics_create_multipart_read_set_upload(
  sequenceStoreId,
  clientToken = NULL,
  sourceFileType,
  subjectId,
  sampleId,
  generatedFrom = NULL,
  referenceArn = NULL,
  name,
  description = NULL,
  tags = NULL
)
}
\arguments{
\item{sequenceStoreId}{[required] The sequence store ID for the store that is the destination of the
multipart uploads.}

\item{clientToken}{An idempotency token that can be used to avoid triggering multiple
multipart uploads.}

\item{sourceFileType}{[required] The type of file being uploaded.}

\item{subjectId}{[required] The source's subject ID.}

\item{sampleId}{[required] The source's sample ID.}

\item{generatedFrom}{Where the source originated.}

\item{referenceArn}{The ARN of the reference.}

\item{name}{[required] The name of the read set.}

\item{description}{The description of the read set.}

\item{tags}{Any tags to add to the read set.}
}
\description{
Begins a multipart read set upload.

See \url{https://www.paws-r-sdk.com/docs/omics_create_multipart_read_set_upload/} for full documentation.
}
\keyword{internal}
