% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_sequence_store}
\alias{omics_create_sequence_store}
\title{Creates a sequence store}
\usage{
omics_create_sequence_store(
  name,
  description = NULL,
  sseConfig = NULL,
  tags = NULL,
  clientToken = NULL,
  fallbackLocation = NULL
)
}
\arguments{
\item{name}{[required] A name for the store.}

\item{description}{A description for the store.}

\item{sseConfig}{Server-side encryption (SSE) settings for the store.}

\item{tags}{Tags for the store.}

\item{clientToken}{To ensure that requests don't run multiple times, specify a unique token
for each request.}

\item{fallbackLocation}{An S3 location that is used to store files that have failed a direct
upload.}
}
\description{
Creates a sequence store.

See \url{https://www.paws-r-sdk.com/docs/omics_create_sequence_store/} for full documentation.
}
\keyword{internal}
