% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_share}
\alias{omics_create_share}
\title{Creates a cross-account shared resource}
\usage{
omics_create_share(resourceArn, principalSubscriber, shareName = NULL)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be shared.}

\item{principalSubscriber}{[required] The principal subscriber is the account being offered shared access to the resource.}

\item{shareName}{A name that the owner defines for the share.}
}
\description{
Creates a cross-account shared resource. The resource owner makes an offer to share the resource with the principal subscriber (an Amazon Web Services user with a different account than the resource owner).

See \url{https://www.paws-r-sdk.com/docs/omics_create_share/} for full documentation.
}
\keyword{internal}
