% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_variant_store}
\alias{omics_create_variant_store}
\title{Amazon Web Services HealthOmics variant stores and annotation stores are
no longer open to new customers}
\usage{
omics_create_variant_store(
  reference,
  name = NULL,
  description = NULL,
  tags = NULL,
  sseConfig = NULL
)
}
\arguments{
\item{reference}{[required] The genome reference for the store's variants.}

\item{name}{A name for the store.}

\item{description}{A description for the store.}

\item{tags}{Tags for the store.}

\item{sseConfig}{Server-side encryption (SSE) settings for the store.}
}
\description{
Amazon Web Services HealthOmics variant stores and annotation stores are no longer open to new customers. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/variant-store-availability-change.html}{Amazon Web Services HealthOmics variant store and annotation store availability change}.

See \url{https://www.paws-r-sdk.com/docs/omics_create_variant_store/} for full documentation.
}
\keyword{internal}
