% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_workflow}
\alias{omics_create_workflow}
\title{Creates a private workflow}
\usage{
omics_create_workflow(
  name = NULL,
  description = NULL,
  engine = NULL,
  definitionZip = NULL,
  definitionUri = NULL,
  main = NULL,
  parameterTemplate = NULL,
  storageCapacity = NULL,
  tags = NULL,
  requestId,
  accelerators = NULL,
  storageType = NULL
)
}
\arguments{
\item{name}{A name for the workflow.}

\item{description}{A description for the workflow.}

\item{engine}{The workflow engine for the workflow.}

\item{definitionZip}{A ZIP archive for the workflow.}

\item{definitionUri}{The URI of a definition for the workflow.}

\item{main}{The path of the main definition file for the workflow.}

\item{parameterTemplate}{A parameter template for the workflow.}

\item{storageCapacity}{The default static storage capacity (in gibibytes) for runs that use
this workflow or workflow version.}

\item{tags}{Tags for the workflow.}

\item{requestId}{[required] To ensure that requests don't run multiple times, specify a unique ID
for each request.}

\item{accelerators}{The computational accelerator specified to run the workflow.}

\item{storageType}{The default storage type for runs that use this workflow. STATIC storage
allocates a fixed amount of storage. DYNAMIC storage dynamically scales
the storage up or down, based on file system utilization. For more
information about static and dynamic storage, see \href{https://docs.aws.amazon.com/omics/latest/dev/}{Running workflows} in the \emph{Amazon
Web Services HealthOmics User Guide}.}
}
\description{
Creates a private workflow.Private workflows depend on a variety of resources that you create and configure before creating the workflow:

See \url{https://www.paws-r-sdk.com/docs/omics_create_workflow/} for full documentation.
}
\keyword{internal}
