% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_workflow_version}
\alias{omics_create_workflow_version}
\title{Creates a new workflow version for the workflow that you specify with
the workflowId parameter}
\usage{
omics_create_workflow_version(
  workflowId,
  versionName,
  definitionZip = NULL,
  definitionUri = NULL,
  accelerators = NULL,
  description = NULL,
  engine = NULL,
  main = NULL,
  parameterTemplate = NULL,
  requestId,
  storageType = NULL,
  storageCapacity = NULL,
  tags = NULL,
  workflowBucketOwnerId = NULL
)
}
\arguments{
\item{workflowId}{[required] The ID of the workflow where you are creating the new version.}

\item{versionName}{[required] A name for the workflow version. Provide a version name that is unique
for this workflow. You cannot change the name after HealthOmics creates
the version.

The version name must start with a letter or number and it can include
upper-case and lower-case letters, numbers, hyphens, periods and
underscores. The maximum length is 64 characters. You can use a simple
naming scheme, such as version1, version2, version3. You can also match
your workflow versions with your own internal versioning conventions,
such as 2.7.0, 2.7.1, 2.7.2.}

\item{definitionZip}{A zip archive containing the workflow definition for this workflow
version.}

\item{definitionUri}{The URI specifies the location of the workflow definition for this
workflow version.}

\item{accelerators}{The computational accelerator for this workflow version.}

\item{description}{A description for this workflow version.}

\item{engine}{The workflow engine for this workflow version.}

\item{main}{The path of the main definition file for this workflow version.}

\item{parameterTemplate}{The parameter template defines the input parameters for runs that use
this workflow version.}

\item{requestId}{[required] To ensure that requests don't run multiple times, specify a unique ID
for each request.}

\item{storageType}{The default storage type for runs that use this workflow. STATIC storage
allocates a fixed amount of storage. DYNAMIC storage dynamically scales
the storage up or down, based on file system utilization. For more
information about static and dynamic storage, see \href{https://docs.aws.amazon.com/omics/latest/dev/}{Running workflows} in the \emph{Amazon
Web Services HealthOmics User Guide}.}

\item{storageCapacity}{The default static storage capacity (in gibibytes) for runs that use
this workflow or workflow version.}

\item{tags}{Optional tags to associate with this workflow version.}

\item{workflowBucketOwnerId}{Amazon Web Services Id of the owner of the S3 bucket that contains the
workflow definition. You need to specify this parameter if your account
is not the bucket owner.}
}
\description{
Creates a new workflow version for the workflow that you specify with the \code{workflowId} parameter.

See \url{https://www.paws-r-sdk.com/docs/omics_create_workflow_version/} for full documentation.
}
\keyword{internal}
