% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_annotation_store_versions}
\alias{omics_delete_annotation_store_versions}
\title{Deletes one or multiple versions of an annotation store}
\usage{
omics_delete_annotation_store_versions(name, versions, force = NULL)
}
\arguments{
\item{name}{[required] The name of the annotation store from which versions are being deleted.}

\item{versions}{[required] The versions of an annotation store to be deleted.}

\item{force}{Forces the deletion of an annotation store version when imports are
in-progress..}
}
\description{
Deletes one or multiple versions of an annotation store.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_annotation_store_versions/} for full documentation.
}
\keyword{internal}
