% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_batch}
\alias{omics_delete_batch}
\title{Deletes a run batch resource and its associated metadata}
\usage{
omics_delete_batch(batchId)
}
\arguments{
\item{batchId}{[required] The identifier portion of the run batch ARN.}
}
\description{
Deletes a run batch resource and its associated metadata. This operation does not delete the individual workflow runs. To delete the runs, call \code{\link[=omics_delete_run_batch]{delete_run_batch}} before calling \code{\link[=omics_delete_batch]{delete_batch}}.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_batch/} for full documentation.
}
\keyword{internal}
