% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_configuration}
\alias{omics_delete_configuration}
\title{Delete an existing configuration}
\usage{
omics_delete_configuration(name)
}
\arguments{
\item{name}{[required] Configuration name to delete.}
}
\description{
Delete an existing configuration.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_configuration/} for full documentation.
}
\keyword{internal}
