% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_reference}
\alias{omics_delete_reference}
\title{Deletes a genome reference}
\usage{
omics_delete_reference(id, referenceStoreId)
}
\arguments{
\item{id}{[required] The reference's ID.}

\item{referenceStoreId}{[required] The reference's store ID.}
}
\description{
Deletes a genome reference.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_reference/} for full documentation.
}
\keyword{internal}
