% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_reference_store}
\alias{omics_delete_reference_store}
\title{Deletes a genome reference store}
\usage{
omics_delete_reference_store(id)
}
\arguments{
\item{id}{[required] The store's ID.}
}
\description{
Deletes a genome reference store.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_reference_store/} for full documentation.
}
\keyword{internal}
