% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_run_batch}
\alias{omics_delete_run_batch}
\title{Deletes the individual workflow runs within a batch}
\usage{
omics_delete_run_batch(batchId)
}
\arguments{
\item{batchId}{[required] The identifier portion of the run batch ARN.}
}
\description{
Deletes the individual workflow runs within a batch. This operation is separate from \code{\link[=omics_delete_batch]{delete_batch}}, which removes the batch metadata.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_run_batch/} for full documentation.
}
\keyword{internal}
