% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_run_cache}
\alias{omics_delete_run_cache}
\title{Delete a run cache}
\usage{
omics_delete_run_cache(id)
}
\arguments{
\item{id}{[required] Run cache identifier for the cache you want to delete.}
}
\description{
Delete a run cache. This action removes the cache metadata stored in the service account, but doesn't delete the data in Amazon S3. You can access the cache data in Amazon S3, for inspection or to troubleshoot issues. You can remove old cache data using standard S3 \code{Delete} operations.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_run_cache/} for full documentation.
}
\keyword{internal}
