% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_run_group}
\alias{omics_delete_run_group}
\title{Deletes a run group and returns a response with no body if the operation
is successful}
\usage{
omics_delete_run_group(id)
}
\arguments{
\item{id}{[required] The run group's ID.}
}
\description{
Deletes a run group and returns a response with no body if the operation is successful.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_run_group/} for full documentation.
}
\keyword{internal}
