% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_s3_access_policy}
\alias{omics_delete_s3_access_policy}
\title{Deletes an access policy for the specified store}
\usage{
omics_delete_s3_access_policy(s3AccessPointArn)
}
\arguments{
\item{s3AccessPointArn}{[required] The S3 access point ARN that has the access policy.}
}
\description{
Deletes an access policy for the specified store.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_s3_access_policy/} for full documentation.
}
\keyword{internal}
