% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_sequence_store}
\alias{omics_delete_sequence_store}
\title{Deletes a sequence store}
\usage{
omics_delete_sequence_store(id)
}
\arguments{
\item{id}{[required] The sequence store's ID.}
}
\description{
Deletes a sequence store.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_sequence_store/} for full documentation.
}
\keyword{internal}
