% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_share}
\alias{omics_delete_share}
\title{Deletes a resource share}
\usage{
omics_delete_share(shareId)
}
\arguments{
\item{shareId}{[required] The ID for the resource share to be deleted.}
}
\description{
Deletes a resource share. If you are the resource owner, the subscriber will no longer have access to the shared resource. If you are the subscriber, this operation deletes your access to the share.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_share/} for full documentation.
}
\keyword{internal}
