% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_workflow_version}
\alias{omics_delete_workflow_version}
\title{Deletes a workflow version}
\usage{
omics_delete_workflow_version(workflowId, versionName)
}
\arguments{
\item{workflowId}{[required] The workflow's ID.}

\item{versionName}{[required] The workflow version name.}
}
\description{
Deletes a workflow version. Deleting a workflow version doesn't affect any ongoing runs that are using the workflow version.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_workflow_version/} for full documentation.
}
\keyword{internal}
