% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_annotation_import_job}
\alias{omics_get_annotation_import_job}
\title{Gets information about an annotation import job}
\usage{
omics_get_annotation_import_job(jobId)
}
\arguments{
\item{jobId}{[required] The job's ID.}
}
\description{
Gets information about an annotation import job.

See \url{https://www.paws-r-sdk.com/docs/omics_get_annotation_import_job/} for full documentation.
}
\keyword{internal}
