% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_annotation_store}
\alias{omics_get_annotation_store}
\title{Gets information about an annotation store}
\usage{
omics_get_annotation_store(name)
}
\arguments{
\item{name}{[required] The store's name.}
}
\description{
Gets information about an annotation store.

See \url{https://www.paws-r-sdk.com/docs/omics_get_annotation_store/} for full documentation.
}
\keyword{internal}
