% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_annotation_store_version}
\alias{omics_get_annotation_store_version}
\title{Retrieves the metadata for an annotation store version}
\usage{
omics_get_annotation_store_version(name, versionName)
}
\arguments{
\item{name}{[required] The name given to an annotation store version to distinguish it from
others.}

\item{versionName}{[required] The name given to an annotation store version to distinguish it from
others.}
}
\description{
Retrieves the metadata for an annotation store version.

See \url{https://www.paws-r-sdk.com/docs/omics_get_annotation_store_version/} for full documentation.
}
\keyword{internal}
