% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_batch}
\alias{omics_get_batch}
\title{Retrieves details and current status for a specific run batch, including
submission progress and run execution counts}
\usage{
omics_get_batch(batchId)
}
\arguments{
\item{batchId}{[required] The identifier portion of the run batch ARN.}
}
\description{
Retrieves details and current status for a specific run batch, including submission progress and run execution counts.

See \url{https://www.paws-r-sdk.com/docs/omics_get_batch/} for full documentation.
}
\keyword{internal}
