% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_configuration}
\alias{omics_get_configuration}
\title{Retrieve configuration details for specified name}
\usage{
omics_get_configuration(name)
}
\arguments{
\item{name}{[required] Configuration name to retrieve.}
}
\description{
Retrieve configuration details for specified name.

See \url{https://www.paws-r-sdk.com/docs/omics_get_configuration/} for full documentation.
}
\keyword{internal}
