% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_read_set_activation_job}
\alias{omics_get_read_set_activation_job}
\title{Returns detailed information about the status of a read set activation
job in JSON format}
\usage{
omics_get_read_set_activation_job(id, sequenceStoreId)
}
\arguments{
\item{id}{[required] The job's ID.}

\item{sequenceStoreId}{[required] The job's sequence store ID.}
}
\description{
Returns detailed information about the status of a read set activation job in JSON format.

See \url{https://www.paws-r-sdk.com/docs/omics_get_read_set_activation_job/} for full documentation.
}
\keyword{internal}
