% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_read_set_export_job}
\alias{omics_get_read_set_export_job}
\title{Gets information about a read set export job}
\usage{
omics_get_read_set_export_job(sequenceStoreId, id)
}
\arguments{
\item{sequenceStoreId}{[required] The job's sequence store ID.}

\item{id}{[required] The job's ID.}
}
\description{
Gets information about a read set export job.

See \url{https://www.paws-r-sdk.com/docs/omics_get_read_set_export_job/} for full documentation.
}
\keyword{internal}
