% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_read_set_import_job}
\alias{omics_get_read_set_import_job}
\title{Gets information about a read set import job}
\usage{
omics_get_read_set_import_job(id, sequenceStoreId)
}
\arguments{
\item{id}{[required] The job's ID.}

\item{sequenceStoreId}{[required] The job's sequence store ID.}
}
\description{
Gets information about a read set import job.

See \url{https://www.paws-r-sdk.com/docs/omics_get_read_set_import_job/} for full documentation.
}
\keyword{internal}
