% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_read_set_metadata}
\alias{omics_get_read_set_metadata}
\title{Gets details about a read set}
\usage{
omics_get_read_set_metadata(id, sequenceStoreId)
}
\arguments{
\item{id}{[required] The read set's ID.}

\item{sequenceStoreId}{[required] The read set's sequence store ID.}
}
\description{
Gets details about a read set.

See \url{https://www.paws-r-sdk.com/docs/omics_get_read_set_metadata/} for full documentation.
}
\keyword{internal}
