% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_reference}
\alias{omics_get_reference}
\title{Gets a reference file}
\usage{
omics_get_reference(
  id,
  referenceStoreId,
  range = NULL,
  partNumber,
  file = NULL
)
}
\arguments{
\item{id}{[required] The reference's ID.}

\item{referenceStoreId}{[required] The reference's store ID.}

\item{range}{The range to retrieve.}

\item{partNumber}{[required] The part number to retrieve.}

\item{file}{The file to retrieve.}
}
\description{
Gets a reference file.

See \url{https://www.paws-r-sdk.com/docs/omics_get_reference/} for full documentation.
}
\keyword{internal}
