% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_reference_import_job}
\alias{omics_get_reference_import_job}
\title{Gets information about a reference import job}
\usage{
omics_get_reference_import_job(id, referenceStoreId)
}
\arguments{
\item{id}{[required] The job's ID.}

\item{referenceStoreId}{[required] The job's reference store ID.}
}
\description{
Gets information about a reference import job.

See \url{https://www.paws-r-sdk.com/docs/omics_get_reference_import_job/} for full documentation.
}
\keyword{internal}
