% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_reference_metadata}
\alias{omics_get_reference_metadata}
\title{Gets information about a genome reference's metadata}
\usage{
omics_get_reference_metadata(id, referenceStoreId)
}
\arguments{
\item{id}{[required] The reference's ID.}

\item{referenceStoreId}{[required] The reference's reference store ID.}
}
\description{
Gets information about a genome reference's metadata.

See \url{https://www.paws-r-sdk.com/docs/omics_get_reference_metadata/} for full documentation.
}
\keyword{internal}
