% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_run_group}
\alias{omics_get_run_group}
\title{Gets information about a run group and returns its metadata}
\usage{
omics_get_run_group(id)
}
\arguments{
\item{id}{[required] The group's ID.}
}
\description{
Gets information about a run group and returns its metadata.

See \url{https://www.paws-r-sdk.com/docs/omics_get_run_group/} for full documentation.
}
\keyword{internal}
