% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_s3_access_policy}
\alias{omics_get_s3_access_policy}
\title{Retrieves details about an access policy on a given store}
\usage{
omics_get_s3_access_policy(s3AccessPointArn)
}
\arguments{
\item{s3AccessPointArn}{[required] The S3 access point ARN that has the access policy.}
}
\description{
Retrieves details about an access policy on a given store.

See \url{https://www.paws-r-sdk.com/docs/omics_get_s3_access_policy/} for full documentation.
}
\keyword{internal}
