% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_sequence_store}
\alias{omics_get_sequence_store}
\title{Gets information about a sequence store}
\usage{
omics_get_sequence_store(id)
}
\arguments{
\item{id}{[required] The store's ID.}
}
\description{
Gets information about a sequence store.

See \url{https://www.paws-r-sdk.com/docs/omics_get_sequence_store/} for full documentation.
}
\keyword{internal}
