% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_variant_import_job}
\alias{omics_get_variant_import_job}
\title{Amazon Web Services HealthOmics variant stores and annotation stores are
no longer open to new customers}
\usage{
omics_get_variant_import_job(jobId)
}
\arguments{
\item{jobId}{[required] The job's ID.}
}
\description{
Amazon Web Services HealthOmics variant stores and annotation stores are no longer open to new customers. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/variant-store-availability-change.html}{Amazon Web Services HealthOmics variant store and annotation store availability change}.

See \url{https://www.paws-r-sdk.com/docs/omics_get_variant_import_job/} for full documentation.
}
\keyword{internal}
