% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_workflow}
\alias{omics_get_workflow}
\title{Gets information about a workflow}
\usage{
omics_get_workflow(id, type = NULL, export = NULL, workflowOwnerId = NULL)
}
\arguments{
\item{id}{[required] The workflow's ID.}

\item{type}{The workflow's type.}

\item{export}{The export format for the workflow.}

\item{workflowOwnerId}{The ID of the workflow owner.}
}
\description{
Gets information about a workflow.

See \url{https://www.paws-r-sdk.com/docs/omics_get_workflow/} for full documentation.
}
\keyword{internal}
