% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_workflow_version}
\alias{omics_get_workflow_version}
\title{Gets information about a workflow version}
\usage{
omics_get_workflow_version(
  workflowId,
  versionName,
  type = NULL,
  export = NULL,
  workflowOwnerId = NULL
)
}
\arguments{
\item{workflowId}{[required] The workflow's ID. The \code{workflowId} is not the UUID.}

\item{versionName}{[required] The workflow version name.}

\item{type}{The workflow's type.}

\item{export}{The export format for the workflow.}

\item{workflowOwnerId}{The 12-digit account ID of the workflow owner. The workflow owner ID can
be retrieved using the \code{\link[=omics_get_share]{get_share}} API operation. If
you are the workflow owner, you do not need to include this ID.}
}
\description{
Gets information about a workflow version. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/workflow-versions.html}{Workflow versioning in Amazon Web Services HealthOmics} in the \emph{Amazon Web Services HealthOmics User Guide}.

See \url{https://www.paws-r-sdk.com/docs/omics_get_workflow_version/} for full documentation.
}
\keyword{internal}
