% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_annotation_import_jobs}
\alias{omics_list_annotation_import_jobs}
\title{Retrieves a list of annotation import jobs}
\usage{
omics_list_annotation_import_jobs(
  maxResults = NULL,
  ids = NULL,
  nextToken = NULL,
  filter = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of jobs to return in one page of results.}

\item{ids}{IDs of annotation import jobs to retrieve.}

\item{nextToken}{Specifies the pagination token from a previous request to retrieve the
next page of results.}

\item{filter}{A filter to apply to the list.}
}
\description{
Retrieves a list of annotation import jobs.

See \url{https://www.paws-r-sdk.com/docs/omics_list_annotation_import_jobs/} for full documentation.
}
\keyword{internal}
