% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_annotation_store_versions}
\alias{omics_list_annotation_store_versions}
\title{Lists the versions of an annotation store}
\usage{
omics_list_annotation_store_versions(
  name,
  maxResults = NULL,
  nextToken = NULL,
  filter = NULL
)
}
\arguments{
\item{name}{[required] The name of an annotation store.}

\item{maxResults}{The maximum number of annotation store versions to return in one page of
results.}

\item{nextToken}{Specifies the pagination token from a previous request to retrieve the
next page of results.}

\item{filter}{A filter to apply to the list of annotation store versions.}
}
\description{
Lists the versions of an annotation store.

See \url{https://www.paws-r-sdk.com/docs/omics_list_annotation_store_versions/} for full documentation.
}
\keyword{internal}
