% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_batch}
\alias{omics_list_batch}
\title{Returns a list of run batches in your account, with optional filtering
by status, name, or run group}
\usage{
omics_list_batch(
  maxItems = NULL,
  startingToken = NULL,
  status = NULL,
  name = NULL,
  runGroupId = NULL
)
}
\arguments{
\item{maxItems}{The maximum number of batches to return. If not specified, defaults to
100.}

\item{startingToken}{A pagination token returned from a prior
\code{\link[=omics_list_batch]{list_batch}} call.}

\item{status}{Filter batches by status.}

\item{name}{Filter batches by name.}

\item{runGroupId}{Filter batches by run group ID.}
}
\description{
Returns a list of run batches in your account, with optional filtering by status, name, or run group. Results are paginated. Only one filter per call is supported.

See \url{https://www.paws-r-sdk.com/docs/omics_list_batch/} for full documentation.
}
\keyword{internal}
