% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_configurations}
\alias{omics_list_configurations}
\title{List all configurations for the account}
\usage{
omics_list_configurations(maxResults = NULL, startingToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of results to return.}

\item{startingToken}{Pagination token for retrieving next page of results.}
}
\description{
List all configurations for the account.

See \url{https://www.paws-r-sdk.com/docs/omics_list_configurations/} for full documentation.
}
\keyword{internal}
