% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_multipart_read_set_uploads}
\alias{omics_list_multipart_read_set_uploads}
\title{Lists all multipart read set uploads and their statuses}
\usage{
omics_list_multipart_read_set_uploads(
  sequenceStoreId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{sequenceStoreId}{[required] The Sequence Store ID used for the multipart uploads.}

\item{maxResults}{The maximum number of multipart uploads returned in a page.}

\item{nextToken}{Next token returned in the response of a previous
ListMultipartReadSetUploads call. Used to get the next page of results.}
}
\description{
Lists all multipart read set uploads and their statuses.

See \url{https://www.paws-r-sdk.com/docs/omics_list_multipart_read_set_uploads/} for full documentation.
}
\keyword{internal}
