% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_read_set_import_jobs}
\alias{omics_list_read_set_import_jobs}
\title{Retrieves a list of read set import jobs and returns the data in JSON
format}
\usage{
omics_list_read_set_import_jobs(
  maxResults = NULL,
  nextToken = NULL,
  sequenceStoreId,
  filter = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of jobs to return in one page of results.}

\item{nextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{sequenceStoreId}{[required] The jobs' sequence store ID.}

\item{filter}{A filter to apply to the list.}
}
\description{
Retrieves a list of read set import jobs and returns the data in JSON format.

See \url{https://www.paws-r-sdk.com/docs/omics_list_read_set_import_jobs/} for full documentation.
}
\keyword{internal}
