% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_read_set_upload_parts}
\alias{omics_list_read_set_upload_parts}
\title{Lists all parts in a multipart read set upload for a sequence store and
returns the metadata in a JSON formatted output}
\usage{
omics_list_read_set_upload_parts(
  sequenceStoreId,
  uploadId,
  partSource,
  maxResults = NULL,
  nextToken = NULL,
  filter = NULL
)
}
\arguments{
\item{sequenceStoreId}{[required] The Sequence Store ID used for the multipart uploads.}

\item{uploadId}{[required] The ID for the initiated multipart upload.}

\item{partSource}{[required] The source file for the upload part.}

\item{maxResults}{The maximum number of read set upload parts returned in a page.}

\item{nextToken}{Next token returned in the response of a previous
ListReadSetUploadPartsRequest call. Used to get the next page of
results.}

\item{filter}{Attributes used to filter for a specific subset of read set part
uploads.}
}
\description{
Lists all parts in a multipart read set upload for a sequence store and returns the metadata in a JSON formatted output.

See \url{https://www.paws-r-sdk.com/docs/omics_list_read_set_upload_parts/} for full documentation.
}
\keyword{internal}
