% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_runs}
\alias{omics_list_runs}
\title{Retrieves a list of runs and returns each run's metadata and status}
\usage{
omics_list_runs(
  name = NULL,
  runGroupId = NULL,
  startingToken = NULL,
  maxResults = NULL,
  status = NULL
)
}
\arguments{
\item{name}{Filter the list by run name.}

\item{runGroupId}{Filter the list by run group ID.}

\item{startingToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{maxResults}{The maximum number of runs to return in one page of results.}

\item{status}{The status of a run.}
}
\description{
Retrieves a list of runs and returns each run's metadata and status.

See \url{https://www.paws-r-sdk.com/docs/omics_list_runs/} for full documentation.
}
\keyword{internal}
