% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_runs_in_batch}
\alias{omics_list_runs_in_batch}
\title{Returns a paginated list of individual workflow runs within a specific
batch}
\usage{
omics_list_runs_in_batch(
  batchId,
  maxItems = NULL,
  startingToken = NULL,
  submissionStatus = NULL,
  runSettingId = NULL,
  runId = NULL
)
}
\arguments{
\item{batchId}{[required] The identifier portion of the run batch ARN.}

\item{maxItems}{The maximum number of runs to return.}

\item{startingToken}{A pagination token returned from a prior
\code{\link[=omics_list_runs_in_batch]{list_runs_in_batch}} call.}

\item{submissionStatus}{Filter runs by submission status.}

\item{runSettingId}{Filter runs by the customer-provided run setting ID.}

\item{runId}{Filter runs by the HealthOmics-generated run ID.}
}
\description{
Returns a paginated list of individual workflow runs within a specific batch. Use this operation to map each \code{runSettingId} to its HealthOmics-generated \code{runId}, and to check the submission status of each run. Only one filter per call is supported.

See \url{https://www.paws-r-sdk.com/docs/omics_list_runs_in_batch/} for full documentation.
}
\keyword{internal}
