% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_sequence_stores}
\alias{omics_list_sequence_stores}
\title{Retrieves a list of sequence stores and returns each sequence store's
metadata}
\usage{
omics_list_sequence_stores(maxResults = NULL, nextToken = NULL, filter = NULL)
}
\arguments{
\item{maxResults}{The maximum number of stores to return in one page of results.}

\item{nextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{filter}{A filter to apply to the list.}
}
\description{
Retrieves a list of sequence stores and returns each sequence store's metadata.

See \url{https://www.paws-r-sdk.com/docs/omics_list_sequence_stores/} for full documentation.
}
\keyword{internal}
