% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_tags_for_resource}
\alias{omics_list_tags_for_resource}
\title{Retrieves a list of tags for a resource}
\usage{
omics_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The resource's ARN.}
}
\description{
Retrieves a list of tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/omics_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
