% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_workflow_versions}
\alias{omics_list_workflow_versions}
\title{Lists the workflow versions for the specified workflow}
\usage{
omics_list_workflow_versions(
  workflowId,
  type = NULL,
  workflowOwnerId = NULL,
  startingToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{workflowId}{[required] The workflow's ID.}

\item{type}{The workflow type.}

\item{workflowOwnerId}{Amazon Web Services Id of the owner of the workflow.}

\item{startingToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{maxResults}{The maximum number of workflows to return in one page of results.}
}
\description{
Lists the workflow versions for the specified workflow. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/workflow-versions.html}{Workflow versioning in Amazon Web Services HealthOmics} in the Amazon Web Services HealthOmics User Guide.

See \url{https://www.paws-r-sdk.com/docs/omics_list_workflow_versions/} for full documentation.
}
\keyword{internal}
