% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_workflows}
\alias{omics_list_workflows}
\title{Retrieves a list of workflows}
\usage{
omics_list_workflows(
  type = NULL,
  name = NULL,
  startingToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{type}{The workflows' type.}

\item{name}{The workflows' name.}

\item{startingToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{maxResults}{The maximum number of workflows to return in one page of results.}
}
\description{
Retrieves a list of workflows.

See \url{https://www.paws-r-sdk.com/docs/omics_list_workflows/} for full documentation.
}
\keyword{internal}
