% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_put_s3_access_policy}
\alias{omics_put_s3_access_policy}
\title{Adds an access policy to the specified store}
\usage{
omics_put_s3_access_policy(s3AccessPointArn, s3AccessPolicy)
}
\arguments{
\item{s3AccessPointArn}{[required] The S3 access point ARN where you want to put the access policy.}

\item{s3AccessPolicy}{[required] The resource policy that controls S3 access to the store.}
}
\description{
Adds an access policy to the specified store.

See \url{https://www.paws-r-sdk.com/docs/omics_put_s3_access_policy/} for full documentation.
}
\keyword{internal}
