% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_start_annotation_import_job}
\alias{omics_start_annotation_import_job}
\title{Starts an annotation import job}
\usage{
omics_start_annotation_import_job(
  destinationName,
  roleArn,
  items,
  versionName = NULL,
  formatOptions = NULL,
  runLeftNormalization = NULL,
  annotationFields = NULL
)
}
\arguments{
\item{destinationName}{[required] A destination annotation store for the job.}

\item{roleArn}{[required] A service role for the job.}

\item{items}{[required] Items to import.}

\item{versionName}{The name of the annotation store version.}

\item{formatOptions}{Formatting options for the annotation file.}

\item{runLeftNormalization}{The job's left normalization setting.}

\item{annotationFields}{The annotation schema generated by the parsed annotation data.}
}
\description{
Starts an annotation import job.

See \url{https://www.paws-r-sdk.com/docs/omics_start_annotation_import_job/} for full documentation.
}
\keyword{internal}
