% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_start_read_set_export_job}
\alias{omics_start_read_set_export_job}
\title{Starts a read set export job}
\usage{
omics_start_read_set_export_job(
  sequenceStoreId,
  destination,
  roleArn,
  clientToken = NULL,
  sources
)
}
\arguments{
\item{sequenceStoreId}{[required] The read set's sequence store ID.}

\item{destination}{[required] A location for exported files in Amazon S3.}

\item{roleArn}{[required] A service role for the job.}

\item{clientToken}{To ensure that jobs don't run multiple times, specify a unique token for
each job.}

\item{sources}{[required] The job's source files.}
}
\description{
Starts a read set export job. When the export job is finished, the read set is exported to an Amazon S3 bucket which can be retrieved using the \code{\link[=omics_get_read_set_export_job]{get_read_set_export_job}} API operation.

See \url{https://www.paws-r-sdk.com/docs/omics_start_read_set_export_job/} for full documentation.
}
\keyword{internal}
