% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_start_variant_import_job}
\alias{omics_start_variant_import_job}
\title{Starts a variant import job}
\usage{
omics_start_variant_import_job(
  destinationName,
  roleArn,
  items,
  runLeftNormalization = NULL,
  annotationFields = NULL
)
}
\arguments{
\item{destinationName}{[required] The destination variant store for the job.}

\item{roleArn}{[required] A service role for the job.}

\item{items}{[required] Items to import.}

\item{runLeftNormalization}{The job's left normalization setting.}

\item{annotationFields}{The annotation schema generated by the parsed annotation data.}
}
\description{
Starts a variant import job.

See \url{https://www.paws-r-sdk.com/docs/omics_start_variant_import_job/} for full documentation.
}
\keyword{internal}
