% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_tag_resource}
\alias{omics_tag_resource}
\title{Tags a resource}
\usage{
omics_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The resource's ARN.}

\item{tags}{[required] Tags for the resource.}
}
\description{
Tags a resource.

See \url{https://www.paws-r-sdk.com/docs/omics_tag_resource/} for full documentation.
}
\keyword{internal}
