% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_untag_resource}
\alias{omics_untag_resource}
\title{Removes tags from a resource}
\usage{
omics_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The resource's ARN.}

\item{tagKeys}{[required] Keys of tags to remove.}
}
\description{
Removes tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/omics_untag_resource/} for full documentation.
}
\keyword{internal}
