% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_update_annotation_store_version}
\alias{omics_update_annotation_store_version}
\title{Updates the description of an annotation store version}
\usage{
omics_update_annotation_store_version(name, versionName, description = NULL)
}
\arguments{
\item{name}{[required] The name of an annotation store.}

\item{versionName}{[required] The name of an annotation store version.}

\item{description}{The description of an annotation store.}
}
\description{
Updates the description of an annotation store version.

See \url{https://www.paws-r-sdk.com/docs/omics_update_annotation_store_version/} for full documentation.
}
\keyword{internal}
