% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_update_sequence_store}
\alias{omics_update_sequence_store}
\title{Update one or more parameters for the sequence store}
\usage{
omics_update_sequence_store(
  id,
  name = NULL,
  description = NULL,
  clientToken = NULL,
  fallbackLocation = NULL,
  propagatedSetLevelTags = NULL,
  s3AccessConfig = NULL
)
}
\arguments{
\item{id}{[required] The ID of the sequence store.}

\item{name}{A name for the sequence store.}

\item{description}{A description for the sequence store.}

\item{clientToken}{To ensure that requests don't run multiple times, specify a unique token
for each request.}

\item{fallbackLocation}{The S3 URI of a bucket and folder to store Read Sets that fail to
upload.}

\item{propagatedSetLevelTags}{The tags keys to propagate to the S3 objects associated with read sets
in the sequence store.}

\item{s3AccessConfig}{S3 access configuration parameters.}
}
\description{
Update one or more parameters for the sequence store.

See \url{https://www.paws-r-sdk.com/docs/omics_update_sequence_store/} for full documentation.
}
\keyword{internal}
