% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_update_variant_store}
\alias{omics_update_variant_store}
\title{Updates a variant store}
\usage{
omics_update_variant_store(name, description = NULL)
}
\arguments{
\item{name}{[required] A name for the store.}

\item{description}{A description for the store.}
}
\description{
Updates a variant store.

See \url{https://www.paws-r-sdk.com/docs/omics_update_variant_store/} for full documentation.
}
\keyword{internal}
