% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_update_workflow}
\alias{omics_update_workflow}
\title{Updates a workflow}
\usage{
omics_update_workflow(id, name = NULL, description = NULL)
}
\arguments{
\item{id}{[required] The workflow's ID.}

\item{name}{A name for the workflow.}

\item{description}{A description for the workflow.}
}
\description{
Updates a workflow.

See \url{https://www.paws-r-sdk.com/docs/omics_update_workflow/} for full documentation.
}
\keyword{internal}
