% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_update_workflow_version}
\alias{omics_update_workflow_version}
\title{Updates information about the workflow version}
\usage{
omics_update_workflow_version(
  workflowId,
  versionName,
  description = NULL,
  storageType = NULL,
  storageCapacity = NULL
)
}
\arguments{
\item{workflowId}{[required] The workflow's ID.}

\item{versionName}{[required] The name of the workflow version.}

\item{description}{Description of the workflow version.}

\item{storageType}{The default storage type for runs that use this workflow. STATIC storage
allocates a fixed amount of storage. DYNAMIC storage dynamically scales
the storage up or down, based on file system utilization. For more
information about static and dynamic storage, see \href{https://docs.aws.amazon.com/omics/latest/dev/}{Running workflows} in the \emph{Amazon
Web Services HealthOmics User Guide}.}

\item{storageCapacity}{The default static storage capacity (in gibibytes) for runs that use
this workflow or workflow version.}
}
\description{
Updates information about the workflow version. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/workflow-versions.html}{Workflow versioning in Amazon Web Services HealthOmics} in the Amazon Web Services HealthOmics User Guide.

See \url{https://www.paws-r-sdk.com/docs/omics_update_workflow_version/} for full documentation.
}
\keyword{internal}
