% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_upload_read_set_part}
\alias{omics_upload_read_set_part}
\title{This operation uploads a specific part of a read set}
\usage{
omics_upload_read_set_part(
  sequenceStoreId,
  uploadId,
  partSource,
  partNumber,
  payload
)
}
\arguments{
\item{sequenceStoreId}{[required] The Sequence Store ID used for the multipart upload.}

\item{uploadId}{[required] The ID for the initiated multipart upload.}

\item{partSource}{[required] The source file for an upload part.}

\item{partNumber}{[required] The number of the part being uploaded.}

\item{payload}{[required] The read set data to upload for a part.}
}
\description{
This operation uploads a specific part of a read set. If you upload a new part using a previously used part number, the previously uploaded part will be overwritten.

See \url{https://www.paws-r-sdk.com/docs/omics_upload_read_set_part/} for full documentation.
}
\keyword{internal}
