% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_delete_rule}
\alias{recyclebin_delete_rule}
\title{Deletes a Recycle Bin retention rule}
\usage{
recyclebin_delete_rule(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique ID of the retention rule.}
}
\description{
Deletes a Recycle Bin retention rule. For more information, see \href{https://docs.aws.amazon.com/ebs/latest/userguide/recycle-bin-working-with-rules.html#recycle-bin-delete-rule}{Delete Recycle Bin retention rules} in the \emph{Amazon Elastic Compute Cloud User Guide}.

See \url{https://www.paws-r-sdk.com/docs/recyclebin_delete_rule/} for full documentation.
}
\keyword{internal}
