% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_list_tags_for_resource}
\alias{recyclebin_list_tags_for_resource}
\title{Lists the tags assigned to a retention rule}
\usage{
recyclebin_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the retention rule.}
}
\description{
Lists the tags assigned to a retention rule.

See \url{https://paws-r.github.io/docs/recyclebin/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
