% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_lock_rule}
\alias{recyclebin_lock_rule}
\title{Locks a retention rule}
\usage{
recyclebin_lock_rule(Identifier, LockConfiguration)
}
\arguments{
\item{Identifier}{[required] The unique ID of the retention rule.}

\item{LockConfiguration}{[required] Information about the retention rule lock configuration.}
}
\description{
Locks a retention rule. A locked retention rule can't be modified or deleted.

See \url{https://paws-r.github.io/docs/recyclebin/lock_rule.html} for full documentation.
}
\keyword{internal}
