% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_tag_resource}
\alias{recyclebin_tag_resource}
\title{Assigns tags to the specified retention rule}
\usage{
recyclebin_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the retention rule.}

\item{Tags}{[required] Information about the tags to assign to the retention rule.}
}
\description{
Assigns tags to the specified retention rule.

See \url{https://paws-r.github.io/docs/recyclebin/tag_resource.html} for full documentation.
}
\keyword{internal}
