% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_unlock_rule}
\alias{recyclebin_unlock_rule}
\title{Unlocks a retention rule}
\usage{
recyclebin_unlock_rule(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique ID of the retention rule.}
}
\description{
Unlocks a retention rule. After a retention rule is unlocked, it can be modified or deleted only after the unlock delay period expires.

See \url{https://www.paws-r-sdk.com/docs/recyclebin_unlock_rule/} for full documentation.
}
\keyword{internal}
