% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_untag_resource}
\alias{recyclebin_untag_resource}
\title{Unassigns a tag from a retention rule}
\usage{
recyclebin_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the retention rule.}

\item{TagKeys}{[required] The tag keys of the tags to unassign. All tags that have the specified
tag key are unassigned.}
}
\description{
Unassigns a tag from a retention rule.

See \url{https://paws-r.github.io/docs/recyclebin/untag_resource.html} for full documentation.
}
\keyword{internal}
