% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports_paws.common.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{paginate}
\alias{paginate_lapply}
\alias{paginate_sapply}
\alias{list_paginators}
\alias{config}
\alias{credentials}
\alias{creds}
\alias{paws_stream_parser}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{paws.common}{\code{\link[paws.common:config]{config()}}, \code{\link[paws.common:credentials]{credentials()}}, \code{\link[paws.common:creds]{creds()}}, \code{\link[paws.common:list_paginators]{list_paginators()}}, \code{\link[paws.common:paginate]{paginate()}}, \code{\link[paws.common:paginate_lapply]{paginate_lapply()}}, \code{\link[paws.common:paginate_sapply]{paginate_sapply()}}, \code{\link[paws.common:paws_stream_parser]{paws_stream_parser()}}}
}}

