% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_create_bucket_metadata_configuration}
\alias{s3_create_bucket_metadata_configuration}
\title{Creates an S3 Metadata V2 metadata configuration for a general purpose
bucket}
\usage{
s3_create_bucket_metadata_configuration(
  Bucket,
  ContentMD5 = NULL,
  ChecksumAlgorithm = NULL,
  MetadataConfiguration,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The general purpose bucket that you want to create the metadata
configuration for.}

\item{ContentMD5}{The \code{Content-MD5} header for the metadata configuration.}

\item{ChecksumAlgorithm}{The checksum algorithm to use with your metadata configuration.}

\item{MetadataConfiguration}{[required] The contents of your metadata configuration.}

\item{ExpectedBucketOwner}{The expected owner of the general purpose bucket that corresponds to
your metadata configuration.}
}
\description{
Creates an S3 Metadata V2 metadata configuration for a general purpose bucket. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/metadata-tables-overview.html}{Accelerating data discovery with S3 Metadata} in the \emph{Amazon S3 User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3_create_bucket_metadata_configuration/} for full documentation.
}
\keyword{internal}
