% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket}
\alias{s3_delete_bucket}
\title{Deletes the S3 bucket}
\usage{
s3_delete_bucket(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] Specifies the bucket being deleted.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Deletes the S3 bucket. All objects (including all object versions and delete markers) in the bucket must be deleted before the bucket itself can be deleted.

See \url{https://paws-r.github.io/docs/s3/delete_bucket.html} for full documentation.
}
\keyword{internal}
