% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_cors}
\alias{s3_delete_bucket_cors}
\title{This operation is not supported by directory buckets}
\usage{
s3_delete_bucket_cors(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] Specifies the bucket whose \code{cors} configuration is being deleted.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide does not match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_delete_bucket_cors/} for full documentation.
}
\keyword{internal}
