% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_intelligent_tiering_configuration}
\alias{s3_delete_bucket_intelligent_tiering_configuration}
\title{This operation is not supported by directory buckets}
\usage{
s3_delete_bucket_intelligent_tiering_configuration(Bucket, Id)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose configuration you want to modify
or retrieve.}

\item{Id}{[required] The ID used to identify the S3 Intelligent-Tiering configuration.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_delete_bucket_intelligent_tiering_configuration/} for full documentation.
}
\keyword{internal}
