% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_inventory_configuration}
\alias{s3_delete_bucket_inventory_configuration}
\title{Deletes an inventory configuration (identified by the inventory ID) from
the bucket}
\usage{
s3_delete_bucket_inventory_configuration(
  Bucket,
  Id,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the inventory configuration to delete.}

\item{Id}{[required] The ID used to identify the inventory configuration.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Deletes an inventory configuration (identified by the inventory ID) from the bucket.

See \url{https://www.paws-r-sdk.com/docs/s3_delete_bucket_inventory_configuration/} for full documentation.
}
\keyword{internal}
