% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_metadata_table_configuration}
\alias{s3_delete_bucket_metadata_table_configuration}
\title{We recommend that you delete your S3 Metadata configurations by using
the V2 DeleteBucketMetadataTableConfiguration API operation}
\usage{
s3_delete_bucket_metadata_table_configuration(
  Bucket,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The general purpose bucket that you want to remove the metadata table
configuration from.}

\item{ExpectedBucketOwner}{The expected bucket owner of the general purpose bucket that you want to
remove the metadata table configuration from.}
}
\description{
We recommend that you delete your S3 Metadata configurations by using the V2 \code{\link[=s3_delete_bucket_metadata_table_configuration]{delete_bucket_metadata_table_configuration}} API operation. We no longer recommend using the V1 \code{\link[=s3_delete_bucket_metadata_table_configuration]{delete_bucket_metadata_table_configuration}} API operation.

See \url{https://www.paws-r-sdk.com/docs/s3_delete_bucket_metadata_table_configuration/} for full documentation.
}
\keyword{internal}
