% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_tagging}
\alias{s3_delete_bucket_tagging}
\title{Deletes the tags from the bucket}
\usage{
s3_delete_bucket_tagging(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The bucket that has the tag set to be removed.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Deletes the tags from the bucket.

See \url{https://www.paws-r-sdk.com/docs/s3_delete_bucket_tagging/} for full documentation.
}
\keyword{internal}
