% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_object_annotation}
\alias{s3_delete_object_annotation}
\title{Deletes a specific annotation from an Amazon S3 object}
\usage{
s3_delete_object_annotation(
  Bucket,
  Key,
  AnnotationName,
  VersionId = NULL,
  RequestPayer = NULL,
  ExpectedBucketOwner = NULL,
  ObjectIfMatch = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket that contains the object.}

\item{Key}{[required] The object key.}

\item{AnnotationName}{[required] The name of the annotation to delete. Annotation names are UTF-8 encoded and cannot start with \code{aws} or \code{s3} (case-insensitive).

Length Constraints: Minimum length of 1. Maximum length of 512 bytes.}

\item{VersionId}{The version ID of the object.}

\item{RequestPayer}{Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. If either the source or destination S3 bucket has Requester Pays enabled, the requester will pay for the corresponding charges. For information about downloading objects from Requester Pays buckets, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/ObjectsinRequesterPaysBuckets.html}{Downloading Objects in Requester Pays Buckets} in the \emph{Amazon S3 User Guide}.

This functionality is not supported for directory buckets.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner.}

\item{ObjectIfMatch}{If specified, the operation only succeeds if the object's ETag matches the provided value.}
}
\description{
Deletes a specific annotation from an Amazon S3 object. Use the \verb{x-amz-object-if-match} header to perform a conditional delete that only succeeds if the object's ETag matches the provided value, preventing race conditions during concurrent updates.

See \url{https://www.paws-r-sdk.com/docs/s3_delete_object_annotation/} for full documentation.
}
\keyword{internal}
