% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_custom.R
\name{s3_generate_presigned_post}
\alias{s3_generate_presigned_post}
\title{Generate the url and form fields used for a presigned s3 post}
\usage{
s3_generate_presigned_post(Bucket, Key, Fields = list(),
Conditions = list(), ExpiresIn = 3600)
}
\arguments{
\item{Bucket}{(character): The name of the bucket to presign the post to.
Note that bucket related conditions should not be included in the
\code{Conditions} parameter.}

\item{Key}{(character): Key name, optionally add \code{${filename}} to the
end to attach the submitted filename. Note that key related conditions and
fields are filled out for you and should not be included in the
\code{Fields} or \code{Conditions} parameter.}

\item{Fields}{(list): A list of prefilled form fields to build on top of.
Note that if a particular element is included in \code{Fields} it will not
be automatically added to \code{Conditions}; you must specify a condition
for it as well.}

\item{Conditions}{(list): A list of conditions to include in the policy.
Each element is either a named list for an exact-match condition, e.g.
\code{list(acl = "public-read")}, or a plain list for an array-style
condition, e.g. \code{list("starts-with", "$key", "uploads/")} or
\code{list("content-length-range", 1, 10485760)}.}

\item{ExpiresIn}{(numeric): The number of seconds the presigned post is
valid for. By default it expires in an hour (3600 seconds).}
}
\value{
A list with two elements: \code{url} and \code{fields}. \code{url} is the
url to post to. \code{fields} is a list of form fields and values to
include in the post.
}
\description{
Generate the url and form fields used for a presigned s3 post
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$generate_presigned_post(
  Bucket = "string",
  Key = "string",
  Fields = list(),
  Conditions = list(),
  ExpiresIn = 3600
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example generates the url and form fields for a presigned
# post that you can give to others so that they can upload a file to an
# S3 bucket without needing AWS credentials.
svc$generate_presigned_post(
  Bucket = "BUCKET_NAME",
  Key = "OBJECT_KEY"
)
}
}
\keyword{internal}
