% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_accelerate_configuration}
\alias{s3_get_bucket_accelerate_configuration}
\title{This operation is not supported for directory buckets}
\usage{
s3_get_bucket_accelerate_configuration(
  Bucket,
  ExpectedBucketOwner = NULL,
  RequestPayer = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket for which the accelerate configuration is retrieved.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you provide does not match the actual owner of the bucket, the request fails with the HTTP status code \verb{403 Forbidden} (access denied).}

\item{RequestPayer}{Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. If either the source or destination S3 bucket has Requester Pays enabled, the requester will pay for the corresponding charges. For information about downloading objects from Requester Pays buckets, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/ObjectsinRequesterPaysBuckets.html}{Downloading Objects in Requester Pays Buckets} in the \emph{Amazon S3 User Guide}.

This functionality is not supported for directory buckets.}
}
\description{
This operation is not supported for directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_get_bucket_accelerate_configuration/} for full documentation.
}
\keyword{internal}
