% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_analytics_configuration}
\alias{s3_get_bucket_analytics_configuration}
\title{This implementation of the GET action returns an analytics configuration
(identified by the analytics configuration ID) from the bucket}
\usage{
s3_get_bucket_analytics_configuration(Bucket, Id, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The name of the bucket from which an analytics configuration is
retrieved.}

\item{Id}{[required] The ID that identifies the analytics configuration.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
This implementation of the GET action returns an analytics configuration (identified by the analytics configuration ID) from the bucket.

See \url{https://paws-r.github.io/docs/s3/get_bucket_analytics_configuration.html} for full documentation.
}
\keyword{internal}
