% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_inventory_configuration}
\alias{s3_get_bucket_inventory_configuration}
\title{Returns an inventory configuration (identified by the inventory ID) from
the bucket}
\usage{
s3_get_bucket_inventory_configuration(Bucket, Id)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the inventory configuration to
retrieve.}

\item{Id}{[required] The ID used to identify the inventory configuration.}
}
\description{
Returns an inventory configuration (identified by the inventory ID) from
the bucket.
}
\section{Request syntax}{
\preformatted{svc$get_bucket_inventory_configuration(
  Bucket = "string",
  Id = "string"
)
}
}

\keyword{internal}
