% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_location}
\alias{s3_get_bucket_location}
\title{Using the GetBucketLocation operation is no longer a best practice}
\usage{
s3_get_bucket_location(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The name of the bucket for which to get the location.

When you use this API operation with an access point, provide the alias of the access point in place of the bucket name.

When you use this API operation with an Object Lambda access point, provide the alias of the Object Lambda access point in place of the bucket name. If the Object Lambda access point alias in a request is not valid, the error code \code{InvalidAccessPointAliasError} is returned. For more information about \code{InvalidAccessPointAliasError}, see \href{https://docs.aws.amazon.com/AmazonS3/latest/API/#ErrorCodeList}{List of Error Codes}.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you provide does not match the actual owner of the bucket, the request fails with the HTTP status code \verb{403 Forbidden} (access denied).}
}
\description{
Using the \code{\link[=s3_get_bucket_location]{get_bucket_location}} operation is no longer a best practice. To return the Region that a bucket resides in, we recommend that you use the \code{\link[=s3_head_bucket]{head_bucket}} operation instead. For backward compatibility, Amazon S3 continues to support the \code{\link[=s3_get_bucket_location]{get_bucket_location}} operation.

See \url{https://www.paws-r-sdk.com/docs/s3_get_bucket_location/} for full documentation.
}
\keyword{internal}
