% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_metadata_table_configuration}
\alias{s3_get_bucket_metadata_table_configuration}
\title{We recommend that you retrieve your S3 Metadata configurations by using
the V2 GetBucketMetadataTableConfiguration API operation}
\usage{
s3_get_bucket_metadata_table_configuration(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The general purpose bucket that corresponds to the metadata table
configuration that you want to retrieve.}

\item{ExpectedBucketOwner}{The expected owner of the general purpose bucket that you want to
retrieve the metadata table configuration for.}
}
\description{
We recommend that you retrieve your S3 Metadata configurations by using the V2 \code{\link[=s3_get_bucket_metadata_table_configuration]{get_bucket_metadata_table_configuration}} API operation. We no longer recommend using the V1 \code{\link[=s3_get_bucket_metadata_table_configuration]{get_bucket_metadata_table_configuration}} API operation.

See \url{https://www.paws-r-sdk.com/docs/s3_get_bucket_metadata_table_configuration/} for full documentation.
}
\keyword{internal}
