% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_versioning}
\alias{s3_get_bucket_versioning}
\title{Returns the versioning state of a bucket}
\usage{
s3_get_bucket_versioning(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The name of the bucket for which to get the versioning information.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Returns the versioning state of a bucket.

See \url{https://paws-r.github.io/docs/s3/get_bucket_versioning.html} for full documentation.
}
\keyword{internal}
