% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_list_object_annotations}
\alias{s3_list_object_annotations}
\title{Lists the annotations attached to an Amazon S3 object}
\usage{
s3_list_object_annotations(
  Bucket,
  Key,
  VersionId = NULL,
  MaxAnnotationResults = NULL,
  AnnotationPrefix = NULL,
  ContinuationToken = NULL,
  RequestPayer = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket that contains the object.}

\item{Key}{[required] The object key.}

\item{VersionId}{The version ID of the object.}

\item{MaxAnnotationResults}{The maximum number of annotations to return in the response. Maximum is 1,000.}

\item{AnnotationPrefix}{Filter results to annotations whose name begins with the specified prefix.}

\item{ContinuationToken}{Continuation token returned by a previous request to retrieve the next page.}

\item{RequestPayer}{Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. If either the source or destination S3 bucket has Requester Pays enabled, the requester will pay for the corresponding charges. For information about downloading objects from Requester Pays buckets, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/ObjectsinRequesterPaysBuckets.html}{Downloading Objects in Requester Pays Buckets} in the \emph{Amazon S3 User Guide}.

This functionality is not supported for directory buckets.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner.}
}
\description{
Lists the annotations attached to an Amazon S3 object. Results are paginated, with a maximum of 1,000 annotations per object. Use the \code{AnnotationPrefix} parameter to filter the results by name prefix.

See \url{https://www.paws-r-sdk.com/docs/s3_list_object_annotations/} for full documentation.
}
\keyword{internal}
