% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_intelligent_tiering_configuration}
\alias{s3_put_bucket_intelligent_tiering_configuration}
\title{Puts a S3 Intelligent-Tiering configuration to the specified bucket}
\usage{
s3_put_bucket_intelligent_tiering_configuration(
  Bucket,
  Id,
  IntelligentTieringConfiguration
)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose configuration you want to modify
or retrieve.}

\item{Id}{[required] The ID used to identify the S3 Intelligent-Tiering configuration.}

\item{IntelligentTieringConfiguration}{[required] Container for S3 Intelligent-Tiering configuration.}
}
\description{
Puts a S3 Intelligent-Tiering configuration to the specified bucket. You can have up to 1,000 S3 Intelligent-Tiering configurations per bucket.

See \url{https://paws-r.github.io/docs/s3/put_bucket_intelligent_tiering_configuration.html} for full documentation.
}
\keyword{internal}
