% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_object_annotation}
\alias{s3_put_object_annotation}
\title{Attaches an annotation to an Amazon S3 object}
\usage{
s3_put_object_annotation(
  Bucket,
  Key,
  VersionId = NULL,
  AnnotationName,
  AnnotationPayload,
  ObjectIfMatch = NULL,
  ChecksumAlgorithm = NULL,
  ChecksumCRC32 = NULL,
  ChecksumCRC32C = NULL,
  ChecksumCRC64NVME = NULL,
  ChecksumSHA1 = NULL,
  ChecksumSHA256 = NULL,
  ChecksumSHA512 = NULL,
  ChecksumMD5 = NULL,
  ChecksumXXHASH64 = NULL,
  ChecksumXXHASH3 = NULL,
  ChecksumXXHASH128 = NULL,
  ContentMD5 = NULL,
  RequestPayer = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket that contains the object.}

\item{Key}{[required] The object key.}

\item{VersionId}{The version ID of the object to attach the annotation to.}

\item{AnnotationName}{[required] The name of the annotation.

Length Constraints: Minimum length of 1. Maximum length of 512 bytes.}

\item{AnnotationPayload}{[required] The annotation payload. Must be between 1 byte and 1 MiB in size, and must be valid UTF-8 encoded text. If the payload contains invalid UTF-8 bytes, the request fails with HTTP 415 (Unsupported Media Type). To store binary data, encode the payload using Base64 before uploading.}

\item{ObjectIfMatch}{If specified, the operation only succeeds if the object's ETag matches the provided value.}

\item{ChecksumAlgorithm}{The checksum algorithm to use. Supported values: \code{CRC32}, \code{CRC32C}, \code{CRC64NVME}, \code{SHA1}, \code{SHA256}, \code{SHA512}, \code{MD5}, \code{XXHASH64}, \code{XXHASH3}, \code{XXHASH128}.}

\item{ChecksumCRC32}{Base64-encoded CRC32 checksum of the annotation payload.}

\item{ChecksumCRC32C}{Base64-encoded CRC32C checksum of the annotation payload.}

\item{ChecksumCRC64NVME}{Base64-encoded CRC64NVME checksum of the annotation payload.}

\item{ChecksumSHA1}{Base64-encoded SHA1 checksum of the annotation payload.}

\item{ChecksumSHA256}{Base64-encoded SHA256 checksum of the annotation payload.}

\item{ChecksumSHA512}{Base64-encoded SHA512 checksum of the annotation payload.}

\item{ChecksumMD5}{Base64-encoded MD5 checksum of the annotation payload.}

\item{ChecksumXXHASH64}{Base64-encoded XXHASH64 checksum of the annotation payload.}

\item{ChecksumXXHASH3}{Base64-encoded XXHASH3 checksum of the annotation payload.}

\item{ChecksumXXHASH128}{Base64-encoded XXHASH128 checksum of the annotation payload.}

\item{ContentMD5}{Base64-encoded MD5 digest of the message.}

\item{RequestPayer}{Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. If either the source or destination S3 bucket has Requester Pays enabled, the requester will pay for the corresponding charges. For information about downloading objects from Requester Pays buckets, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/ObjectsinRequesterPaysBuckets.html}{Downloading Objects in Requester Pays Buckets} in the \emph{Amazon S3 User Guide}.

This functionality is not supported for directory buckets.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a different account, the request fails with an HTTP 403 (Access Denied) error.}
}
\description{
Attaches an annotation to an Amazon S3 object. An annotation is a named payload of 1 byte to 1 MiB that you can associate with a specific object or object version. Each object can have up to 1,000 annotations.

See \url{https://www.paws-r-sdk.com/docs/s3_put_object_annotation/} for full documentation.
}
\keyword{internal}
