% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_object_lock_configuration}
\alias{s3_put_object_lock_configuration}
\title{Places an Object Lock configuration on the specified bucket}
\usage{
s3_put_object_lock_configuration(Bucket, ObjectLockConfiguration,
  RequestPayer, Token, ContentMD5, ExpectedBucketOwner)
}
\arguments{
\item{Bucket}{[required] The bucket whose Object Lock configuration you want to create or
replace.}

\item{ObjectLockConfiguration}{The Object Lock configuration that you want to apply to the specified
bucket.}

\item{RequestPayer}{}

\item{Token}{A token to allow Object Lock to be enabled for an existing bucket.}

\item{ContentMD5}{The MD5 hash for the request body.

For requests made using the AWS Command Line Interface (CLI) or AWS
SDKs, this field is calculated automatically.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request will fail with an HTTP
\verb{403 (Access Denied)} error.}
}
\value{
A list with the following syntax:\preformatted{list(
  RequestCharged = "requester"
)
}
}
\description{
Places an Object Lock configuration on the specified bucket. The rule
specified in the Object Lock configuration will be applied by default to
every new object placed in the specified bucket. For more information,
see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-lock.html}{Locking Objects}.
\itemize{
\item The \code{DefaultRetention} settings require both a mode and a period.
\item The \code{DefaultRetention} period can be either \code{Days} or \code{Years} but
you must select one. You cannot specify \code{Days} and \code{Years} at the
same time.
\item You can only enable Object Lock for new buckets. If you want to turn
on Object Lock for an existing bucket, contact AWS Support.
}
}
\section{Request syntax}{
\preformatted{svc$put_object_lock_configuration(
  Bucket = "string",
  ObjectLockConfiguration = list(
    ObjectLockEnabled = "Enabled",
    Rule = list(
      DefaultRetention = list(
        Mode = "GOVERNANCE"|"COMPLIANCE",
        Days = 123,
        Years = 123
      )
    )
  ),
  RequestPayer = "requester",
  Token = "string",
  ContentMD5 = "string",
  ExpectedBucketOwner = "string"
)
}
}

\keyword{internal}
