% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_rename_object}
\alias{s3_rename_object}
\title{Renames an existing object in a directory bucket that uses the S3
Express One Zone storage class}
\usage{
s3_rename_object(
  Bucket,
  Key,
  RenameSource,
  DestinationIfMatch = NULL,
  DestinationIfNoneMatch = NULL,
  DestinationIfModifiedSince = NULL,
  DestinationIfUnmodifiedSince = NULL,
  SourceIfMatch = NULL,
  SourceIfNoneMatch = NULL,
  SourceIfModifiedSince = NULL,
  SourceIfUnmodifiedSince = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{Bucket}{[required] The bucket name of the directory bucket containing the object.

You must use virtual-hosted-style requests in the format
\code{Bucket-name.s3express-zone-id.region-code.amazonaws.com}. Path-style
requests are not supported. Directory bucket names must be unique in the
chosen Availability Zone. Bucket names must follow the format
\code{bucket-base-name--zone-id--x-s3 } (for example,
\code{amzn-s3-demo-bucket--usw2-az1--x-s3}). For information about bucket
naming restrictions, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/directory-bucket-naming-rules.html}{Directory bucket naming rules}
in the \emph{Amazon S3 User Guide}.}

\item{Key}{[required] Key name of the object to rename.}

\item{RenameSource}{[required] Specifies the source for the rename operation. The value must be URL
encoded.}

\item{DestinationIfMatch}{Renames the object only if the ETag (entity tag) value provided during
the operation matches the ETag of the object in S3. The \code{If-Match}
header field makes the request method conditional on ETags. If the ETag
values do not match, the operation returns a \verb{412 Precondition Failed}
error.

Expects the ETag value as a string.}

\item{DestinationIfNoneMatch}{Renames the object only if the destination does not already exist in the
specified directory bucket. If the object does exist when you send a
request with \verb{If-None-Match:*}, the S3 API will return a
\verb{412 Precondition Failed} error, preventing an overwrite. The
\code{If-None-Match} header prevents overwrites of existing data by
validating that there's not an object with the same key name already in
your directory bucket.

Expects the \code{*} character (asterisk).}

\item{DestinationIfModifiedSince}{Renames the object if the destination exists and if it has been modified
since the specified time.}

\item{DestinationIfUnmodifiedSince}{Renames the object if it hasn't been modified since the specified time.}

\item{SourceIfMatch}{Renames the object if the source exists and if its entity tag (ETag)
matches the specified ETag.}

\item{SourceIfNoneMatch}{Renames the object if the source exists and if its entity tag (ETag) is
different than the specified ETag. If an asterisk (\code{*}) character is
provided, the operation will fail and return a \verb{412 Precondition Failed}
error.}

\item{SourceIfModifiedSince}{Renames the object if the source exists and if it has been modified
since the specified time.}

\item{SourceIfUnmodifiedSince}{Renames the object if the source exists and hasn't been modified since
the specified time.}

\item{ClientToken}{A unique string with a max of 64 ASCII characters in the ASCII range of
33 - 126.

\code{\link[=s3_rename_object]{rename_object}} supports idempotency using a client
token. To make an idempotent API request using
\code{\link[=s3_rename_object]{rename_object}}, specify a client token in the
request. You should not reuse the same client token for other API
requests. If you retry a request that completed successfully using the
same client token and the same parameters, the retry succeeds without
performing any further actions. If you retry a successful request using
the same client token, but one or more of the parameters are different,
the retry fails and an \code{IdempotentParameterMismatch} error is returned.}
}
\description{
Renames an existing object in a directory bucket that uses the S3 Express One Zone storage class. You can use \code{\link[=s3_rename_object]{rename_object}} by specifying an existing object’s name as the source and the new name of the object as the destination within the same directory bucket.

See \url{https://www.paws-r-sdk.com/docs/s3_rename_object/} for full documentation.
}
\keyword{internal}
