% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_update_bucket_metadata_annotation_table_configuration}
\alias{s3_update_bucket_metadata_annotation_table_configuration}
\title{Updates the annotation table configuration for an Amazon S3 bucket's
metadata configuration}
\usage{
s3_update_bucket_metadata_annotation_table_configuration(
  Bucket,
  ContentMD5 = NULL,
  ChecksumAlgorithm = NULL,
  AnnotationTableConfiguration,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket whose annotation table configuration to update.}

\item{ContentMD5}{Base64-encoded MD5 digest of the message body.}

\item{ChecksumAlgorithm}{Checksum algorithm for the request payload.}

\item{AnnotationTableConfiguration}{[required] The annotation table configuration updates to apply.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner.}
}
\description{
Updates the annotation table configuration for an Amazon S3 bucket's metadata configuration. Use this operation to enable or disable the annotation table, or to update its associated IAM role.

See \url{https://www.paws-r-sdk.com/docs/s3_update_bucket_metadata_annotation_table_configuration/} for full documentation.
}
\keyword{internal}
