% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_update_bucket_metadata_inventory_table_configuration}
\alias{s3_update_bucket_metadata_inventory_table_configuration}
\title{Enables or disables a live inventory table for an S3 Metadata
configuration on a general purpose bucket}
\usage{
s3_update_bucket_metadata_inventory_table_configuration(
  Bucket,
  ContentMD5 = NULL,
  ChecksumAlgorithm = NULL,
  InventoryTableConfiguration,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The general purpose bucket that corresponds to the metadata
configuration that you want to enable or disable an inventory table for.}

\item{ContentMD5}{The \code{Content-MD5} header for the inventory table configuration.}

\item{ChecksumAlgorithm}{The checksum algorithm to use with your inventory table configuration.}

\item{InventoryTableConfiguration}{[required] The contents of your inventory table configuration.}

\item{ExpectedBucketOwner}{The expected owner of the general purpose bucket that corresponds to the
metadata table configuration that you want to enable or disable an
inventory table for.}
}
\description{
Enables or disables a live inventory table for an S3 Metadata configuration on a general purpose bucket. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/metadata-tables-overview.html}{Accelerating data discovery with S3 Metadata} in the \emph{Amazon S3 User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3_update_bucket_metadata_inventory_table_configuration/} for full documentation.
}
\keyword{internal}
