% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_update_object_encryption}
\alias{s3_update_object_encryption}
\title{This operation is not supported for directory buckets or Amazon S3 on
Outposts buckets}
\usage{
s3_update_object_encryption(
  Bucket,
  Key,
  VersionId = NULL,
  ObjectEncryption,
  RequestPayer = NULL,
  ExpectedBucketOwner = NULL,
  ContentMD5 = NULL,
  ChecksumAlgorithm = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the general purpose bucket that contains the specified
object key name.

When you use this operation with an access point attached to a general
purpose bucket, you must either provide the alias of the access point in
place of the bucket name or you must specify the access point Amazon
Resource Name (ARN). When using the access point ARN, you must direct
requests to the access point hostname. The access point hostname takes
the form
\code{ AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com}. When
using this operation with an access point through the Amazon Web
Services SDKs, you provide the access point ARN in place of the bucket
name. For more information about access point ARNs, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/access-points-naming.html}{Referencing access points}
in the \emph{Amazon S3 User Guide}.}

\item{Key}{[required] The key name of the object that you want to update the server-side
encryption type for.}

\item{VersionId}{The version ID of the object that you want to update the server-side
encryption type for.}

\item{ObjectEncryption}{[required] The updated server-side encryption type for this object. The
\code{\link[=s3_update_object_encryption]{update_object_encryption}} operation
supports the SSE-S3 and SSE-KMS encryption types.

Valid Values: \code{SSES3} | \code{SSEKMS}}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide doesn't match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}

\item{ContentMD5}{The MD5 hash for the request body. For requests made using the Amazon
Web Services Command Line Interface (CLI) or Amazon Web Services SDKs,
this field is calculated automatically.}

\item{ChecksumAlgorithm}{Indicates the algorithm used to create the checksum for the object when
you use an Amazon Web Services SDK. This header doesn't provide any
additional functionality if you don't use the SDK. When you send this
header, there must be a corresponding \code{x-amz-checksum} or
\code{x-amz-trailer} header sent. Otherwise, Amazon S3 fails the request with
the HTTP status code \verb{400 Bad Request}. For more information, see
\href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/checking-object-integrity.html}{Checking object integrity}
in the \emph{Amazon S3 User Guide}.

If you provide an individual checksum, Amazon S3 ignores any provided
\code{ChecksumAlgorithm} parameter.}
}
\description{
This operation is not supported for directory buckets or Amazon S3 on Outposts buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_update_object_encryption/} for full documentation.
}
\keyword{internal}
