% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_create_access_point_for_object_lambda}
\alias{s3control_create_access_point_for_object_lambda}
\title{Creates an Object Lambda Access Point}
\usage{
s3control_create_access_point_for_object_lambda(AccountId, Name,
  Configuration)
}
\arguments{
\item{AccountId}{[required] The AWS account ID for owner of the specified Object Lambda Access
Point.}

\item{Name}{[required] The name you want to assign to this Object Lambda Access Point.}

\item{Configuration}{[required] Object Lambda Access Point configuration as a JSON document.}
}
\value{
A list with the following syntax:\preformatted{list(
  ObjectLambdaAccessPointArn = "string"
)
}
}
\description{
Creates an Object Lambda Access Point. For more information, see
\href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/transforming-objects.html}{Transforming objects with Object Lambda Access Points}
in the \emph{Amazon S3 User Guide}.

The following actions are related to
\code{\link[=s3control_create_access_point_for_object_lambda]{create_access_point_for_object_lambda}}:
\itemize{
\item \code{\link[=s3control_delete_access_point_for_object_lambda]{delete_access_point_for_object_lambda}}
\item \code{\link[=s3control_get_access_point_for_object_lambda]{get_access_point_for_object_lambda}}
\item \code{\link[=s3control_list_access_points_for_object_lambda]{list_access_points_for_object_lambda}}
}
}
\section{Request syntax}{
\preformatted{svc$create_access_point_for_object_lambda(
  AccountId = "string",
  Name = "string",
  Configuration = list(
    SupportingAccessPoint = "string",
    CloudWatchMetricsEnabled = TRUE|FALSE,
    AllowedFeatures = list(
      "GetObject-Range"|"GetObject-PartNumber"
    ),
    TransformationConfigurations = list(
      list(
        Actions = list(
          "GetObject"
        ),
        ContentTransformation = list(
          AwsLambda = list(
            FunctionArn = "string",
            FunctionPayload = "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
