% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_delete_access_grants_instance_resource_policy}
\alias{s3control_delete_access_grants_instance_resource_policy}
\title{Deletes the resource policy of the S3 Access Grants instance}
\usage{
s3control_delete_access_grants_instance_resource_policy(AccountId)
}
\arguments{
\item{AccountId}{[required] The ID of the Amazon Web Services account that is making this request.}
}
\description{
Deletes the resource policy of the S3 Access Grants instance. The resource policy is used to manage cross-account access to your S3 Access Grants instance. By deleting the resource policy, you delete any cross-account permissions to your S3 Access Grants instance.

See \url{https://www.paws-r-sdk.com/docs/s3control_delete_access_grants_instance_resource_policy/} for full documentation.
}
\keyword{internal}
