% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_delete_access_point_for_object_lambda}
\alias{s3control_delete_access_point_for_object_lambda}
\title{Deletes the specified Object Lambda Access Point}
\usage{
s3control_delete_access_point_for_object_lambda(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{Name}{[required] The name of the access point you want to delete.}
}
\description{
Deletes the specified Object Lambda Access Point.

See \url{https://www.paws-r-sdk.com/docs/s3control_delete_access_point_for_object_lambda/} for full documentation.
}
\keyword{internal}
