% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_delete_access_point_scope}
\alias{s3control_delete_access_point_scope}
\title{Deletes an existing access point scope for a directory bucket}
\usage{
s3control_delete_access_point_scope(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the access point with the
scope that you want to delete.}

\item{Name}{[required] The name of the access point with the scope that you want to delete.}
}
\description{
Deletes an existing access point scope for a directory bucket.

See \url{https://www.paws-r-sdk.com/docs/s3control_delete_access_point_scope/} for full documentation.
}
\keyword{internal}
