% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_delete_multi_region_access_point}
\alias{s3control_delete_multi_region_access_point}
\title{Deletes a Multi-Region Access Point}
\usage{
s3control_delete_multi_region_access_point(AccountId, ClientToken, Details)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{ClientToken}{[required] An idempotency token used to identify the request and guarantee that
requests are unique.}

\item{Details}{[required] A container element containing details about the Multi-Region Access
Point.}
}
\description{
Deletes a Multi-Region Access Point. This action does not delete the buckets associated with the Multi-Region Access Point, only the Multi-Region Access Point itself.

See \url{https://paws-r.github.io/docs/s3control/delete_multi_region_access_point.html} for full documentation.
}
\keyword{internal}
