% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_delete_storage_lens_group}
\alias{s3control_delete_storage_lens_group}
\title{Deletes an existing S3 Storage Lens group}
\usage{
s3control_delete_storage_lens_group(Name, AccountId)
}
\arguments{
\item{Name}{[required] The name of the Storage Lens group that you're trying to delete.}

\item{AccountId}{[required] The Amazon Web Services account ID used to create the Storage Lens group
that you're trying to delete.}
}
\description{
Deletes an existing S3 Storage Lens group.

See \url{https://www.paws-r-sdk.com/docs/s3control_delete_storage_lens_group/} for full documentation.
}
\keyword{internal}
