% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_dissociate_access_grants_identity_center}
\alias{s3control_dissociate_access_grants_identity_center}
\title{Dissociates the Amazon Web Services IAM Identity Center instance from
the S3 Access Grants instance}
\usage{
s3control_dissociate_access_grants_identity_center(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID of the S3 Access Grants instance.}
}
\description{
Dissociates the Amazon Web Services IAM Identity Center instance from the S3 Access Grants instance.

See \url{https://www.paws-r-sdk.com/docs/s3control_dissociate_access_grants_identity_center/} for full documentation.
}
\keyword{internal}
