% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_grants_instance}
\alias{s3control_get_access_grants_instance}
\title{Retrieves the S3 Access Grants instance for a Region in your account}
\usage{
s3control_get_access_grants_instance(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID of the S3 Access Grants instance.}
}
\description{
Retrieves the S3 Access Grants instance for a Region in your account.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_access_grants_instance/} for full documentation.
}
\keyword{internal}
