% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_grants_instance_for_prefix}
\alias{s3control_get_access_grants_instance_for_prefix}
\title{Retrieve the S3 Access Grants instance that contains a particular prefix}
\usage{
s3control_get_access_grants_instance_for_prefix(AccountId, S3Prefix)
}
\arguments{
\item{AccountId}{[required] The ID of the Amazon Web Services account that is making this request.}

\item{S3Prefix}{[required] The S3 prefix of the access grants that you would like to retrieve.}
}
\description{
Retrieve the S3 Access Grants instance that contains a particular prefix.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_access_grants_instance_for_prefix/} for full documentation.
}
\keyword{internal}
