% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_grants_instance_resource_policy}
\alias{s3control_get_access_grants_instance_resource_policy}
\title{Returns the resource policy of the S3 Access Grants instance}
\usage{
s3control_get_access_grants_instance_resource_policy(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID of the S3 Access Grants instance.}
}
\description{
Returns the resource policy of the S3 Access Grants instance.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_access_grants_instance_resource_policy/} for full documentation.
}
\keyword{internal}
