% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_point_for_object_lambda}
\alias{s3control_get_access_point_for_object_lambda}
\title{Returns configuration information about the specified Object Lambda
Access Point}
\usage{
s3control_get_access_point_for_object_lambda(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{Name}{[required] The name of the Object Lambda Access Point.}
}
\value{
A list with the following syntax:\preformatted{list(
  Name = "string",
  PublicAccessBlockConfiguration = list(
    BlockPublicAcls = TRUE|FALSE,
    IgnorePublicAcls = TRUE|FALSE,
    BlockPublicPolicy = TRUE|FALSE,
    RestrictPublicBuckets = TRUE|FALSE
  ),
  CreationDate = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Returns configuration information about the specified Object Lambda
Access Point

The following actions are related to
\code{\link[=s3control_get_access_point_for_object_lambda]{get_access_point_for_object_lambda}}:
\itemize{
\item \code{\link[=s3control_create_access_point_for_object_lambda]{create_access_point_for_object_lambda}}
\item \code{\link[=s3control_delete_access_point_for_object_lambda]{delete_access_point_for_object_lambda}}
\item \code{\link[=s3control_list_access_points_for_object_lambda]{list_access_points_for_object_lambda}}
}
}
\section{Request syntax}{
\preformatted{svc$get_access_point_for_object_lambda(
  AccountId = "string",
  Name = "string"
)
}
}

\keyword{internal}
