% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_point_policy_status}
\alias{s3control_get_access_point_policy_status}
\title{Indicates whether the specified access point currently has a policy that
allows public access}
\usage{
s3control_get_access_point_policy_status(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified access point.}

\item{Name}{[required] The name of the access point whose policy status you want to retrieve.}
}
\description{
Indicates whether the specified access point currently has a policy that allows public access. For more information about public access through access points, see Managing Data Access with Amazon S3 access points in the \emph{Amazon S3 User Guide}.

See \url{https://paws-r.github.io/docs/s3control/get_access_point_policy_status.html} for full documentation.
}
\keyword{internal}
