% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_point_policy_status_for_object_lambda}
\alias{s3control_get_access_point_policy_status_for_object_lambda}
\title{Returns the status of the resource policy associated with an Object
Lambda Access Point}
\usage{
s3control_get_access_point_policy_status_for_object_lambda(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{Name}{[required] The name of the Object Lambda Access Point.}
}
\description{
Returns the status of the resource policy associated with an Object Lambda Access Point.

See \url{https://paws-r.github.io/docs/s3control/get_access_point_policy_status_for_object_lambda.html} for full documentation.
}
\keyword{internal}
