% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_multi_region_access_point}
\alias{s3control_get_multi_region_access_point}
\title{Returns configuration information about the specified Multi-Region
Access Point}
\usage{
s3control_get_multi_region_access_point(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{Name}{[required] The name of the Multi-Region Access Point whose configuration
information you want to receive. The name of the Multi-Region Access
Point is different from the alias. For more information about the
distinction between the name and the alias of an Multi-Region Access
Point, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/CreatingMultiRegionAccessPoints.html#multi-region-access-point-naming}{Managing Multi-Region Access Points}
in the \emph{Amazon S3 User Guide}.}
}
\description{
Returns configuration information about the specified Multi-Region Access Point.

See \url{https://paws-r.github.io/docs/s3control/get_multi_region_access_point.html} for full documentation.
}
\keyword{internal}
