% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_multi_region_access_point_policy_status}
\alias{s3control_get_multi_region_access_point_policy_status}
\title{Indicates whether the specified Multi-Region Access Point has an access
control policy that allows public access}
\usage{
s3control_get_multi_region_access_point_policy_status(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{Name}{[required] Specifies the Multi-Region Access Point. The name of the Multi-Region
Access Point is different from the alias. For more information about the
distinction between the name and the alias of an Multi-Region Access
Point, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/CreatingMultiRegionAccessPoints.html#multi-region-access-point-naming}{Managing Multi-Region Access Points}
in the \emph{Amazon S3 User Guide}.}
}
\description{
Indicates whether the specified Multi-Region Access Point has an access control policy that allows public access.

See \url{https://paws-r.github.io/docs/s3control/get_multi_region_access_point_policy_status.html} for full documentation.
}
\keyword{internal}
