% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_multi_region_access_point_routes}
\alias{s3control_get_multi_region_access_point_routes}
\title{Returns the routing configuration for a Multi-Region Access Point,
indicating which Regions are active or passive}
\usage{
s3control_get_multi_region_access_point_routes(AccountId, Mrap)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{Mrap}{[required] The Multi-Region Access Point ARN.}
}
\description{
Returns the routing configuration for a Multi-Region Access Point, indicating which Regions are active or passive.

See \url{https://paws-r.github.io/docs/s3control/get_multi_region_access_point_routes.html} for full documentation.
}
\keyword{internal}
