% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_storage_lens_group}
\alias{s3control_get_storage_lens_group}
\title{Retrieves the Storage Lens group configuration details}
\usage{
s3control_get_storage_lens_group(Name, AccountId)
}
\arguments{
\item{Name}{[required] The name of the Storage Lens group that you're trying to retrieve the
configuration details for.}

\item{AccountId}{[required] The Amazon Web Services account ID associated with the Storage Lens
group that you're trying to retrieve the details for.}
}
\description{
Retrieves the Storage Lens group configuration details.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_storage_lens_group/} for full documentation.
}
\keyword{internal}
