% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_access_points_for_directory_buckets}
\alias{s3control_list_access_points_for_directory_buckets}
\title{Returns a list of the access points that are owned by the Amazon Web
Services account and that are associated with the specified directory
bucket}
\usage{
s3control_list_access_points_for_directory_buckets(
  AccountId,
  DirectoryBucket = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the access points.}

\item{DirectoryBucket}{The name of the directory bucket associated with the access points you
want to list.}

\item{NextToken}{If \code{NextToken} is returned, there are more access points available than
requested in the \code{maxResults} value. The value of \code{NextToken} is a
unique pagination token for each page. Make the call again using the
returned token to retrieve the next page. Keep all other arguments
unchanged. Each pagination token expires after 24 hours.}

\item{MaxResults}{The maximum number of access points that you would like returned in the
\code{\link[=s3control_list_access_points_for_directory_buckets]{list_access_points_for_directory_buckets}}
response. If the directory bucket is associated with more than this
number of access points, the results include the pagination token
\code{NextToken}. Make another call using the \code{NextToken} to retrieve more
results.}
}
\description{
Returns a list of the access points that are owned by the Amazon Web Services account and that are associated with the specified directory bucket.

See \url{https://www.paws-r-sdk.com/docs/s3control_list_access_points_for_directory_buckets/} for full documentation.
}
\keyword{internal}
