% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_access_points_for_object_lambda}
\alias{s3control_list_access_points_for_object_lambda}
\title{Returns a list of the access points associated with the Object Lambda
Access Point}
\usage{
s3control_list_access_points_for_object_lambda(AccountId, NextToken,
  MaxResults)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{NextToken}{If the list has more access points than can be returned in one call to
this API, this field contains a continuation token that you can provide
in subsequent calls to this API to retrieve additional access points.}

\item{MaxResults}{The maximum number of access points that you want to include in the
list. If there are more than this number of access points, then the
response will include a continuation token in the \code{NextToken} field that
you can use to retrieve the next page of access points.}
}
\value{
A list with the following syntax:\preformatted{list(
  ObjectLambdaAccessPointList = list(
    list(
      Name = "string",
      ObjectLambdaAccessPointArn = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of the access points associated with the Object Lambda
Access Point. You can retrieve up to 1000 access points per call. If
there are more than 1,000 access points (or the number specified in
\code{maxResults}, whichever is less), the response will include a
continuation token that you can use to list the additional access
points.

The following actions are related to
\code{\link[=s3control_list_access_points_for_object_lambda]{list_access_points_for_object_lambda}}:
\itemize{
\item \code{\link[=s3control_create_access_point_for_object_lambda]{create_access_point_for_object_lambda}}
\item \code{\link[=s3control_delete_access_point_for_object_lambda]{delete_access_point_for_object_lambda}}
\item \code{\link[=s3control_get_access_point_for_object_lambda]{get_access_point_for_object_lambda}}
}
}
\section{Request syntax}{
\preformatted{svc$list_access_points_for_object_lambda(
  AccountId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
