% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_multi_region_access_points}
\alias{s3control_list_multi_region_access_points}
\title{Returns a list of the Multi-Region Access Points currently associated
with the specified Amazon Web Services account}
\usage{
s3control_list_multi_region_access_points(
  AccountId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{NextToken}{Not currently used. Do not use this parameter.}

\item{MaxResults}{Not currently used. Do not use this parameter.}
}
\description{
Returns a list of the Multi-Region Access Points currently associated with the specified Amazon Web Services account. Each call can return up to 100 Multi-Region Access Points, the maximum number of Multi-Region Access Points that can be associated with a single account.

See \url{https://paws-r.github.io/docs/s3control/list_multi_region_access_points.html} for full documentation.
}
\keyword{internal}
