% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_storage_lens_configurations}
\alias{s3control_list_storage_lens_configurations}
\title{This operation is not supported by directory buckets}
\usage{
s3control_list_storage_lens_configurations(AccountId, NextToken = NULL)
}
\arguments{
\item{AccountId}{[required] The account ID of the requester.}

\item{NextToken}{A pagination token to request the next page of results.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_list_storage_lens_configurations/} for full documentation.
}
\keyword{internal}
