% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_storage_lens_groups}
\alias{s3control_list_storage_lens_groups}
\title{Lists all the Storage Lens groups in the specified home Region}
\usage{
s3control_list_storage_lens_groups(AccountId, NextToken = NULL)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the Storage Lens groups.}

\item{NextToken}{The token for the next set of results, or \code{null} if there are no more
results.}
}
\description{
Lists all the Storage Lens groups in the specified home Region.

See \url{https://www.paws-r-sdk.com/docs/s3control_list_storage_lens_groups/} for full documentation.
}
\keyword{internal}
