% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_access_grants_instance_resource_policy}
\alias{s3control_put_access_grants_instance_resource_policy}
\title{Updates the resource policy of the S3 Access Grants instance}
\usage{
s3control_put_access_grants_instance_resource_policy(
  AccountId,
  Policy,
  Organization = NULL
)
}
\arguments{
\item{AccountId}{[required] The ID of the Amazon Web Services account that is making this request.}

\item{Policy}{[required] The resource policy of the S3 Access Grants instance that you are
updating.}

\item{Organization}{The Organization of the resource policy of the S3 Access Grants
instance.}
}
\description{
Updates the resource policy of the S3 Access Grants instance.

See \url{https://www.paws-r-sdk.com/docs/s3control_put_access_grants_instance_resource_policy/} for full documentation.
}
\keyword{internal}
