% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_access_point_configuration_for_object_lambda}
\alias{s3control_put_access_point_configuration_for_object_lambda}
\title{Replaces configuration for an Object Lambda Access Point}
\usage{
s3control_put_access_point_configuration_for_object_lambda(AccountId,
  Name, Configuration)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{Name}{[required] The name of the Object Lambda Access Point.}

\item{Configuration}{[required] Object Lambda Access Point configuration document.}
}
\value{
An empty list.
}
\description{
Replaces configuration for an Object Lambda Access Point.

The following actions are related to
\code{\link[=s3control_put_access_point_configuration_for_object_lambda]{put_access_point_configuration_for_object_lambda}}:
\itemize{
\item \code{\link[=s3control_get_access_point_configuration_for_object_lambda]{get_access_point_configuration_for_object_lambda}}
}
}
\section{Request syntax}{
\preformatted{svc$put_access_point_configuration_for_object_lambda(
  AccountId = "string",
  Name = "string",
  Configuration = list(
    SupportingAccessPoint = "string",
    CloudWatchMetricsEnabled = TRUE|FALSE,
    AllowedFeatures = list(
      "GetObject-Range"|"GetObject-PartNumber"
    ),
    TransformationConfigurations = list(
      list(
        Actions = list(
          "GetObject"
        ),
        ContentTransformation = list(
          AwsLambda = list(
            FunctionArn = "string",
            FunctionPayload = "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
