% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_access_point_scope}
\alias{s3control_put_access_point_scope}
\title{Creates or replaces the access point scope for a directory bucket}
\usage{
s3control_put_access_point_scope(AccountId, Name, Scope)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the access point with scope
that you want to create or replace.}

\item{Name}{[required] The name of the access point with the scope that you want to create or
replace.}

\item{Scope}{[required] Object prefixes, API operations, or a combination of both.}
}
\description{
Creates or replaces the access point scope for a directory bucket. You can use the access point scope to restrict access to specific prefixes, API operations, or a combination of both.

See \url{https://www.paws-r-sdk.com/docs/s3control_put_access_point_scope/} for full documentation.
}
\keyword{internal}
