% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_bucket_lifecycle_configuration}
\alias{s3control_put_bucket_lifecycle_configuration}
\title{This action puts a lifecycle configuration to an Amazon S3 on Outposts
bucket}
\usage{
s3control_put_bucket_lifecycle_configuration(
  AccountId,
  Bucket,
  LifecycleConfiguration = NULL
)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID of the Outposts bucket.}

\item{Bucket}{[required] The name of the bucket for which to set the configuration.}

\item{LifecycleConfiguration}{Container for lifecycle rules. You can add as many as 1,000 rules.}
}
\description{
This action puts a lifecycle configuration to an Amazon S3 on Outposts bucket. To put a lifecycle configuration to an S3 bucket, see \code{\link[=s3control_put_bucket_lifecycle_configuration]{put_bucket_lifecycle_configuration}} in the \emph{Amazon S3 API Reference}.

See \url{https://paws-r.github.io/docs/s3control/put_bucket_lifecycle_configuration.html} for full documentation.
}
\keyword{internal}
