% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_job_tagging}
\alias{s3control_put_job_tagging}
\title{Sets the supplied tag-set on an S3 Batch Operations job}
\usage{
s3control_put_job_tagging(AccountId, JobId, Tags)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID associated with the S3 Batch
Operations job.}

\item{JobId}{[required] The ID for the S3 Batch Operations job whose tags you want to replace.}

\item{Tags}{[required] The set of tags to associate with the S3 Batch Operations job.}
}
\description{
Sets the supplied tag-set on an S3 Batch Operations job.

See \url{https://www.paws-r-sdk.com/docs/s3control_put_job_tagging/} for full documentation.
}
\keyword{internal}
