% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_multi_region_access_point_policy}
\alias{s3control_put_multi_region_access_point_policy}
\title{Associates an access control policy with the specified Multi-Region
Access Point}
\usage{
s3control_put_multi_region_access_point_policy(AccountId, ClientToken, Details)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{ClientToken}{[required] An idempotency token used to identify the request and guarantee that
requests are unique.}

\item{Details}{[required] A container element containing the details of the policy for the
Multi-Region Access Point.}
}
\description{
Associates an access control policy with the specified Multi-Region Access Point. Each Multi-Region Access Point can have only one policy, so a request made to this action replaces any existing policy that is associated with the specified Multi-Region Access Point.

See \url{https://paws-r.github.io/docs/s3control/put_multi_region_access_point_policy.html} for full documentation.
}
\keyword{internal}
