% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_public_access_block}
\alias{s3control_put_public_access_block}
\title{This operation is not supported by directory buckets}
\usage{
s3control_put_public_access_block(PublicAccessBlockConfiguration, AccountId)
}
\arguments{
\item{PublicAccessBlockConfiguration}{[required] The \code{PublicAccessBlock} configuration that you want to apply to the
specified Amazon Web Services account.}

\item{AccountId}{[required] The account ID for the Amazon Web Services account whose
\code{PublicAccessBlock} configuration you want to set.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_put_public_access_block/} for full documentation.
}
\keyword{internal}
