% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_storage_lens_configuration_tagging}
\alias{s3control_put_storage_lens_configuration_tagging}
\title{Put or replace tags on an existing Amazon S3 Storage Lens configuration}
\usage{
s3control_put_storage_lens_configuration_tagging(ConfigId, AccountId,
  Tags)
}
\arguments{
\item{ConfigId}{[required] The ID of the S3 Storage Lens configuration.}

\item{AccountId}{[required] The account ID of the requester.}

\item{Tags}{[required] The tag set of the S3 Storage Lens configuration.

You can set up to a maximum of 50 tags.}
}
\value{
An empty list.
}
\description{
Put or replace tags on an existing Amazon S3 Storage Lens configuration.
For more information about S3 Storage Lens, see Working with Amazon S3
Storage Lens in the \emph{Amazon Simple Storage Service Developer Guide}.

To use this action, you must have permission to perform the
\code{s3:PutStorageLensConfigurationTagging} action. For more information,
see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage_lens.html#storage_lens_IAM}{Setting permissions to use Amazon S3 Storage Lens}
in the \emph{Amazon Simple Storage Service Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_storage_lens_configuration_tagging(
  ConfigId = "string",
  AccountId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
