% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_untag_resource}
\alias{s3control_untag_resource}
\title{This operation removes the specified Amazon Web Services resource tags
from an S3 resource}
\usage{
s3control_untag_resource(AccountId, ResourceArn, TagKeys)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the resource that you're
trying to remove the tags from.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the S3 resource that you're trying to
remove the tags from.}

\item{TagKeys}{[required] The array of tag key-value pairs that you're trying to remove from of
the S3 resource.}
}
\description{
This operation removes the specified Amazon Web Services resource tags from an S3 resource. Each tag is a label consisting of a user-defined key and value. Tags can help you manage, identify, organize, search for, and filter resources.

See \url{https://www.paws-r-sdk.com/docs/s3control_untag_resource/} for full documentation.
}
\keyword{internal}
