% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_update_storage_lens_group}
\alias{s3control_update_storage_lens_group}
\title{Updates the existing Storage Lens group}
\usage{
s3control_update_storage_lens_group(Name, AccountId, StorageLensGroup)
}
\arguments{
\item{Name}{[required] The name of the Storage Lens group that you want to update.}

\item{AccountId}{[required] The Amazon Web Services account ID of the Storage Lens group owner.}

\item{StorageLensGroup}{[required] The JSON file that contains the Storage Lens group configuration.}
}
\description{
Updates the existing Storage Lens group.

See \url{https://www.paws-r-sdk.com/docs/s3control_update_storage_lens_group/} for full documentation.
}
\keyword{internal}
