% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_operations.R
\name{s3outposts_delete_endpoint}
\alias{s3outposts_delete_endpoint}
\title{Deletes an endpoint}
\usage{
s3outposts_delete_endpoint(EndpointId, OutpostId)
}
\arguments{
\item{EndpointId}{[required] The ID of the endpoint.}

\item{OutpostId}{[required] The ID of the Outposts.}
}
\description{
Deletes an endpoint.

See \url{https://www.paws-r-sdk.com/docs/s3outposts_delete_endpoint/} for full documentation.
}
\keyword{internal}
