% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_operations.R
\name{s3outposts_list_endpoints}
\alias{s3outposts_list_endpoints}
\title{Lists endpoints associated with the specified Outpost}
\usage{
s3outposts_list_endpoints(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If a previous response from this operation included a \code{NextToken} value,
provide that value here to retrieve the next page of results.}

\item{MaxResults}{The maximum number of endpoints that will be returned in the response.}
}
\description{
Lists endpoints associated with the specified Outpost.

See \url{https://paws-r.github.io/docs/s3outposts/list_endpoints.html} for full documentation.
}
\keyword{internal}
