% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_operations.R
\name{s3outposts_list_outposts_with_s3}
\alias{s3outposts_list_outposts_with_s3}
\title{Lists the Outposts with S3 on Outposts capacity for your Amazon Web
Services account}
\usage{
s3outposts_list_outposts_with_s3(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{When you can get additional results from the
\code{\link[=s3outposts_list_outposts_with_s3]{list_outposts_with_s3}} call, a
\code{NextToken} parameter is returned in the output. You can then pass in a
subsequent command to the \code{NextToken} parameter to continue listing
additional Outposts.}

\item{MaxResults}{The maximum number of Outposts to return. The limit is 100.}
}
\description{
Lists the Outposts with S3 on Outposts capacity for your Amazon Web Services account. Includes S3 on Outposts that you have access to as the Outposts owner, or as a shared user from Resource Access Manager (RAM).

See \url{https://www.paws-r-sdk.com/docs/s3outposts_list_outposts_with_s3/} for full documentation.
}
\keyword{internal}
