% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3outposts_operations.R
\name{s3outposts_list_shared_endpoints}
\alias{s3outposts_list_shared_endpoints}
\title{Lists all endpoints associated with an Outpost that has been shared by
Amazon Web Services Resource Access Manager (RAM)}
\usage{
s3outposts_list_shared_endpoints(
  NextToken = NULL,
  MaxResults = NULL,
  OutpostId
)
}
\arguments{
\item{NextToken}{If a previous response from this operation included a \code{NextToken} value,
you can provide that value here to retrieve the next page of results.}

\item{MaxResults}{The maximum number of endpoints that will be returned in the response.}

\item{OutpostId}{[required] The ID of the Amazon Web Services Outpost.}
}
\description{
Lists all endpoints associated with an Outpost that has been shared by Amazon Web Services Resource Access Manager (RAM).

See \url{https://www.paws-r-sdk.com/docs/s3outposts_list_shared_endpoints/} for full documentation.
}
\keyword{internal}
