% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_create_namespace}
\alias{s3tables_create_namespace}
\title{Creates a namespace}
\usage{
s3tables_create_namespace(tableBucketARN, namespace)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket to create the
namespace in.}

\item{namespace}{[required] A name for the namespace.}
}
\description{
Creates a namespace. A namespace is a logical grouping of tables within your table bucket, which you can use to organize tables. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-namespace-create.html}{Create a namespace} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_create_namespace/} for full documentation.
}
\keyword{internal}
