% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_create_table_bucket}
\alias{s3tables_create_table_bucket}
\title{Creates a table bucket}
\usage{
s3tables_create_table_bucket(name)
}
\arguments{
\item{name}{[required] The name for the table bucket.}
}
\description{
Creates a table bucket. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-buckets-create.html}{Creating a table bucket} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_create_table_bucket/} for full documentation.
}
\keyword{internal}
