% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_delete_namespace}
\alias{s3tables_delete_namespace}
\title{Deletes a namespace}
\usage{
s3tables_delete_namespace(tableBucketARN, namespace)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket associated with the
namespace.}

\item{namespace}{[required] The name of the namespace.}
}
\description{
Deletes a namespace. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-namespace-delete.html}{Delete a namespace} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_delete_namespace/} for full documentation.
}
\keyword{internal}
