% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_delete_table}
\alias{s3tables_delete_table}
\title{Deletes a table}
\usage{
s3tables_delete_table(tableBucketARN, namespace, name, versionToken = NULL)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket that contains the
table.}

\item{namespace}{[required] The namespace associated with the table.}

\item{name}{[required] The name of the table.}

\item{versionToken}{The version token of the table.}
}
\description{
Deletes a table. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-delete.html}{Deleting an Amazon S3 table} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_delete_table/} for full documentation.
}
\keyword{internal}
