% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_delete_table_bucket}
\alias{s3tables_delete_table_bucket}
\title{Deletes a table bucket}
\usage{
s3tables_delete_table_bucket(tableBucketARN)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}
}
\description{
Deletes a table bucket. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-buckets-delete.html}{Deleting a table bucket} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_delete_table_bucket/} for full documentation.
}
\keyword{internal}
