% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_delete_table_bucket_policy}
\alias{s3tables_delete_table_bucket_policy}
\title{Deletes a table bucket policy}
\usage{
s3tables_delete_table_bucket_policy(tableBucketARN)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}
}
\description{
Deletes a table bucket policy. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-bucket-policy.html#table-bucket-policy-delete}{Deleting a table bucket policy} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_delete_table_bucket_policy/} for full documentation.
}
\keyword{internal}
