% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_delete_table_bucket_replication}
\alias{s3tables_delete_table_bucket_replication}
\title{Deletes the replication configuration for a table bucket}
\usage{
s3tables_delete_table_bucket_replication(tableBucketARN, versionToken = NULL)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{versionToken}{A version token from a previous GetTableBucketReplication call. Use this
token to ensure you're deleting the expected version of the
configuration.}
}
\description{
Deletes the replication configuration for a table bucket. After deletion, new table updates will no longer be replicated to destination buckets, though existing replicated tables will remain in destination buckets.

See \url{https://www.paws-r-sdk.com/docs/s3tables_delete_table_bucket_replication/} for full documentation.
}
\keyword{internal}
