% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_delete_table_replication}
\alias{s3tables_delete_table_replication}
\title{Deletes the replication configuration for a specific table}
\usage{
s3tables_delete_table_replication(tableArn, versionToken)
}
\arguments{
\item{tableArn}{[required] The Amazon Resource Name (ARN) of the table.}

\item{versionToken}{[required] A version token from a previous GetTableReplication call. Use this token
to ensure you're deleting the expected version of the configuration.}
}
\description{
Deletes the replication configuration for a specific table. After deletion, new updates to this table will no longer be replicated to destination tables, though existing replicated copies will remain in destination buckets.

See \url{https://www.paws-r-sdk.com/docs/s3tables_delete_table_replication/} for full documentation.
}
\keyword{internal}
