% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_namespace}
\alias{s3tables_get_namespace}
\title{Gets details about a namespace}
\usage{
s3tables_get_namespace(tableBucketARN, namespace)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{namespace}{[required] The name of the namespace.}
}
\description{
Gets details about a namespace. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-namespace.html}{Table namespaces} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_namespace/} for full documentation.
}
\keyword{internal}
