% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table}
\alias{s3tables_get_table}
\title{Gets details about a table}
\usage{
s3tables_get_table(tableBucketARN, namespace, name)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket associated with the
table.}

\item{namespace}{[required] The name of the namespace the table is associated with.}

\item{name}{[required] The name of the table.}
}
\description{
Gets details about a table. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-tables.html}{S3 Tables} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table/} for full documentation.
}
\keyword{internal}
