% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_bucket}
\alias{s3tables_get_table_bucket}
\title{Gets details on a table bucket}
\usage{
s3tables_get_table_bucket(tableBucketARN)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}
}
\description{
Gets details on a table bucket. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-buckets-details.html}{Viewing details about an Amazon S3 table bucket} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_bucket/} for full documentation.
}
\keyword{internal}
