% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_bucket_replication}
\alias{s3tables_get_table_bucket_replication}
\title{Retrieves the replication configuration for a table bucket}
\usage{
s3tables_get_table_bucket_replication(tableBucketARN)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}
}
\description{
Retrieves the replication configuration for a table bucket.This operation returns the IAM role, \code{versionToken}, and replication rules that define how tables in this bucket are replicated to other buckets.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_bucket_replication/} for full documentation.
}
\keyword{internal}
