% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_encryption}
\alias{s3tables_get_table_encryption}
\title{Gets the encryption configuration for a table}
\usage{
s3tables_get_table_encryption(tableBucketARN, namespace, name)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket containing the table.}

\item{namespace}{[required] The namespace associated with the table.}

\item{name}{[required] The name of the table.}
}
\description{
Gets the encryption configuration for a table.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_encryption/} for full documentation.
}
\keyword{internal}
