% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_maintenance_configuration}
\alias{s3tables_get_table_maintenance_configuration}
\title{Gets details about the maintenance configuration of a table}
\usage{
s3tables_get_table_maintenance_configuration(tableBucketARN, namespace, name)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{namespace}{[required] The namespace associated with the table.}

\item{name}{[required] The name of the table.}
}
\description{
Gets details about the maintenance configuration of a table. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-maintenance.html}{S3 Tables maintenance} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_maintenance_configuration/} for full documentation.
}
\keyword{internal}
