% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_metadata_location}
\alias{s3tables_get_table_metadata_location}
\title{Gets the location of the table metadata}
\usage{
s3tables_get_table_metadata_location(tableBucketARN, namespace, name)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{namespace}{[required] The namespace of the table.}

\item{name}{[required] The name of the table.}
}
\description{
Gets the location of the table metadata.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_metadata_location/} for full documentation.
}
\keyword{internal}
