% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_policy}
\alias{s3tables_get_table_policy}
\title{Gets details about a table policy}
\usage{
s3tables_get_table_policy(tableBucketARN, namespace, name)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket that contains the
table.}

\item{namespace}{[required] The namespace associated with the table.}

\item{name}{[required] The name of the table.}
}
\description{
Gets details about a table policy. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-table-policy.html#table-policy-get}{Viewing a table policy} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_policy/} for full documentation.
}
\keyword{internal}
