% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_record_expiration_configuration}
\alias{s3tables_get_table_record_expiration_configuration}
\title{Retrieves the expiration configuration settings for records in a table,
and the status of the configuration}
\usage{
s3tables_get_table_record_expiration_configuration(tableArn)
}
\arguments{
\item{tableArn}{[required] The Amazon Resource Name (ARN) of the table.}
}
\description{
Retrieves the expiration configuration settings for records in a table, and the status of the configuration. If the status of the configuration is \code{enabled}, records expire and are automatically removed from the table after the specified number of days.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_record_expiration_configuration/} for full documentation.
}
\keyword{internal}
