% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_record_expiration_job_status}
\alias{s3tables_get_table_record_expiration_job_status}
\title{Retrieves the status, metrics, and details of the latest record
expiration job for a table}
\usage{
s3tables_get_table_record_expiration_job_status(tableArn)
}
\arguments{
\item{tableArn}{[required] The Amazon Resource Name (ARN) of the table.}
}
\description{
Retrieves the status, metrics, and details of the latest record expiration job for a table. This includes when the job ran, and whether it succeeded or failed. If the job ran successfully, this also includes statistics about the records that were removed.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_record_expiration_job_status/} for full documentation.
}
\keyword{internal}
