% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_replication}
\alias{s3tables_get_table_replication}
\title{Retrieves the replication configuration for a specific table}
\usage{
s3tables_get_table_replication(tableArn)
}
\arguments{
\item{tableArn}{[required] The Amazon Resource Name (ARN) of the table.}
}
\description{
Retrieves the replication configuration for a specific table.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_replication/} for full documentation.
}
\keyword{internal}
