% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_replication_status}
\alias{s3tables_get_table_replication_status}
\title{Retrieves the replication status for a table, including the status of
replication to each destination}
\usage{
s3tables_get_table_replication_status(tableArn)
}
\arguments{
\item{tableArn}{[required] The Amazon Resource Name (ARN) of the table.}
}
\description{
Retrieves the replication status for a table, including the status of replication to each destination. This operation provides visibility into replication health and progress.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_replication_status/} for full documentation.
}
\keyword{internal}
