% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_storage_class}
\alias{s3tables_get_table_storage_class}
\title{Retrieves the storage class configuration for a specific table}
\usage{
s3tables_get_table_storage_class(tableBucketARN, namespace, name)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket that contains the
table.}

\item{namespace}{[required] The namespace associated with the table.}

\item{name}{[required] The name of the table.}
}
\description{
Retrieves the storage class configuration for a specific table. This allows you to view the storage class settings that apply to an individual table, which may differ from the table bucket's default configuration.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_storage_class/} for full documentation.
}
\keyword{internal}
