% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_list_table_buckets}
\alias{s3tables_list_table_buckets}
\title{Lists table buckets for your account}
\usage{
s3tables_list_table_buckets(
  prefix = NULL,
  continuationToken = NULL,
  maxBuckets = NULL,
  type = NULL
)
}
\arguments{
\item{prefix}{The prefix of the table buckets.}

\item{continuationToken}{\code{ContinuationToken} indicates to Amazon S3 that the list is being
continued on this bucket with a token. \code{ContinuationToken} is obfuscated
and is not a real key. You can use this \code{ContinuationToken} for
pagination of the list results.}

\item{maxBuckets}{The maximum number of table buckets to return in the list.}

\item{type}{The type of table buckets to filter by in the list.}
}
\description{
Lists table buckets for your account. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-buckets.html}{S3 Table buckets} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_list_table_buckets/} for full documentation.
}
\keyword{internal}
