% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_put_table_bucket_maintenance_configuration}
\alias{s3tables_put_table_bucket_maintenance_configuration}
\title{Creates a new maintenance configuration or replaces an existing
maintenance configuration for a table bucket}
\usage{
s3tables_put_table_bucket_maintenance_configuration(
  tableBucketARN,
  type,
  value
)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket associated with the
maintenance configuration.}

\item{type}{[required] The type of the maintenance configuration.}

\item{value}{[required] Defines the values of the maintenance configuration for the table
bucket.}
}
\description{
Creates a new maintenance configuration or replaces an existing maintenance configuration for a table bucket. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-table-buckets-maintenance.html}{Amazon S3 table bucket maintenance} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_put_table_bucket_maintenance_configuration/} for full documentation.
}
\keyword{internal}
