% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_put_table_bucket_metrics_configuration}
\alias{s3tables_put_table_bucket_metrics_configuration}
\title{Sets the metrics configuration for a table bucket}
\usage{
s3tables_put_table_bucket_metrics_configuration(tableBucketARN)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}
}
\description{
Sets the metrics configuration for a table bucket.

See \url{https://www.paws-r-sdk.com/docs/s3tables_put_table_bucket_metrics_configuration/} for full documentation.
}
\keyword{internal}
