% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_put_table_bucket_storage_class}
\alias{s3tables_put_table_bucket_storage_class}
\title{Sets or updates the storage class configuration for a table bucket}
\usage{
s3tables_put_table_bucket_storage_class(
  tableBucketARN,
  storageClassConfiguration
)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{storageClassConfiguration}{[required] The storage class configuration to apply to the table bucket. This
configuration will serve as the default for new tables created in this
bucket.}
}
\description{
Sets or updates the storage class configuration for a table bucket. This configuration serves as the default storage class for all new tables created in the bucket, allowing you to optimize storage costs at the bucket level.

See \url{https://www.paws-r-sdk.com/docs/s3tables_put_table_bucket_storage_class/} for full documentation.
}
\keyword{internal}
