% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_put_table_maintenance_configuration}
\alias{s3tables_put_table_maintenance_configuration}
\title{Creates a new maintenance configuration or replaces an existing
maintenance configuration for a table}
\usage{
s3tables_put_table_maintenance_configuration(
  tableBucketARN,
  namespace,
  name,
  type,
  value
)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table associated with the
maintenance configuration.}

\item{namespace}{[required] The namespace of the table.}

\item{name}{[required] The name of the table.}

\item{type}{[required] The type of the maintenance configuration.}

\item{value}{[required] Defines the values of the maintenance configuration for the table.}
}
\description{
Creates a new maintenance configuration or replaces an existing maintenance configuration for a table. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-maintenance.html}{S3 Tables maintenance} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_put_table_maintenance_configuration/} for full documentation.
}
\keyword{internal}
