% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_put_table_record_expiration_configuration}
\alias{s3tables_put_table_record_expiration_configuration}
\title{Creates or updates the expiration configuration settings for records in
a table, including the status of the configuration}
\usage{
s3tables_put_table_record_expiration_configuration(tableArn, value)
}
\arguments{
\item{tableArn}{[required] The Amazon Resource Name (ARN) of the table.}

\item{value}{[required] The record expiration configuration to apply to the table, including the
status (\code{enabled} or \code{disabled}) and retention period in days.}
}
\description{
Creates or updates the expiration configuration settings for records in a table, including the status of the configuration. If you enable record expiration for a table, records expire and are automatically removed from the table after the number of days that you specify.

See \url{https://www.paws-r-sdk.com/docs/s3tables_put_table_record_expiration_configuration/} for full documentation.
}
\keyword{internal}
