% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_rename_table}
\alias{s3tables_rename_table}
\title{Renames a table or a namespace}
\usage{
s3tables_rename_table(
  tableBucketARN,
  namespace,
  name,
  newNamespaceName = NULL,
  newName = NULL,
  versionToken = NULL
)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{namespace}{[required] The namespace associated with the table.}

\item{name}{[required] The current name of the table.}

\item{newNamespaceName}{The new name for the namespace.}

\item{newName}{The new name for the table.}

\item{versionToken}{The version token of the table.}
}
\description{
Renames a table or a namespace. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-tables.html}{S3 Tables} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_rename_table/} for full documentation.
}
\keyword{internal}
