% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_untag_resource}
\alias{s3tables_untag_resource}
\title{Removes the specified user-defined tags from an Amazon S3 Tables
resource}
\usage{
s3tables_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon S3 Tables resource that
you're removing tags from. The tagged resource can be a table bucket or
a table. For a list of all S3 resources that support tagging, see
\href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/tagging.html#manage-tags}{Managing tags for Amazon S3 resources}.}

\item{tagKeys}{[required] The array of tag keys that you're removing from the S3 Tables resource.
For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/tagging.html}{Tagging for cost allocation or attribute-based access control (ABAC)}.}
}
\description{
Removes the specified user-defined tags from an Amazon S3 Tables resource. You can pass one or more tag keys.

See \url{https://www.paws-r-sdk.com/docs/s3tables_untag_resource/} for full documentation.
}
\keyword{internal}
