% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_update_table_metadata_location}
\alias{s3tables_update_table_metadata_location}
\title{Updates the metadata location for a table}
\usage{
s3tables_update_table_metadata_location(
  tableBucketARN,
  namespace,
  name,
  versionToken,
  metadataLocation
)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{namespace}{[required] The namespace of the table.}

\item{name}{[required] The name of the table.}

\item{versionToken}{[required] The version token of the table.}

\item{metadataLocation}{[required] The new metadata location for the table.}
}
\description{
Updates the metadata location for a table. The metadata location of a table must be an S3 URI that begins with the table's warehouse location. The metadata location for an Apache Iceberg table must end with \code{.metadata.json}, or if the metadata file is Gzip-compressed, \code{.metadata.json.gz}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_update_table_metadata_location/} for full documentation.
}
\keyword{internal}
