% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_add_cache}
\alias{storagegateway_add_cache}
\title{Configures one or more gateway local disks as cache for a gateway}
\usage{
storagegateway_add_cache(GatewayARN, DiskIds)
}
\arguments{
\item{GatewayARN}{[required]}

\item{DiskIds}{[required] An array of strings that identify disks that are to be configured as
working storage. Each string has a minimum length of 1 and maximum
length of 300. You can get the disk IDs from the
\code{\link[=storagegateway_list_local_disks]{list_local_disks}} API.}
}
\description{
Configures one or more gateway local disks as cache for a gateway. This operation is only supported in the cached volume, tape, and file gateway type (see \href{https://docs.aws.amazon.com/storagegateway/index.html}{How Storage Gateway works (architecture)}.

See \url{https://paws-r.github.io/docs/storagegateway/add_cache.html} for full documentation.
}
\keyword{internal}
