% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_add_upload_buffer}
\alias{storagegateway_add_upload_buffer}
\title{Configures one or more gateway local disks as upload buffer for a
specified gateway}
\usage{
storagegateway_add_upload_buffer(GatewayARN, DiskIds)
}
\arguments{
\item{GatewayARN}{[required]}

\item{DiskIds}{[required] An array of strings that identify disks that are to be configured as
working storage. Each string has a minimum length of 1 and maximum
length of 300. You can get the disk IDs from the
\code{\link[=storagegateway_list_local_disks]{list_local_disks}} API.}
}
\description{
Configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for the stored volume, cached volume, and tape gateway types.

See \url{https://paws-r.github.io/docs/storagegateway/add_upload_buffer.html} for full documentation.
}
\keyword{internal}
