% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_add_working_storage}
\alias{storagegateway_add_working_storage}
\title{Configures one or more gateway local disks as working storage for a
gateway}
\usage{
storagegateway_add_working_storage(GatewayARN, DiskIds)
}
\arguments{
\item{GatewayARN}{[required]}

\item{DiskIds}{[required] An array of strings that identify disks that are to be configured as
working storage. Each string has a minimum length of 1 and maximum
length of 300. You can get the disk IDs from the
\code{\link[=storagegateway_list_local_disks]{list_local_disks}} API.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GatewayARN = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Configures one or more gateway local disks as working storage for a
gateway. This operation is only supported in the stored volume gateway
type. This operation is deprecated in cached volume API version
20120630. Use \code{\link[=storagegateway_add_upload_buffer]{add_upload_buffer}}
instead.

Working storage is also referred to as upload buffer. You can also use
the \code{\link[=storagegateway_add_upload_buffer]{add_upload_buffer}} operation to
add upload buffer to a stored volume gateway.

In the request, you specify the gateway Amazon Resource Name (ARN) to
which you want to add working storage, and one or more disk IDs that you
want to configure as working storage.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$add_working_storage(
  GatewayARN = "string",
  DiskIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Configures one or more gateway local disks as working storage for a
# gateway. (Working storage is also referred to as upload buffer.)
svc$add_working_storage(
  DiskIds = list(
    "pci-0000:03:00.0-scsi-0:0:0:0",
    "pci-0000:03:00.0-scsi-0:0:1:0"
  ),
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)
}

}
\keyword{internal}
