% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_delete_automatic_tape_creation_policy}
\alias{storagegateway_delete_automatic_tape_creation_policy}
\title{Deletes the automatic tape creation policy of a gateway}
\usage{
storagegateway_delete_automatic_tape_creation_policy(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Deletes the automatic tape creation policy of a gateway. If you delete this policy, new virtual tapes must be created manually. Use the Amazon Resource Name (ARN) of the gateway in your request to remove the policy.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_delete_automatic_tape_creation_policy/} for full documentation.
}
\keyword{internal}
