% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_delete_snapshot_schedule}
\alias{storagegateway_delete_snapshot_schedule}
\title{Deletes a snapshot of a volume}
\usage{
storagegateway_delete_snapshot_schedule(VolumeARN)
}
\arguments{
\item{VolumeARN}{[required] The volume which snapshot schedule to delete.}
}
\description{
Deletes a snapshot of a volume.

See \url{https://paws-r.github.io/docs/storagegateway/delete_snapshot_schedule.html} for full documentation.
}
\keyword{internal}
