% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_availability_monitor_test}
\alias{storagegateway_describe_availability_monitor_test}
\title{Returns information about the most recent high availability monitoring
test that was performed on the host in a cluster}
\usage{
storagegateway_describe_availability_monitor_test(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Returns information about the most recent high availability monitoring test that was performed on the host in a cluster. If a test isn't performed, the status and start time in the response would be null.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_availability_monitor_test/} for full documentation.
}
\keyword{internal}
