% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_bandwidth_rate_limit}
\alias{storagegateway_describe_bandwidth_rate_limit}
\title{Returns the bandwidth rate limits of a gateway}
\usage{
storagegateway_describe_bandwidth_rate_limit(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect. This operation is supported only for the stored volume, cached volume, and tape gateway types. To describe bandwidth rate limits for S3 file gateways, use \code{\link[=storagegateway_describe_bandwidth_rate_limit_schedule]{describe_bandwidth_rate_limit_schedule}}.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_bandwidth_rate_limit/} for full documentation.
}
\keyword{internal}
