% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_cache_report}
\alias{storagegateway_describe_cache_report}
\title{Returns information about the specified cache report, including
completion status and generation progress}
\usage{
storagegateway_describe_cache_report(CacheReportARN)
}
\arguments{
\item{CacheReportARN}{[required] The Amazon Resource Name (ARN) of the cache report you want to describe.}
}
\description{
Returns information about the specified cache report, including completion status and generation progress.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_cache_report/} for full documentation.
}
\keyword{internal}
