% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_cachedi_scsi_volumes}
\alias{storagegateway_describe_cachedi_scsi_volumes}
\title{Returns a description of the gateway volumes specified in the request}
\usage{
storagegateway_describe_cachedi_scsi_volumes(VolumeARNs)
}
\arguments{
\item{VolumeARNs}{[required] An array of strings where each string represents the Amazon Resource
Name (ARN) of a cached volume. All of the specified cached volumes must
be from the same gateway. Use
\code{\link[=storagegateway_list_volumes]{list_volumes}} to get volume ARNs for a
gateway.}
}
\description{
Returns a description of the gateway volumes specified in the request. This operation is only supported in the cached volume gateway types.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_cachedi_scsi_volumes/} for full documentation.
}
\keyword{internal}
