% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_file_system_associations}
\alias{storagegateway_describe_file_system_associations}
\title{Gets the file system association information}
\usage{
storagegateway_describe_file_system_associations(
  FileSystemAssociationARNList)
}
\arguments{
\item{FileSystemAssociationARNList}{[required] An array containing the Amazon Resource Name (ARN) of each file system
association to be described.}
}
\value{
A list with the following syntax:\preformatted{list(
  FileSystemAssociationInfoList = list(
    list(
      FileSystemAssociationARN = "string",
      LocationARN = "string",
      FileSystemAssociationStatus = "string",
      AuditDestinationARN = "string",
      GatewayARN = "string",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      ),
      CacheAttributes = list(
        CacheStaleTimeoutInSeconds = 123
      )
    )
  )
)
}
}
\description{
Gets the file system association information. This operation is only
supported for Amazon FSx file gateways.
}
\section{Request syntax}{
\preformatted{svc$describe_file_system_associations(
  FileSystemAssociationARNList = list(
    "string"
  )
)
}
}

\keyword{internal}
