% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_gateway_information}
\alias{storagegateway_describe_gateway_information}
\title{Returns metadata about a gateway such as its name, network interfaces,
time zone, status, and software version}
\usage{
storagegateway_describe_gateway_information(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Returns metadata about a gateway such as its name, network interfaces, time zone, status, and software version. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_gateway_information/} for full documentation.
}
\keyword{internal}
