% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_maintenance_start_time}
\alias{storagegateway_describe_maintenance_start_time}
\title{Returns your gateway's maintenance window schedule information, with
values for monthly or weekly cadence, specific day and time to begin
maintenance, and which types of updates to apply}
\usage{
storagegateway_describe_maintenance_start_time(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Returns your gateway's maintenance window schedule information, with values for monthly or weekly cadence, specific day and time to begin maintenance, and which types of updates to apply. Time values returned are for the gateway's time zone.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_maintenance_start_time/} for full documentation.
}
\keyword{internal}
