% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_nfs_file_shares}
\alias{storagegateway_describe_nfs_file_shares}
\title{Gets a description for one or more Network File System (NFS) file shares
from an S3 File Gateway}
\usage{
storagegateway_describe_nfs_file_shares(FileShareARNList)
}
\arguments{
\item{FileShareARNList}{[required] An array containing the Amazon Resource Name (ARN) of each file share to
be described.}
}
\description{
Gets a description for one or more Network File System (NFS) file shares from an S3 File Gateway. This operation is only supported for S3 File Gateways.

See \url{https://paws-r.github.io/docs/storagegateway/describe_nfs_file_shares.html} for full documentation.
}
\keyword{internal}
