% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_upload_buffer}
\alias{storagegateway_describe_upload_buffer}
\title{Returns information about the upload buffer of a gateway}
\usage{
storagegateway_describe_upload_buffer(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns information about the upload buffer of a gateway. This operation is supported for the stored volume, cached volume, and tape gateway types.

See \url{https://paws-r.github.io/docs/storagegateway/describe_upload_buffer.html} for full documentation.
}
\keyword{internal}
