% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_working_storage}
\alias{storagegateway_describe_working_storage}
\title{Returns information about the working storage of a gateway}
\usage{
storagegateway_describe_working_storage(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns information about the working storage of a gateway. This operation is only supported in the stored volumes gateway type. This operation is deprecated in cached volumes API version (20120630). Use DescribeUploadBuffer instead.

See \url{https://paws-r.github.io/docs/storagegateway/describe_working_storage.html} for full documentation.
}
\keyword{internal}
