% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_file_system_associations}
\alias{storagegateway_list_file_system_associations}
\title{Gets a list of FileSystemAssociationSummary objects}
\usage{
storagegateway_list_file_system_associations(GatewayARN, Limit, Marker)
}
\arguments{
\item{GatewayARN}{}

\item{Limit}{The maximum number of file system associations to return in the
response. If present, \code{Limit} must be an integer with a value greater
than zero. Optional.}

\item{Marker}{Opaque pagination token returned from a previous
\code{\link[=storagegateway_list_file_system_associations]{list_file_system_associations}}
operation. If present, \code{Marker} specifies where to continue the list
from after a previous call to
\code{\link[=storagegateway_list_file_system_associations]{list_file_system_associations}}.
Optional.}
}
\value{
A list with the following syntax:\preformatted{list(
  Marker = "string",
  NextMarker = "string",
  FileSystemAssociationSummaryList = list(
    list(
      FileSystemAssociationId = "string",
      FileSystemAssociationARN = "string",
      FileSystemAssociationStatus = "string",
      GatewayARN = "string"
    )
  )
)
}
}
\description{
Gets a list of \code{FileSystemAssociationSummary} objects. Each object
contains a summary of a file system association. This operation is only
supported for Amazon FSx file gateways.
}
\section{Request syntax}{
\preformatted{svc$list_file_system_associations(
  GatewayARN = "string",
  Limit = 123,
  Marker = "string"
)
}
}

\keyword{internal}
