% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_gateways}
\alias{storagegateway_list_gateways}
\title{Lists gateways owned by an Amazon Web Services account in an Amazon Web
Services Region specified in the request}
\usage{
storagegateway_list_gateways(Marker = NULL, Limit = NULL)
}
\arguments{
\item{Marker}{An opaque string that indicates the position at which to begin the
returned list of gateways.}

\item{Limit}{Specifies that the list of gateways returned be limited to the specified
number of items.}
}
\description{
Lists gateways owned by an Amazon Web Services account in an Amazon Web Services Region specified in the request. The returned list is ordered by gateway Amazon Resource Name (ARN).

See \url{https://www.paws-r-sdk.com/docs/storagegateway_list_gateways/} for full documentation.
}
\keyword{internal}
