% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_local_disks}
\alias{storagegateway_list_local_disks}
\title{Returns a list of the gateway's local disks}
\usage{
storagegateway_list_local_disks(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Returns a list of the gateway's local disks. To specify which gateway to describe, you use the Amazon Resource Name (ARN) of the gateway in the body of the request.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_list_local_disks/} for full documentation.
}
\keyword{internal}
