% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_tags_for_resource}
\alias{storagegateway_list_tags_for_resource}
\title{Lists the tags that have been added to the specified resource}
\usage{
storagegateway_list_tags_for_resource(ResourceARN, Marker = NULL, Limit = NULL)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource for which you want to
list tags.}

\item{Marker}{An opaque string that indicates the position at which to begin returning
the list of tags.}

\item{Limit}{Specifies that the list of tags returned be limited to the specified
number of items.}
}
\description{
Lists the tags that have been added to the specified resource. This operation is supported in storage gateways of all types.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
