% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_volume_recovery_points}
\alias{storagegateway_list_volume_recovery_points}
\title{Lists the recovery points for a specified gateway}
\usage{
storagegateway_list_volume_recovery_points(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Lists the recovery points for a specified gateway. This operation is only supported in the cached volume gateway type.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_list_volume_recovery_points/} for full documentation.
}
\keyword{internal}
