% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_notify_when_uploaded}
\alias{storagegateway_notify_when_uploaded}
\title{Sends you notification through CloudWatch Events when all files written
to your file share have been uploaded to Amazon S3}
\usage{
storagegateway_notify_when_uploaded(FileShareARN)
}
\arguments{
\item{FileShareARN}{[required]}
}
\description{
Sends you notification through CloudWatch Events when all files written to your file share have been uploaded to Amazon S3.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_notify_when_uploaded/} for full documentation.
}
\keyword{internal}
