% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_remove_tags_from_resource}
\alias{storagegateway_remove_tags_from_resource}
\title{Removes one or more tags from the specified resource}
\usage{
storagegateway_remove_tags_from_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource you want to remove the
tags from.}

\item{TagKeys}{[required] The keys of the tags you want to remove from the specified resource. A
tag is composed of a key-value pair.}
}
\description{
Removes one or more tags from the specified resource. This operation is supported in storage gateways of all types.

See \url{https://paws-r.github.io/docs/storagegateway/remove_tags_from_resource.html} for full documentation.
}
\keyword{internal}
