% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_reset_cache}
\alias{storagegateway_reset_cache}
\title{Resets all cache disks that have encountered an error and makes the
disks available for reconfiguration as cache storage}
\usage{
storagegateway_reset_cache(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Resets all cache disks that have encountered an error and makes the disks available for reconfiguration as cache storage. If your cache disk encounters an error, the gateway prevents read and write operations on virtual tapes in the gateway. For example, an error can occur when a disk is corrupted or removed from the gateway. When a cache is reset, the gateway loses its cache storage. At this point, you can reconfigure the disks as cache disks. This operation is only supported in the cached volume and tape types.

See \url{https://paws-r.github.io/docs/storagegateway/reset_cache.html} for full documentation.
}
\keyword{internal}
