% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_retrieve_tape_recovery_point}
\alias{storagegateway_retrieve_tape_recovery_point}
\title{Retrieves the recovery point for the specified virtual tape}
\usage{
storagegateway_retrieve_tape_recovery_point(TapeARN, GatewayARN)
}
\arguments{
\item{TapeARN}{[required] The Amazon Resource Name (ARN) of the virtual tape for which you want to
retrieve the recovery point.}

\item{GatewayARN}{[required]}
}
\description{
Retrieves the recovery point for the specified virtual tape. This operation is only supported in the tape gateway type.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_retrieve_tape_recovery_point/} for full documentation.
}
\keyword{internal}
