% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_set_local_console_password}
\alias{storagegateway_set_local_console_password}
\title{Sets the password for your VM local console}
\usage{
storagegateway_set_local_console_password(GatewayARN, LocalConsolePassword)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}

\item{LocalConsolePassword}{[required] The password you want to set for your VM local console.}
}
\description{
Sets the password for your VM local console. When you log in to the local console for the first time, you log in to the VM with the default credentials. We recommend that you set a new password. You don't need to know the default password to set a new password.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_set_local_console_password/} for full documentation.
}
\keyword{internal}
