% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_shutdown_gateway}
\alias{storagegateway_shutdown_gateway}
\title{Shuts down a Tape Gateway or Volume Gateway}
\usage{
storagegateway_shutdown_gateway(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required] The Amazon Resource Name (ARN) of the gateway. Use the \code{\link[=storagegateway_list_gateways]{list_gateways}} operation to return a list of gateways for your account and Amazon Web Services Region.}
}
\description{
Shuts down a Tape Gateway or Volume Gateway. To specify which gateway to shut down, use the Amazon Resource Name (ARN) of the gateway in the body of your request.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_shutdown_gateway/} for full documentation.
}
\keyword{internal}
