% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_start_availability_monitor_test}
\alias{storagegateway_start_availability_monitor_test}
\title{Start a test that verifies that the specified gateway is configured for
High Availability monitoring in your host environment}
\usage{
storagegateway_start_availability_monitor_test(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Start a test that verifies that the specified gateway is configured for High Availability monitoring in your host environment. This request only initiates the test and that a successful response only indicates that the test was started. It doesn't indicate that the test passed. For the status of the test, invoke the \code{\link[=storagegateway_describe_availability_monitor_test]{describe_availability_monitor_test}} API.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_start_availability_monitor_test/} for full documentation.
}
\keyword{internal}
