% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_automatic_tape_creation_policy}
\alias{storagegateway_update_automatic_tape_creation_policy}
\title{Updates the automatic tape creation policy of a gateway}
\usage{
storagegateway_update_automatic_tape_creation_policy(
  AutomaticTapeCreationRules, GatewayARN)
}
\arguments{
\item{AutomaticTapeCreationRules}{[required] An automatic tape creation policy consists of a list of automatic tape
creation rules. The rules determine when and how to automatically create
new tapes.}

\item{GatewayARN}{[required]}
}
\description{
Updates the automatic tape creation policy of a gateway. Use this to
update the policy with a new set of automatic tape creation rules. This
is only supported for tape gateways.
}
\details{
By default, there is no automatic tape creation policy.

A gateway can have only one automatic tape creation policy.
}
\section{Request syntax}{
\preformatted{svc$update_automatic_tape_creation_policy(
  AutomaticTapeCreationRules = list(
    list(
      TapeBarcodePrefix = "string",
      PoolId = "string",
      TapeSizeInBytes = 123,
      MinimumNumTapes = 123,
      Worm = TRUE|FALSE
    )
  ),
  GatewayARN = "string"
)
}
}

\keyword{internal}
