% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_gateway_information}
\alias{storagegateway_update_gateway_information}
\title{Updates a gateway's metadata, which includes the gateway's name and time
zone}
\usage{
storagegateway_update_gateway_information(
  GatewayARN,
  GatewayName = NULL,
  GatewayTimezone = NULL,
  CloudWatchLogGroupARN = NULL,
  GatewayCapacity = NULL
)
}
\arguments{
\item{GatewayARN}{[required]}

\item{GatewayName}{}

\item{GatewayTimezone}{A value that indicates the time zone of the gateway.}

\item{CloudWatchLogGroupARN}{The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that
you want to use to monitor and log events in the gateway.

For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/WhatIsCloudWatchLogs.html}{What is Amazon CloudWatch Logs?}}

\item{GatewayCapacity}{Specifies the size of the gateway's metadata cache.}
}
\description{
Updates a gateway's metadata, which includes the gateway's name and time zone. To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request.

See \url{https://paws-r.github.io/docs/storagegateway/update_gateway_information.html} for full documentation.
}
\keyword{internal}
