% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_smb_file_share_visibility}
\alias{storagegateway_update_smb_file_share_visibility}
\title{Controls whether the shares on a gateway are visible in a net view or
browse list}
\usage{
storagegateway_update_smb_file_share_visibility(GatewayARN,
  FileSharesVisible)
}
\arguments{
\item{GatewayARN}{[required]}

\item{FileSharesVisible}{[required] The shares on this gateway appear when listing shares.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GatewayARN = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Controls whether the shares on a gateway are visible in a net view or
browse list.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_smb_file_share_visibility(
  GatewayARN = "string",
  FileSharesVisible = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
