% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_create_backup_plan}
\alias{backup_create_backup_plan}
\title{Creates a backup plan using a backup plan name and backup rules}
\usage{
backup_create_backup_plan(
  BackupPlan,
  BackupPlanTags = NULL,
  CreatorRequestId = NULL
)
}
\arguments{
\item{BackupPlan}{[required] Specifies the body of a backup plan. Includes a \code{BackupPlanName} and one
or more sets of \code{Rules}.}

\item{BackupPlanTags}{To help organize your resources, you can assign your own metadata to the
resources that you create. Each tag is a key-value pair. The specified
tags are assigned to all backups created with this plan.}

\item{CreatorRequestId}{Identifies the request and allows failed requests to be retried without
the risk of running the operation twice. If the request includes a
\code{CreatorRequestId} that matches an existing backup plan, that plan is
returned. This parameter is optional.

If used, this parameter must contain 1 to 50 alphanumeric or '-_.'
characters.}
}
\description{
Creates a backup plan using a backup plan name and backup rules. A backup plan is a document that contains information that Backup uses to schedule tasks that create recovery points for resources.

See \url{https://www.paws-r-sdk.com/docs/backup_create_backup_plan/} for full documentation.
}
\keyword{internal}
