% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_create_restore_testing_plan}
\alias{backup_create_restore_testing_plan}
\title{Creates a restore testing plan}
\usage{
backup_create_restore_testing_plan(
  CreatorRequestId = NULL,
  RestoreTestingPlan,
  Tags = NULL
)
}
\arguments{
\item{CreatorRequestId}{This is a unique string that identifies the request and allows failed
requests to be retriedwithout the risk of running the operation twice.
This parameter is optional. If used, this parameter must contain 1 to 50
alphanumeric or '-_.' characters.}

\item{RestoreTestingPlan}{[required] A restore testing plan must contain a unique \code{RestoreTestingPlanName}
string you create and must contain a \code{ScheduleExpression} cron. You may
optionally include a \code{StartWindowHours} integer and a \code{CreatorRequestId}
string.

The \code{RestoreTestingPlanName} is a unique string that is the name of the
restore testing plan. This cannot be changed after creation, and it must
consist of only alphanumeric characters and underscores.}

\item{Tags}{The tags to assign to the restore testing plan.}
}
\description{
Creates a restore testing plan.

See \url{https://www.paws-r-sdk.com/docs/backup_create_restore_testing_plan/} for full documentation.
}
\keyword{internal}
