% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_delete_backup_plan}
\alias{backup_delete_backup_plan}
\title{Deletes a backup plan}
\usage{
backup_delete_backup_plan(BackupPlanId)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}
}
\description{
Deletes a backup plan. A backup plan can only be deleted after all associated selections of resources have been deleted. Deleting a backup plan deletes the current version of a backup plan. Previous versions, if any, will still exist.

See \url{https://paws-r.github.io/docs/backup/delete_backup_plan.html} for full documentation.
}
\keyword{internal}
