% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_backup_job}
\alias{backup_describe_backup_job}
\title{Returns backup job details for the specified BackupJobId}
\usage{
backup_describe_backup_job(BackupJobId)
}
\arguments{
\item{BackupJobId}{[required] Uniquely identifies a request to Backup to back up a resource.}
}
\description{
Returns backup job details for the specified \code{BackupJobId}.

See \url{https://paws-r.github.io/docs/backup/describe_backup_job.html} for full documentation.
}
\keyword{internal}
