% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_backup_vault}
\alias{backup_describe_backup_vault}
\title{Returns metadata about a backup vault specified by its name}
\usage{
backup_describe_backup_vault(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults
are identified by names that are unique to the account used to create
them and the AWS Region where they are created. They consist of
lowercase letters, numbers, and hyphens.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BackupVaultName = "string",
  BackupVaultArn = "string",
  EncryptionKeyArn = "string",
  CreationDate = as.POSIXct(
    "2015-01-01"
  ),
  CreatorRequestId = "string",
  NumberOfRecoveryPoints = 123
)
}\if{html}{\out{</div>}}
}
\description{
Returns metadata about a backup vault specified by its name.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_backup_vault(
  BackupVaultName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
