% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_copy_job}
\alias{backup_describe_copy_job}
\title{Returns metadata associated with creating a copy of a resource}
\usage{
backup_describe_copy_job(CopyJobId)
}
\arguments{
\item{CopyJobId}{[required] Uniquely identifies a copy job.}
}
\description{
Returns metadata associated with creating a copy of a resource.
}
\section{Request syntax}{
\preformatted{svc$describe_copy_job(
  CopyJobId = "string"
)
}
}

\keyword{internal}
