% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_framework}
\alias{backup_describe_framework}
\title{Returns the framework details for the specified FrameworkName}
\usage{
backup_describe_framework(FrameworkName)
}
\arguments{
\item{FrameworkName}{[required] The unique name of a framework.}
}
\description{
Returns the framework details for the specified \code{FrameworkName}.

See \url{https://www.paws-r-sdk.com/docs/backup_describe_framework/} for full documentation.
}
\keyword{internal}
