% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_global_settings}
\alias{backup_describe_global_settings}
\title{Describes the global settings of the AWS account, including whether it
is opted in to cross-account backup}
\usage{
backup_describe_global_settings()
}
\value{
A list with the following syntax:\preformatted{list(
  GlobalSettings = list(
    "string"
  ),
  LastUpdateTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Describes the global settings of the AWS account, including whether it
is opted in to cross-account backup.
}
\section{Request syntax}{
\preformatted{svc$describe_global_settings()
}
}

\keyword{internal}
