% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_global_settings}
\alias{backup_describe_global_settings}
\title{Describes whether the Amazon Web Services account is opted in to
cross-account backup}
\usage{
backup_describe_global_settings()
}
\description{
Describes whether the Amazon Web Services account is opted in to cross-account backup. Returns an error if the account is not a member of an Organizations organization. Example: \verb{describe-global-settings --region us-west-2}

See \url{https://paws-r.github.io/docs/backup/describe_global_settings.html} for full documentation.
}
\keyword{internal}
