% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_protected_resource}
\alias{backup_describe_protected_resource}
\title{Returns information about a saved resource, including the last time it
was backed-up, its Amazon Resource Name (ARN), and the AWS service type
of the saved resource}
\usage{
backup_describe_protected_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies a resource. The
format of the ARN depends on the resource type.}
}
\description{
Returns information about a saved resource, including the last time it
was backed-up, its Amazon Resource Name (ARN), and the AWS service type
of the saved resource.
}
\section{Request syntax}{
\preformatted{svc$describe_protected_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
