% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_region_settings}
\alias{backup_describe_region_settings}
\title{Returns the current service opt-in settings for the Region}
\usage{
backup_describe_region_settings()
}
\description{
Returns the current service opt-in settings for the Region. If service opt-in is enabled for a service, Backup tries to protect that service's resources in this Region, when the resource is included in an on-demand backup or scheduled backup plan. Otherwise, Backup does not try to protect that service's resources in this Region.

See \url{https://www.paws-r-sdk.com/docs/backup_describe_region_settings/} for full documentation.
}
\keyword{internal}
