% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_restore_job}
\alias{backup_describe_restore_job}
\title{Returns metadata associated with a restore job that is specified by a
job ID}
\usage{
backup_describe_restore_job(RestoreJobId)
}
\arguments{
\item{RestoreJobId}{[required] Uniquely identifies the job that restores a recovery point.}
}
\description{
Returns metadata associated with a restore job that is specified by a job ID.

See \url{https://www.paws-r-sdk.com/docs/backup_describe_restore_job/} for full documentation.
}
\keyword{internal}
