% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_disassociate_recovery_point}
\alias{backup_disassociate_recovery_point}
\title{Deletes the specified continuous backup recovery point from Backup and
releases control of that continuous backup to the source service, such
as Amazon RDS}
\usage{
backup_disassociate_recovery_point(BackupVaultName, RecoveryPointArn)
}
\arguments{
\item{BackupVaultName}{[required] The unique name of an Backup vault.}

\item{RecoveryPointArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies an Backup
recovery point.}
}
\description{
Deletes the specified continuous backup recovery point from Backup and releases control of that continuous backup to the source service, such as Amazon RDS. The source service will continue to create and retain continuous backups using the lifecycle that you specified in your original backup plan.

See \url{https://www.paws-r-sdk.com/docs/backup_disassociate_recovery_point/} for full documentation.
}
\keyword{internal}
