% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_backup_plan}
\alias{backup_get_backup_plan}
\title{Returns BackupPlan details for the specified BackupPlanId}
\usage{
backup_get_backup_plan(
  BackupPlanId,
  VersionId = NULL,
  MaxScheduledRunsPreview = NULL
)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}

\item{VersionId}{Unique, randomly generated, Unicode, UTF-8 encoded strings that are at
most 1,024 bytes long. Version IDs cannot be edited.}

\item{MaxScheduledRunsPreview}{Number of future scheduled backup runs to preview. When set to 0
(default), no scheduled runs preview is included in the response. Valid
range is 0-10.}
}
\description{
Returns \code{BackupPlan} details for the specified \code{BackupPlanId}. The details are the body of a backup plan in JSON format, in addition to plan metadata.

See \url{https://www.paws-r-sdk.com/docs/backup_get_backup_plan/} for full documentation.
}
\keyword{internal}
