% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_backup_plan_from_json}
\alias{backup_get_backup_plan_from_json}
\title{Returns a valid JSON document specifying a backup plan or an error}
\usage{
backup_get_backup_plan_from_json(BackupPlanTemplateJson)
}
\arguments{
\item{BackupPlanTemplateJson}{[required] A customer-supplied backup plan document in JSON format.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BackupPlan = list(
    BackupPlanName = "string",
    Rules = list(
      list(
        RuleName = "string",
        TargetBackupVaultName = "string",
        ScheduleExpression = "string",
        StartWindowMinutes = 123,
        CompletionWindowMinutes = 123,
        Lifecycle = list(
          MoveToColdStorageAfterDays = 123,
          DeleteAfterDays = 123
        ),
        RecoveryPointTags = list(
          "string"
        ),
        RuleId = "string",
        CopyActions = list(
          list(
            Lifecycle = list(
              MoveToColdStorageAfterDays = 123,
              DeleteAfterDays = 123
            ),
            DestinationBackupVaultArn = "string"
          )
        )
      )
    ),
    AdvancedBackupSettings = list(
      list(
        ResourceType = "string",
        BackupOptions = list(
          "string"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a valid JSON document specifying a backup plan or an error.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_backup_plan_from_json(
  BackupPlanTemplateJson = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
