% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_legal_hold}
\alias{backup_get_legal_hold}
\title{This action returns details for a specified legal hold}
\usage{
backup_get_legal_hold(LegalHoldId)
}
\arguments{
\item{LegalHoldId}{[required] This is the ID required to use
\code{\link[=backup_get_legal_hold]{get_legal_hold}}. This unique ID is associated
with a specific legal hold.}
}
\description{
This action returns details for a specified legal hold. The details are the body of a legal hold in JSON format, in addition to metadata.

See \url{https://paws-r.github.io/docs/backup/get_legal_hold.html} for full documentation.
}
\keyword{internal}
