% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_supported_resource_types}
\alias{backup_get_supported_resource_types}
\title{Returns the AWS resource types supported by AWS Backup}
\usage{
backup_get_supported_resource_types()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceTypes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the AWS resource types supported by AWS Backup.
}
\keyword{internal}
