% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_get_supported_resource_types}
\alias{backup_get_supported_resource_types}
\title{Returns the AWS resource types supported by AWS Backup}
\usage{
backup_get_supported_resource_types()
}
\description{
Returns the AWS resource types supported by AWS Backup.
}
\keyword{internal}
