% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_backup_plan_templates}
\alias{backup_list_backup_plan_templates}
\title{Returns metadata of your saved backup plan templates, including the
template ID, name, and the creation and deletion dates}
\usage{
backup_list_backup_plan_templates(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The next item following a partial list of returned items. For example,
if a request is made to return \code{maxResults} number of items, \code{NextToken}
allows you to return more items in your list starting at the location
pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextToken = "string",
  BackupPlanTemplatesList = list(
    list(
      BackupPlanTemplateId = "string",
      BackupPlanTemplateName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns metadata of your saved backup plan templates, including the
template ID, name, and the creation and deletion dates.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_backup_plan_templates(
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
