% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_tags}
\alias{backup_list_tags}
\title{Returns a list of key-value pairs assigned to a target recovery point,
backup plan, or backup vault}
\usage{
backup_list_tags(ResourceArn, NextToken, MaxResults)
}
\arguments{
\item{ResourceArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies a resource. The
format of the ARN depends on the type of resource. Valid targets for
\code{ListTags} are recovery points, backup plans, and backup vaults.}

\item{NextToken}{The next item following a partial list of returned items. For example,
if a request is made to return \code{maxResults} number of items, \code{NextToken}
allows you to return more items in your list starting at the location
pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\description{
Returns a list of key-value pairs assigned to a target recovery point,
backup plan, or backup vault.

\code{ListTags} are currently only supported with Amazon EFS backups.
}
\section{Request syntax}{
\preformatted{svc$list_tags(
  ResourceArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
