% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_tag_resource}
\alias{backup_tag_resource}
\title{Assigns a set of key-value pairs to a resource}
\usage{
backup_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN that uniquely identifies the resource.}

\item{Tags}{[required] Key-value pairs that are used to help organize your resources. You can
assign your own metadata to the resources you create. For clarity, this
is the structure to assign tags: \verb{[\{"Key":"string","Value":"string"\}]}.}
}
\description{
Assigns a set of key-value pairs to a resource.

See \url{https://www.paws-r-sdk.com/docs/backup_tag_resource/} for full documentation.
}
\keyword{internal}
