% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_update_global_settings}
\alias{backup_update_global_settings}
\title{Updates the current global settings for the AWS account}
\usage{
backup_update_global_settings(GlobalSettings)
}
\arguments{
\item{GlobalSettings}{A list of resources along with the opt-in preferences for the account.}
}
\value{
An empty list.
}
\description{
Updates the current global settings for the AWS account. Use the
\code{\link[=backup_describe_global_settings]{describe_global_settings}} API to
determine the current settings.
}
\section{Request syntax}{
\preformatted{svc$update_global_settings(
  GlobalSettings = list(
    "string"
  )
)
}
}

\keyword{internal}
