% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_update_region_settings}
\alias{backup_update_region_settings}
\title{Updates the current service opt-in settings for the Region}
\usage{
backup_update_region_settings(ResourceTypeOptInPreference)
}
\arguments{
\item{ResourceTypeOptInPreference}{Updates the list of services along with the opt-in preferences for the
Region.}
}
\description{
Updates the current service opt-in settings for the Region. If
service-opt-in is enabled for a service, AWS Backup tries to protect
that service's resources in this Region, when the resource is included
in an on-demand backup or scheduled backup plan. Otherwise, AWS Backup
does not try to protect that service's resources in this Region. Use the
\code{DescribeRegionSettings} API to determine the resource types that are
supported.
}
\section{Request syntax}{
\preformatted{svc$update_region_settings(
  ResourceTypeOptInPreference = list(
    TRUE|FALSE
  )
)
}
}

\keyword{internal}
