% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_list_chunks}
\alias{backupstorage_list_chunks}
\title{List chunks in a given Object}
\usage{
backupstorage_list_chunks(
  StorageJobId,
  ObjectToken,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StorageJobId}{[required] Storage job id}

\item{ObjectToken}{[required] Object token}

\item{MaxResults}{Maximum number of chunks}

\item{NextToken}{Pagination token}
}
\description{
List chunks in a given Object

See \url{https://paws-r.github.io/docs/backupstorage/list_chunks.html} for full documentation.
}
\keyword{internal}
