% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupstorage_operations.R
\name{backupstorage_start_object}
\alias{backupstorage_start_object}
\title{Start upload containing one or many chunks}
\usage{
backupstorage_start_object(BackupJobId, ObjectName, ThrowOnDuplicate = NULL)
}
\arguments{
\item{BackupJobId}{[required] Backup job Id for the in-progress backup}

\item{ObjectName}{[required] Name for the object.}

\item{ThrowOnDuplicate}{Throw an exception if Object name is already exist.}
}
\description{
Start upload containing one or many chunks.

See \url{https://paws-r.github.io/docs/backupstorage/start_object.html} for full documentation.
}
\keyword{internal}
