% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_create_lifecycle_policy}
\alias{dlm_create_lifecycle_policy}
\title{Creates a policy to manage the lifecycle of the specified Amazon Web
Services resources}
\usage{
dlm_create_lifecycle_policy(
  ExecutionRoleArn,
  Description,
  State,
  PolicyDetails,
  Tags = NULL
)
}
\arguments{
\item{ExecutionRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role used to run the
operations specified by the lifecycle policy.}

\item{Description}{[required] A description of the lifecycle policy. The characters ^[0-9A-Za-z
_-]+$ are supported.}

\item{State}{[required] The desired activation state of the lifecycle policy after creation.}

\item{PolicyDetails}{[required] The configuration details of the lifecycle policy.}

\item{Tags}{The tags to apply to the lifecycle policy during creation.}
}
\description{
Creates a policy to manage the lifecycle of the specified Amazon Web Services resources. You can create up to 100 lifecycle policies.

See \url{https://www.paws-r-sdk.com/docs/dlm_create_lifecycle_policy/} for full documentation.
}
\keyword{internal}
