% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_get_lifecycle_policies}
\alias{dlm_get_lifecycle_policies}
\title{Gets summary information about all or the specified data lifecycle
policies}
\usage{
dlm_get_lifecycle_policies(
  PolicyIds = NULL,
  State = NULL,
  ResourceTypes = NULL,
  TargetTags = NULL,
  TagsToAdd = NULL
)
}
\arguments{
\item{PolicyIds}{The identifiers of the data lifecycle policies.}

\item{State}{The activation state.}

\item{ResourceTypes}{The resource type.}

\item{TargetTags}{The target tag for a policy.

Tags are strings in the format \code{key=value}.}

\item{TagsToAdd}{The tags to add to objects created by the policy.

Tags are strings in the format \code{key=value}.

These user-defined tags are added in addition to the Amazon Web
Services-added lifecycle tags.}
}
\description{
Gets summary information about all or the specified data lifecycle policies.

See \url{https://paws-r.github.io/docs/dlm/get_lifecycle_policies.html} for full documentation.
}
\keyword{internal}
