% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_tag_resource}
\alias{dlm_tag_resource}
\title{Adds the specified tags to the specified resource}
\usage{
dlm_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] One or more tags.}
}
\description{
Adds the specified tags to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/dlm_tag_resource/} for full documentation.
}
\keyword{internal}
