% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_operations.R
\name{dlm_update_lifecycle_policy}
\alias{dlm_update_lifecycle_policy}
\title{Updates the specified lifecycle policy}
\usage{
dlm_update_lifecycle_policy(
  PolicyId,
  ExecutionRoleArn = NULL,
  State = NULL,
  Description = NULL,
  PolicyDetails = NULL
)
}
\arguments{
\item{PolicyId}{[required] The identifier of the lifecycle policy.}

\item{ExecutionRoleArn}{The Amazon Resource Name (ARN) of the IAM role used to run the
operations specified by the lifecycle policy.}

\item{State}{The desired activation state of the lifecycle policy after creation.}

\item{Description}{A description of the lifecycle policy.}

\item{PolicyDetails}{The configuration of the lifecycle policy. You cannot update the policy
type or the resource type.}
}
\description{
Updates the specified lifecycle policy.

See \url{https://www.paws-r-sdk.com/docs/dlm_update_lifecycle_policy/} for full documentation.
}
\keyword{internal}
