% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_create_replication_configuration}
\alias{efs_create_replication_configuration}
\title{Creates a replication configuration that replicates an existing EFS file
system to a new, read-only file system}
\usage{
efs_create_replication_configuration(SourceFileSystemId, Destinations)
}
\arguments{
\item{SourceFileSystemId}{[required] Specifies the Amazon EFS file system that you want to replicate. This
file system cannot already be a source or destination file system in
another replication configuration.}

\item{Destinations}{[required] An array of destination configuration objects. Only one destination
configuration object is supported.}
}
\description{
Creates a replication configuration that replicates an existing EFS file system to a new, read-only file system. For more information, see \href{https://docs.aws.amazon.com/efs/latest/ug/efs-replication.html}{Amazon EFS replication} in the \emph{Amazon EFS User Guide}. The replication configuration specifies the following:

See \url{https://paws-r.github.io/docs/efs/create_replication_configuration.html} for full documentation.
}
\keyword{internal}
