% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_delete_access_point}
\alias{efs_delete_access_point}
\title{Deletes the specified access point}
\usage{
efs_delete_access_point(AccessPointId)
}
\arguments{
\item{AccessPointId}{[required] The ID of the access point that you want to delete.}
}
\description{
Deletes the specified access point. After deletion is complete, new clients can no longer connect to the access points. Clients connected to the access point at the time of deletion will continue to function until they terminate their connection.

See \url{https://paws-r.github.io/docs/efs/delete_access_point.html} for full documentation.
}
\keyword{internal}
