% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_delete_mount_target}
\alias{efs_delete_mount_target}
\title{Deletes the specified mount target}
\usage{
efs_delete_mount_target(MountTargetId)
}
\arguments{
\item{MountTargetId}{[required] The ID of the mount target to delete (String).}
}
\description{
Deletes the specified mount target.

See \url{https://www.paws-r-sdk.com/docs/efs_delete_mount_target/} for full documentation.
}
\keyword{internal}
