% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_delete_replication_configuration}
\alias{efs_delete_replication_configuration}
\title{Deletes an existing replication configuration}
\usage{
efs_delete_replication_configuration(SourceFileSystemId)
}
\arguments{
\item{SourceFileSystemId}{[required] The ID of the source file system in the replication configuration.}
}
\description{
Deletes an existing replication configuration. To delete a replication configuration, you must make the request from the Amazon Web Services Region in which the destination file system is located. Deleting a replication configuration ends the replication process. After a replication configuration is deleted, the destination file system is no longer read-only. You can write to the destination file system after its status becomes \code{Writeable}.

See \url{https://www.paws-r-sdk.com/docs/efs_delete_replication_configuration/} for full documentation.
}
\keyword{internal}
