% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_delete_replication_configuration}
\alias{efs_delete_replication_configuration}
\title{Deletes a replication configuration}
\usage{
efs_delete_replication_configuration(SourceFileSystemId)
}
\arguments{
\item{SourceFileSystemId}{[required] The ID of the source file system in the replication configuration.}
}
\description{
Deletes a replication configuration. Deleting a replication configuration ends the replication process. After a replication configuration is deleted, the destination file system becomes \code{Writeable} and its replication overwrite protection is re-enabled. For more information, see \href{https://docs.aws.amazon.com/efs/latest/ug/delete-replications.html}{Delete a replication configuration}.

See \url{https://www.paws-r-sdk.com/docs/efs_delete_replication_configuration/} for full documentation.
}
\keyword{internal}
