% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_describe_file_system_policy}
\alias{efs_describe_file_system_policy}
\title{Returns the FileSystemPolicy for the specified EFS file system}
\usage{
efs_describe_file_system_policy(FileSystemId)
}
\arguments{
\item{FileSystemId}{[required] Specifies which EFS file system to retrieve the \code{FileSystemPolicy} for.}
}
\description{
Returns the \code{FileSystemPolicy} for the specified EFS file system.

See \url{https://paws-r.github.io/docs/efs/describe_file_system_policy.html} for full documentation.
}
\keyword{internal}
