% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_put_backup_policy}
\alias{efs_put_backup_policy}
\title{Updates the file system's backup policy}
\usage{
efs_put_backup_policy(FileSystemId, BackupPolicy)
}
\arguments{
\item{FileSystemId}{[required] Specifies which EFS file system to update the backup policy for.}

\item{BackupPolicy}{[required] The backup policy included in the
\code{\link[=efs_put_backup_policy]{put_backup_policy}} request.}
}
\description{
Updates the file system's backup policy. Use this action to start or stop automatic backups of the file system.

See \url{https://paws-r.github.io/docs/efs/put_backup_policy.html} for full documentation.
}
\keyword{internal}
