% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_tag_resource}
\alias{efs_tag_resource}
\title{Creates a tag for an EFS resource}
\usage{
efs_tag_resource(ResourceId, Tags)
}
\arguments{
\item{ResourceId}{[required] The ID specifying the EFS resource that you want to create a tag for.}

\item{Tags}{[required]}
}
\description{
Creates a tag for an EFS resource. You can create tags for EFS file
systems and access points using this API operation.
}
\details{
This operation requires permissions for the
\code{elasticfilesystem:TagResource} action.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
