% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_tag_resource}
\alias{efs_tag_resource}
\title{Creates a tag for an EFS resource}
\usage{
efs_tag_resource(ResourceId, Tags)
}
\arguments{
\item{ResourceId}{[required] The ID specifying the EFS resource that you want to create a tag for.}

\item{Tags}{[required]}
}
\value{
An empty list.
}
\description{
Creates a tag for an EFS resource. You can create tags for EFS file
systems and access points using this API operation.

This operation requires permissions for the
\code{elasticfilesystem:TagResource} action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$tag_resource(
  ResourceId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
