% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_delete_dataset}
\alias{finspacedata_delete_dataset}
\title{Deletes a FinSpace Dataset}
\usage{
finspacedata_delete_dataset(clientToken = NULL, datasetId)
}
\arguments{
\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}

\item{datasetId}{[required] The unique identifier of the Dataset to be deleted.}
}
\description{
Deletes a FinSpace Dataset.

See \url{https://paws-r.github.io/docs/finspacedata/delete_dataset.html} for full documentation.
}
\keyword{internal}
