% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_data_view}
\alias{finspacedata_get_data_view}
\title{Gets information about a Dataview}
\usage{
finspacedata_get_data_view(dataViewId, datasetId)
}
\arguments{
\item{dataViewId}{[required] The unique identifier for the Dataview.}

\item{datasetId}{[required] The unique identifier for the Dataset used in the Dataview.}
}
\description{
Gets information about a Dataview.

See \url{https://paws-r.github.io/docs/finspacedata/get_data_view.html} for full documentation.
}
\keyword{internal}
