% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_permission_group}
\alias{finspacedata_get_permission_group}
\title{Retrieves the details of a specific permission group}
\usage{
finspacedata_get_permission_group(permissionGroupId)
}
\arguments{
\item{permissionGroupId}{[required] The unique identifier for the permission group.}
}
\description{
Retrieves the details of a specific permission group.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_get_permission_group/} for full documentation.
}
\keyword{internal}
