% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspacedata_operations.R
\name{finspacedata_get_programmatic_access_credentials}
\alias{finspacedata_get_programmatic_access_credentials}
\title{Request programmatic credentials to use with FinSpace SDK}
\usage{
finspacedata_get_programmatic_access_credentials(
  durationInMinutes = NULL,
  environmentId
)
}
\arguments{
\item{durationInMinutes}{The time duration in which the credentials remain valid.}

\item{environmentId}{[required] The FinSpace environment identifier.}
}
\description{
Request programmatic credentials to use with FinSpace SDK.

See \url{https://www.paws-r-sdk.com/docs/finspacedata_get_programmatic_access_credentials/} for full documentation.
}
\keyword{internal}
