% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_create_snapshot}
\alias{fsx_create_snapshot}
\title{Creates a snapshot of an existing Amazon FSx for OpenZFS volume}
\usage{
fsx_create_snapshot(ClientRequestToken = NULL, Name, VolumeId, Tags = NULL)
}
\arguments{
\item{ClientRequestToken}{}

\item{Name}{[required] The name of the snapshot.}

\item{VolumeId}{[required] The ID of the volume that you are taking a snapshot of.}

\item{Tags}{}
}
\description{
Creates a snapshot of an existing Amazon FSx for OpenZFS volume. With snapshots, you can easily undo file changes and compare file versions by restoring the volume to a previous version.

See \url{https://paws-r.github.io/docs/fsx/create_snapshot.html} for full documentation.
}
\keyword{internal}
