% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_create_volume_from_backup}
\alias{fsx_create_volume_from_backup}
\title{Creates a new Amazon FSx for NetApp ONTAP volume from an existing Amazon
FSx volume backup}
\usage{
fsx_create_volume_from_backup(
  BackupId,
  ClientRequestToken = NULL,
  Name,
  OntapConfiguration = NULL,
  Tags = NULL
)
}
\arguments{
\item{BackupId}{[required]}

\item{ClientRequestToken}{}

\item{Name}{[required] The name of the new volume you're creating.}

\item{OntapConfiguration}{Specifies the configuration of the ONTAP volume that you are creating.}

\item{Tags}{}
}
\description{
Creates a new Amazon FSx for NetApp ONTAP volume from an existing Amazon FSx volume backup.

See \url{https://www.paws-r-sdk.com/docs/fsx_create_volume_from_backup/} for full documentation.
}
\keyword{internal}
