% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_delete_backup}
\alias{fsx_delete_backup}
\title{Deletes an Amazon FSx backup}
\usage{
fsx_delete_backup(BackupId, ClientRequestToken = NULL)
}
\arguments{
\item{BackupId}{[required] The ID of the backup that you want to delete.}

\item{ClientRequestToken}{A string of up to 64 ASCII characters that Amazon FSx uses to ensure
idempotent deletion. This parameter is automatically filled on your
behalf when using the CLI or SDK.}
}
\description{
Deletes an Amazon FSx backup. After deletion, the backup no longer exists, and its data is gone.

See \url{https://paws-r.github.io/docs/fsx/delete_backup.html} for full documentation.
}
\keyword{internal}
