% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_delete_storage_virtual_machine}
\alias{fsx_delete_storage_virtual_machine}
\title{Deletes an existing Amazon FSx for ONTAP storage virtual machine (SVM)}
\usage{
fsx_delete_storage_virtual_machine(
  ClientRequestToken = NULL,
  StorageVirtualMachineId
)
}
\arguments{
\item{ClientRequestToken}{}

\item{StorageVirtualMachineId}{[required] The ID of the SVM that you want to delete.}
}
\description{
Deletes an existing Amazon FSx for ONTAP storage virtual machine (SVM). Prior to deleting an SVM, you must delete all non-root volumes in the SVM, otherwise the operation will fail.

See \url{https://paws-r.github.io/docs/fsx/delete_storage_virtual_machine.html} for full documentation.
}
\keyword{internal}
