% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_describe_s3_access_point_attachments}
\alias{fsx_describe_s3_access_point_attachments}
\title{Describes one or more S3 access points attached to Amazon FSx volumes}
\usage{
fsx_describe_s3_access_point_attachments(
  Names = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Names}{The names of the S3 access point attachments whose descriptions you want
to retrieve.}

\item{Filters}{Enter a filter Name and Values pair to view a select set of S3 access
point attachments.}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Describes one or more S3 access points attached to Amazon FSx volumes.

See \url{https://www.paws-r-sdk.com/docs/fsx_describe_s3_access_point_attachments/} for full documentation.
}
\keyword{internal}
