% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_describe_storage_virtual_machines}
\alias{fsx_describe_storage_virtual_machines}
\title{Describes one or more Amazon FSx for NetApp ONTAP storage virtual
machines (SVMs)}
\usage{
fsx_describe_storage_virtual_machines(
  StorageVirtualMachineIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StorageVirtualMachineIds}{Enter the ID of one or more SVMs that you want to view.}

\item{Filters}{Enter a filter name:value pair to view a select set of SVMs.}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Describes one or more Amazon FSx for NetApp ONTAP storage virtual machines (SVMs).

See \url{https://paws-r.github.io/docs/fsx/describe_storage_virtual_machines.html} for full documentation.
}
\keyword{internal}
