% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_detach_and_delete_s3_access_point}
\alias{fsx_detach_and_delete_s3_access_point}
\title{Detaches an S3 access point from an Amazon FSx volume and deletes the S3
access point}
\usage{
fsx_detach_and_delete_s3_access_point(ClientRequestToken = NULL, Name)
}
\arguments{
\item{ClientRequestToken}{}

\item{Name}{[required] The name of the S3 access point attachment that you want to delete.}
}
\description{
Detaches an S3 access point from an Amazon FSx volume and deletes the S3 access point.

See \url{https://www.paws-r-sdk.com/docs/fsx_detach_and_delete_s3_access_point/} for full documentation.
}
\keyword{internal}
