% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_update_file_cache}
\alias{fsx_update_file_cache}
\title{Updates the configuration of an existing Amazon File Cache resource}
\usage{
fsx_update_file_cache(
  FileCacheId,
  ClientRequestToken = NULL,
  LustreConfiguration = NULL
)
}
\arguments{
\item{FileCacheId}{[required] The ID of the cache that you are updating.}

\item{ClientRequestToken}{}

\item{LustreConfiguration}{The configuration updates for an Amazon File Cache resource.}
}
\description{
Updates the configuration of an existing Amazon File Cache resource. You can update multiple properties in a single request.

See \url{https://www.paws-r-sdk.com/docs/fsx_update_file_cache/} for full documentation.
}
\keyword{internal}
