% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsx_operations.R
\name{fsx_update_storage_virtual_machine}
\alias{fsx_update_storage_virtual_machine}
\title{Updates an FSx for ONTAP storage virtual machine (SVM)}
\usage{
fsx_update_storage_virtual_machine(
  ActiveDirectoryConfiguration = NULL,
  ClientRequestToken = NULL,
  StorageVirtualMachineId,
  SvmAdminPassword = NULL
)
}
\arguments{
\item{ActiveDirectoryConfiguration}{Specifies updates to an SVM's Microsoft Active Directory (AD)
configuration.}

\item{ClientRequestToken}{}

\item{StorageVirtualMachineId}{[required] The ID of the SVM that you want to update, in the format
\verb{svm-0123456789abcdef0}.}

\item{SvmAdminPassword}{Specifies a new SvmAdminPassword.}
}
\description{
Updates an FSx for ONTAP storage virtual machine (SVM).

See \url{https://www.paws-r-sdk.com/docs/fsx_update_storage_virtual_machine/} for full documentation.
}
\keyword{internal}
