% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier_operations.R
\name{glacier_list_jobs}
\alias{glacier_list_jobs}
\title{This operation lists jobs for a vault, including jobs that are
in-progress and jobs that have recently finished}
\usage{
glacier_list_jobs(
  accountId,
  vaultName,
  limit = NULL,
  marker = NULL,
  statuscode = NULL,
  completed = NULL
)
}
\arguments{
\item{accountId}{[required] The \code{AccountId} value is the AWS account ID of the account that owns the
vault. You can either specify an AWS account ID or optionally a single
'\code{-}' (hyphen), in which case Amazon Glacier uses the AWS account ID
associated with the credentials used to sign the request. If you use an
account ID, do not include any hyphens ('-') in the ID.}

\item{vaultName}{[required] The name of the vault.}

\item{limit}{The maximum number of jobs to be returned. The default limit is 50. The
number of jobs returned might be fewer than the specified limit, but the
number of returned jobs never exceeds the limit.}

\item{marker}{An opaque string used for pagination. This value specifies the job at
which the listing of jobs should begin. Get the marker value from a
previous List Jobs response. You only need to include the marker if you
are continuing the pagination of results started in a previous List Jobs
request.}

\item{statuscode}{The type of job status to return. You can specify the following values:
\code{InProgress}, \code{Succeeded}, or \code{Failed}.}

\item{completed}{The state of the jobs to return. You can specify \code{true} or \code{false}.}
}
\description{
This operation lists jobs for a vault, including jobs that are in-progress and jobs that have recently finished. The List Job operation returns a list of these jobs sorted by job initiation time.

See \url{https://www.paws-r-sdk.com/docs/glacier_list_jobs/} for full documentation.
}
\keyword{internal}
