% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier_operations.R
\name{glacier_purchase_provisioned_capacity}
\alias{glacier_purchase_provisioned_capacity}
\title{This operation purchases a provisioned capacity unit for an AWS account}
\usage{
glacier_purchase_provisioned_capacity(accountId)
}
\arguments{
\item{accountId}{[required] The AWS account ID of the account that owns the vault. You can either
specify an AWS account ID or optionally a single '-' (hyphen), in which
case Amazon Glacier uses the AWS account ID associated with the
credentials used to sign the request. If you use an account ID, don't
include any hyphens ('-') in the ID.}
}
\description{
This operation purchases a provisioned capacity unit for an AWS account.

See \url{https://www.paws-r-sdk.com/docs/glacier_purchase_provisioned_capacity/} for full documentation.
}
\keyword{internal}
