% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_accept_share}
\alias{omics_accept_share}
\title{Accept a resource share request}
\usage{
omics_accept_share(shareId)
}
\arguments{
\item{shareId}{[required] The ID of the resource share.}
}
\description{
Accept a resource share request.

See \url{https://www.paws-r-sdk.com/docs/omics_accept_share/} for full documentation.
}
\keyword{internal}
