% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_batch_delete_read_set}
\alias{omics_batch_delete_read_set}
\title{Deletes one or more read sets}
\usage{
omics_batch_delete_read_set(ids, sequenceStoreId)
}
\arguments{
\item{ids}{[required] The read sets' IDs.}

\item{sequenceStoreId}{[required] The read sets' sequence store ID.}
}
\description{
Deletes one or more read sets. If the operation is successful, it returns a response with no body. If there is an error with deleting one of the read sets, the operation returns an error list. If the operation successfully deletes only a subset of files, it will return an error list for the remaining files that fail to be deleted. There is a limit of 100 read sets that can be deleted in each \code{\link[=omics_batch_delete_read_set]{batch_delete_read_set}} API call.

See \url{https://www.paws-r-sdk.com/docs/omics_batch_delete_read_set/} for full documentation.
}
\keyword{internal}
