% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_multipart_read_set_upload}
\alias{omics_create_multipart_read_set_upload}
\title{Initiates a multipart read set upload for uploading partitioned source
files into a sequence store}
\usage{
omics_create_multipart_read_set_upload(
  sequenceStoreId,
  clientToken = NULL,
  sourceFileType,
  subjectId,
  sampleId,
  generatedFrom = NULL,
  referenceArn = NULL,
  name,
  description = NULL,
  tags = NULL
)
}
\arguments{
\item{sequenceStoreId}{[required] The sequence store ID for the store that is the destination of the
multipart uploads.}

\item{clientToken}{An idempotency token that can be used to avoid triggering multiple
multipart uploads.}

\item{sourceFileType}{[required] The type of file being uploaded.}

\item{subjectId}{[required] The source's subject ID.}

\item{sampleId}{[required] The source's sample ID.}

\item{generatedFrom}{Where the source originated.}

\item{referenceArn}{The ARN of the reference.}

\item{name}{[required] The name of the read set.}

\item{description}{The description of the read set.}

\item{tags}{Any tags to add to the read set.}
}
\description{
Initiates a multipart read set upload for uploading partitioned source files into a sequence store. You can directly import source files from an EC2 instance and other local compute, or from an S3 bucket. To separate these source files into parts, use the \code{split} operation. Each part cannot be larger than 100 MB. If the operation is successful, it provides an \code{uploadId} which is required by the \code{\link[=omics_upload_read_set_part]{upload_read_set_part}} API operation to upload parts into a sequence store.

See \url{https://www.paws-r-sdk.com/docs/omics_create_multipart_read_set_upload/} for full documentation.
}
\keyword{internal}
