% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_run_group}
\alias{omics_create_run_group}
\title{You can optionally create a run group to limit the compute resources for
the runs that you add to the group}
\usage{
omics_create_run_group(
  name = NULL,
  maxCpus = NULL,
  maxRuns = NULL,
  maxDuration = NULL,
  tags = NULL,
  requestId,
  maxGpus = NULL
)
}
\arguments{
\item{name}{A name for the group.}

\item{maxCpus}{The maximum number of CPUs that can run concurrently across all active
runs in the run group.}

\item{maxRuns}{The maximum number of runs that can be running at the same time.}

\item{maxDuration}{The maximum time for each run (in minutes). If a run exceeds the maximum
run time, the run fails automatically.}

\item{tags}{Tags for the group.}

\item{requestId}{[required] To ensure that requests don't run multiple times, specify a unique ID
for each request.}

\item{maxGpus}{The maximum number of GPUs that can run concurrently across all active
runs in the run group.}
}
\description{
You can optionally create a run group to limit the compute resources for the runs that you add to the group.

See \url{https://www.paws-r-sdk.com/docs/omics_create_run_group/} for full documentation.
}
\keyword{internal}
