% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_share}
\alias{omics_create_share}
\title{Creates a share offer that can be accepted outside the account by a
subscriber}
\usage{
omics_create_share(resourceArn, principalSubscriber, shareName = NULL)
}
\arguments{
\item{resourceArn}{[required] The resource ARN for the analytics store to be shared.}

\item{principalSubscriber}{[required] The principal subscriber is the account being given access to the
analytics store data through the share offer.}

\item{shareName}{A name given to the share.}
}
\description{
Creates a share offer that can be accepted outside the account by a subscriber. The share is created by the owner and accepted by the principal subscriber.

See \url{https://www.paws-r-sdk.com/docs/omics_create_share/} for full documentation.
}
\keyword{internal}
