% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_workflow}
\alias{omics_create_workflow}
\title{Creates a workflow}
\usage{
omics_create_workflow(
  name = NULL,
  description = NULL,
  engine = NULL,
  definitionZip = NULL,
  definitionUri = NULL,
  main = NULL,
  parameterTemplate = NULL,
  storageCapacity = NULL,
  tags = NULL,
  requestId,
  accelerators = NULL
)
}
\arguments{
\item{name}{A name for the workflow.}

\item{description}{A description for the workflow.}

\item{engine}{An engine for the workflow.}

\item{definitionZip}{A ZIP archive for the workflow.}

\item{definitionUri}{The URI of a definition for the workflow.}

\item{main}{The path of the main definition file for the workflow.}

\item{parameterTemplate}{A parameter template for the workflow.}

\item{storageCapacity}{A storage capacity for the workflow in gigabytes.}

\item{tags}{Tags for the workflow.}

\item{requestId}{[required] To ensure that requests don't run multiple times, specify a unique ID
for each request.}

\item{accelerators}{The computational accelerator specified to run the workflow.}
}
\description{
Creates a workflow.

See \url{https://www.paws-r-sdk.com/docs/omics_create_workflow/} for full documentation.
}
\keyword{internal}
