% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_create_workflow_version}
\alias{omics_create_workflow_version}
\title{Creates a new workflow version for the workflow that you specify with
the workflowId parameter}
\usage{
omics_create_workflow_version(
  workflowId,
  versionName,
  definitionZip = NULL,
  definitionUri = NULL,
  accelerators = NULL,
  description = NULL,
  engine = NULL,
  main = NULL,
  parameterTemplate = NULL,
  requestId,
  storageType = NULL,
  storageCapacity = NULL,
  tags = NULL,
  workflowBucketOwnerId = NULL,
  containerRegistryMap = NULL,
  containerRegistryMapUri = NULL,
  readmeMarkdown = NULL,
  parameterTemplatePath = NULL,
  readmePath = NULL,
  definitionRepository = NULL,
  readmeUri = NULL
)
}
\arguments{
\item{workflowId}{[required] The ID of the workflow where you are creating the new version. The
\code{workflowId} is not the UUID.}

\item{versionName}{[required] A name for the workflow version. Provide a version name that is unique
for this workflow. You cannot change the name after HealthOmics creates
the version.

The version name must start with a letter or number and it can include
upper-case and lower-case letters, numbers, hyphens, periods and
underscores. The maximum length is 64 characters. You can use a simple
naming scheme, such as version1, version2, version3. You can also match
your workflow versions with your own internal versioning conventions,
such as 2.7.0, 2.7.1, 2.7.2.}

\item{definitionZip}{A ZIP archive containing the main workflow definition file and
dependencies that it imports for this workflow version. You can use a
file with a ://fileb prefix instead of the Base64 string. For more
information, see Workflow definition requirements in the \emph{Amazon Web
Services HealthOmics User Guide}.}

\item{definitionUri}{The S3 URI of a definition for this workflow version. The S3 bucket must
be in the same region as this workflow version.}

\item{accelerators}{The computational accelerator for this workflow version.}

\item{description}{A description for this workflow version.}

\item{engine}{The workflow engine for this workflow version. This is only required if
you have workflow definition files from more than one engine in your zip
file. Otherwise, the service can detect the engine automatically from
your workflow definition.}

\item{main}{The path of the main definition file for this workflow version. This
parameter is not required if the ZIP archive contains only one workflow
definition file, or if the main definition file is named “main”. An
example path is: \code{workflow-definition/main-file.wdl}.}

\item{parameterTemplate}{A parameter template for this workflow version. If this field is blank,
Amazon Web Services HealthOmics will automatically parse the parameter
template values from your workflow definition file. To override these
service generated default values, provide a parameter template. To view
an example of a parameter template, see \href{https://docs.aws.amazon.com/omics/latest/dev/parameter-templates.html}{Parameter template files}
in the \emph{Amazon Web Services HealthOmics User Guide}.}

\item{requestId}{[required] An idempotency token to ensure that duplicate workflows are not created
when Amazon Web Services HealthOmics submits retry requests.}

\item{storageType}{The default storage type for runs that use this workflow version. The
\code{storageType} can be overridden at run time. \code{DYNAMIC} storage
dynamically scales the storage up or down, based on file system
utilization. STATIC storage allocates a fixed amount of storage. For
more information about dynamic and static storage types, see \href{https://docs.aws.amazon.com/omics/latest/dev/workflows-run-types.html}{Run storage types}
in the \emph{Amazon Web Services HealthOmics User Guide}.}

\item{storageCapacity}{The default static storage capacity (in gibibytes) for runs that use
this workflow version. The \code{storageCapacity} can be overwritten at run
time. The storage capacity is not required for runs with a \code{DYNAMIC}
storage type.}

\item{tags}{Tags for this workflow version. You can define up to 50 tags for the
workflow. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/add-a-tag.html}{Adding a tag} in the
\emph{Amazon Web Services HealthOmics User Guide}.}

\item{workflowBucketOwnerId}{Amazon Web Services Id of the owner of the S3 bucket that contains the
workflow definition. You need to specify this parameter if your account
is not the bucket owner.}

\item{containerRegistryMap}{(Optional) Use a container registry map to specify mappings between the
ECR private repository and one or more upstream registries. For more
information, see \href{https://docs.aws.amazon.com/omics/latest/dev/workflows-ecr.html}{Container images}
in the \emph{Amazon Web Services HealthOmics User Guide}.}

\item{containerRegistryMapUri}{(Optional) URI of the S3 location for the registry mapping file.}

\item{readmeMarkdown}{The markdown content for the workflow version's README file. This
provides documentation and usage information for users of this specific
workflow version.}

\item{parameterTemplatePath}{The path to the workflow version parameter template JSON file within the
repository. This file defines the input parameters for runs that use
this workflow version. If not specified, the workflow version will be
created without a parameter template.}

\item{readmePath}{The path to the workflow version README markdown file within the
repository. This file provides documentation and usage information for
the workflow. If not specified, the \code{README.md} file from the root
directory of the repository will be used.}

\item{definitionRepository}{The repository information for the workflow version definition. This
allows you to source your workflow version definition directly from a
code repository.}

\item{readmeUri}{The S3 URI of the README file for the workflow version. This file
provides documentation and usage information for the workflow version.
Requirements include:
\itemize{
\item The S3 URI must begin with \verb{s3://USER-OWNED-BUCKET/}
\item The requester must have access to the S3 bucket and object.
\item The max README content length is 500 KiB.
}}
}
\description{
Creates a new workflow version for the workflow that you specify with the \code{workflowId} parameter.

See \url{https://www.paws-r-sdk.com/docs/omics_create_workflow_version/} for full documentation.
}
\keyword{internal}
