% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_annotation_store}
\alias{omics_delete_annotation_store}
\title{Deletes an annotation store}
\usage{
omics_delete_annotation_store(name, force = NULL)
}
\arguments{
\item{name}{[required] The store's name.}

\item{force}{Whether to force deletion.}
}
\description{
Deletes an annotation store.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_annotation_store/} for full documentation.
}
\keyword{internal}
