% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_reference}
\alias{omics_delete_reference}
\title{Deletes a reference genome and returns a response with no body if the
operation is successful}
\usage{
omics_delete_reference(id, referenceStoreId)
}
\arguments{
\item{id}{[required] The reference's ID.}

\item{referenceStoreId}{[required] The reference's store ID.}
}
\description{
Deletes a reference genome and returns a response with no body if the operation is successful. The read set associated with the reference genome must first be deleted before deleting the reference genome. After the reference genome is deleted, you can delete the reference store using the \code{\link[=omics_delete_reference_store]{delete_reference_store}} API operation.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_reference/} for full documentation.
}
\keyword{internal}
