% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_reference_store}
\alias{omics_delete_reference_store}
\title{Deletes a reference store and returns a response with no body if the
operation is successful}
\usage{
omics_delete_reference_store(id)
}
\arguments{
\item{id}{[required] The store's ID.}
}
\description{
Deletes a reference store and returns a response with no body if the operation is successful. You can only delete a reference store when it does not contain any reference genomes. To empty a reference store, use \code{\link[=omics_delete_reference]{delete_reference}}.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_reference_store/} for full documentation.
}
\keyword{internal}
