% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_sequence_store}
\alias{omics_delete_sequence_store}
\title{Deletes a sequence store and returns a response with no body if the
operation is successful}
\usage{
omics_delete_sequence_store(id)
}
\arguments{
\item{id}{[required] The sequence store's ID.}
}
\description{
Deletes a sequence store and returns a response with no body if the operation is successful. You can only delete a sequence store when it does not contain any read sets.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_sequence_store/} for full documentation.
}
\keyword{internal}
