% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_delete_share}
\alias{omics_delete_share}
\title{Deletes a share of an analytics store}
\usage{
omics_delete_share(shareId)
}
\arguments{
\item{shareId}{[required] The ID for the share request to be deleted.}
}
\description{
Deletes a share of an analytics store.

See \url{https://www.paws-r-sdk.com/docs/omics_delete_share/} for full documentation.
}
\keyword{internal}
