% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_read_set}
\alias{omics_get_read_set}
\title{Retrieves detailed information from parts of a read set and returns the
read set in the same format that it was uploaded}
\usage{
omics_get_read_set(id, sequenceStoreId, file = NULL, partNumber)
}
\arguments{
\item{id}{[required] The read set's ID.}

\item{sequenceStoreId}{[required] The read set's sequence store ID.}

\item{file}{The file to retrieve.}

\item{partNumber}{[required] The part number to retrieve.}
}
\description{
Retrieves detailed information from parts of a read set and returns the read set in the same format that it was uploaded. You must have read sets uploaded to your sequence store in order to run this operation.

See \url{https://www.paws-r-sdk.com/docs/omics_get_read_set/} for full documentation.
}
\keyword{internal}
