% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_reference_import_job}
\alias{omics_get_reference_import_job}
\title{Monitors the status of a reference import job}
\usage{
omics_get_reference_import_job(id, referenceStoreId)
}
\arguments{
\item{id}{[required] The job's ID.}

\item{referenceStoreId}{[required] The job's reference store ID.}
}
\description{
Monitors the status of a reference import job. This operation can be called after calling the \code{\link[=omics_start_reference_import_job]{start_reference_import_job}} operation.

See \url{https://www.paws-r-sdk.com/docs/omics_get_reference_import_job/} for full documentation.
}
\keyword{internal}
