% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_reference_metadata}
\alias{omics_get_reference_metadata}
\title{Retrieves metadata for a reference genome}
\usage{
omics_get_reference_metadata(id, referenceStoreId)
}
\arguments{
\item{id}{[required] The reference's ID.}

\item{referenceStoreId}{[required] The reference's reference store ID.}
}
\description{
Retrieves metadata for a reference genome. This operation returns the number of parts, part size, and MD5 of an entire file. This operation does not return tags. To retrieve the list of tags for a read set, use the \code{\link[=omics_list_tags_for_resource]{list_tags_for_resource}} API operation.

See \url{https://www.paws-r-sdk.com/docs/omics_get_reference_metadata/} for full documentation.
}
\keyword{internal}
