% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_run}
\alias{omics_get_run}
\title{Gets detailed information about a specific run using its ID}
\usage{
omics_get_run(id, export = NULL)
}
\arguments{
\item{id}{[required] The run's ID.}

\item{export}{The run's export format.}
}
\description{
Gets detailed information about a specific run using its ID.

See \url{https://www.paws-r-sdk.com/docs/omics_get_run/} for full documentation.
}
\keyword{internal}
