% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_run_cache}
\alias{omics_get_run_cache}
\title{Retrieves detailed information about the specified run cache using its
ID}
\usage{
omics_get_run_cache(id)
}
\arguments{
\item{id}{[required] The identifier of the run cache to retrieve.}
}
\description{
Retrieves detailed information about the specified run cache using its ID.

See \url{https://www.paws-r-sdk.com/docs/omics_get_run_cache/} for full documentation.
}
\keyword{internal}
