% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_sequence_store}
\alias{omics_get_sequence_store}
\title{Retrieves metadata for a sequence store using its ID and returns it in
JSON format}
\usage{
omics_get_sequence_store(id)
}
\arguments{
\item{id}{[required] The store's ID.}
}
\description{
Retrieves metadata for a sequence store using its ID and returns it in JSON format.

See \url{https://www.paws-r-sdk.com/docs/omics_get_sequence_store/} for full documentation.
}
\keyword{internal}
