% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_get_share}
\alias{omics_get_share}
\title{Retrieves the metadata for a share}
\usage{
omics_get_share(shareId)
}
\arguments{
\item{shareId}{[required] The generated ID for a share.}
}
\description{
Retrieves the metadata for a share.

See \url{https://www.paws-r-sdk.com/docs/omics_get_share/} for full documentation.
}
\keyword{internal}
