% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_references}
\alias{omics_list_references}
\title{Retrieves the metadata of one or more reference genomes in a reference
store}
\usage{
omics_list_references(
  referenceStoreId,
  maxResults = NULL,
  nextToken = NULL,
  filter = NULL
)
}
\arguments{
\item{referenceStoreId}{[required] The references' reference store ID.}

\item{maxResults}{The maximum number of references to return in one page of results.}

\item{nextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{filter}{A filter to apply to the list.}
}
\description{
Retrieves the metadata of one or more reference genomes in a reference store.

See \url{https://www.paws-r-sdk.com/docs/omics_list_references/} for full documentation.
}
\keyword{internal}
