% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_shares}
\alias{omics_list_shares}
\title{Retrieves the resource shares associated with an account}
\usage{
omics_list_shares(
  resourceOwner,
  filter = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{resourceOwner}{[required] The account that owns the resource shares.}

\item{filter}{Attributes that you use to filter for a specific subset of resource
shares.}

\item{nextToken}{Next token returned in the response of a previous
ListReadSetUploadPartsRequest call. Used to get the next page of
results.}

\item{maxResults}{The maximum number of shares to return in one page of results.}
}
\description{
Retrieves the resource shares associated with an account. Use the filter parameter to retrieve a specific subset of the shares.

See \url{https://www.paws-r-sdk.com/docs/omics_list_shares/} for full documentation.
}
\keyword{internal}
