% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_workflow_versions}
\alias{omics_list_workflow_versions}
\title{Lists the workflow versions for the specified workflow}
\usage{
omics_list_workflow_versions(
  workflowId,
  type = NULL,
  workflowOwnerId = NULL,
  startingToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{workflowId}{[required] The workflow's ID. The \code{workflowId} is not the UUID.}

\item{type}{The workflow type.}

\item{workflowOwnerId}{The 12-digit account ID of the workflow owner. The workflow owner ID can
be retrieved using the \code{\link[=omics_get_share]{get_share}} API operation. If
you are the workflow owner, you do not need to include this ID.}

\item{startingToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{maxResults}{The maximum number of workflows to return in one page of results.}
}
\description{
Lists the workflow versions for the specified workflow. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/workflow-versions.html}{Workflow versioning in Amazon Web Services HealthOmics} in the \emph{Amazon Web Services HealthOmics User Guide}.

See \url{https://www.paws-r-sdk.com/docs/omics_list_workflow_versions/} for full documentation.
}
\keyword{internal}
