% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_list_workflows}
\alias{omics_list_workflows}
\title{Retrieves a list of existing workflows}
\usage{
omics_list_workflows(
  type = NULL,
  name = NULL,
  startingToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{type}{Filter the list by workflow type.}

\item{name}{Filter the list by workflow name.}

\item{startingToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{maxResults}{The maximum number of workflows to return in one page of results.}
}
\description{
Retrieves a list of existing workflows. You can filter for specific workflows by their name and type. Using the type parameter, specify \code{PRIVATE} to retrieve a list of private workflows or specify \code{READY2RUN} for a list of all Ready2Run workflows. If you do not specify the type of workflow, this operation returns a list of existing workflows.

See \url{https://www.paws-r-sdk.com/docs/omics_list_workflows/} for full documentation.
}
\keyword{internal}
