% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_start_annotation_import_job}
\alias{omics_start_annotation_import_job}
\title{Amazon Web Services HealthOmics variant stores and annotation stores
will no longer be open to new customers starting November 7, 2025}
\usage{
omics_start_annotation_import_job(
  destinationName,
  roleArn,
  items,
  versionName = NULL,
  formatOptions = NULL,
  runLeftNormalization = NULL,
  annotationFields = NULL
)
}
\arguments{
\item{destinationName}{[required] A destination annotation store for the job.}

\item{roleArn}{[required] A service role for the job.}

\item{items}{[required] Items to import.}

\item{versionName}{The name of the annotation store version.}

\item{formatOptions}{Formatting options for the annotation file.}

\item{runLeftNormalization}{The job's left normalization setting.}

\item{annotationFields}{The annotation schema generated by the parsed annotation data.}
}
\description{
Amazon Web Services HealthOmics variant stores and annotation stores will no longer be open to new customers starting November 7, 2025. If you would like to use variant stores or annotation stores, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/variant-store-availability-change.html}{Amazon Web Services HealthOmics variant store and annotation store availability change}.

See \url{https://www.paws-r-sdk.com/docs/omics_start_annotation_import_job/} for full documentation.
}
\keyword{internal}
