% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_start_read_set_import_job}
\alias{omics_start_read_set_import_job}
\title{Imports a read set from the sequence store}
\usage{
omics_start_read_set_import_job(
  sequenceStoreId,
  roleArn,
  clientToken = NULL,
  sources
)
}
\arguments{
\item{sequenceStoreId}{[required] The read set's sequence store ID.}

\item{roleArn}{[required] A service role for the job.}

\item{clientToken}{To ensure that jobs don't run multiple times, specify a unique token for
each job.}

\item{sources}{[required] The job's source files.}
}
\description{
Imports a read set from the sequence store. Read set import jobs support a maximum of 100 read sets of different types. Monitor the progress of your read set import job by calling the \code{\link[=omics_get_read_set_import_job]{get_read_set_import_job}} API operation.

See \url{https://www.paws-r-sdk.com/docs/omics_start_read_set_import_job/} for full documentation.
}
\keyword{internal}
