% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_start_reference_import_job}
\alias{omics_start_reference_import_job}
\title{Imports a reference genome from Amazon S3 into a specified reference
store}
\usage{
omics_start_reference_import_job(
  referenceStoreId,
  roleArn,
  clientToken = NULL,
  sources
)
}
\arguments{
\item{referenceStoreId}{[required] The job's reference store ID.}

\item{roleArn}{[required] A service role for the job.}

\item{clientToken}{To ensure that jobs don't run multiple times, specify a unique token for
each job.}

\item{sources}{[required] The job's source files.}
}
\description{
Imports a reference genome from Amazon S3 into a specified reference store. You can have multiple reference genomes in a reference store. You can only import reference genomes one at a time into each reference store. Monitor the status of your reference import job by using the \code{\link[=omics_get_reference_import_job]{get_reference_import_job}} API operation.

See \url{https://www.paws-r-sdk.com/docs/omics_start_reference_import_job/} for full documentation.
}
\keyword{internal}
