% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_start_run}
\alias{omics_start_run}
\title{Starts a new run or duplicates an existing run}
\usage{
omics_start_run(
  workflowId = NULL,
  workflowType = NULL,
  runId = NULL,
  roleArn,
  name = NULL,
  cacheId = NULL,
  cacheBehavior = NULL,
  runGroupId = NULL,
  priority = NULL,
  parameters = NULL,
  storageCapacity = NULL,
  outputUri = NULL,
  logLevel = NULL,
  tags = NULL,
  requestId,
  retentionMode = NULL,
  storageType = NULL,
  workflowOwnerId = NULL,
  workflowVersionName = NULL
)
}
\arguments{
\item{workflowId}{The run's workflow ID.}

\item{workflowType}{The run's workflow type.}

\item{runId}{The ID of a run to duplicate.}

\item{roleArn}{[required] A service role for the run.}

\item{name}{A name for the run.}

\item{cacheId}{Identifier of the cache associated with this run. If you don't specify a
cache ID, no task outputs are cached for this run.}

\item{cacheBehavior}{The cache behavior for the run. You specify this value if you want to
override the default behavior for the cache. You had set the default
value when you created the cache. For more information, see \href{https://docs.aws.amazon.com/omics/latest/dev/how-run-cache.html#run-cache-behavior}{Run cache behavior}
in the Amazon Web Services HealthOmics User Guide.}

\item{runGroupId}{The run's group ID.}

\item{priority}{A priority for the run.}

\item{parameters}{Parameters for the run.}

\item{storageCapacity}{The static storage capacity (in gibibytes) for this run. This field is
not required if the storage type is dynamic (the system ignores any
value that you enter).}

\item{outputUri}{An output URI for the run.}

\item{logLevel}{A log level for the run.}

\item{tags}{Tags for the run.}

\item{requestId}{[required] To ensure that requests don't run multiple times, specify a unique ID
for each request.}

\item{retentionMode}{The retention mode for the run. The default value is RETAIN.

Amazon Web Services HealthOmics stores a fixed number of runs that are
available to the console and API. In the default mode (RETAIN), you need
to remove runs manually when the number of run exceeds the maximum. If
you set the retention mode to \code{REMOVE}, Amazon Web Services HealthOmics
automatically removes runs (that have mode set to REMOVE) when the
number of run exceeds the maximum. All run logs are available in
CloudWatch logs, if you need information about a run that is no longer
available to the API.

For more information about retention mode, see \href{https://docs.aws.amazon.com/omics/latest/dev/starting-a-run.html}{Specifying run retention mode}
in the \emph{Amazon Web Services HealthOmics User Guide}.}

\item{storageType}{The storage type for the run. By default, the run uses STATIC storage
type, which allocates a fixed amount of storage. If you set the storage
type to DYNAMIC, Amazon Web Services HealthOmics dynamically scales the
storage up or down, based on file system utilization. For more
information about static and dynamic storage, see \href{https://docs.aws.amazon.com/omics/latest/dev/}{Running workflows} in the \emph{Amazon
Web Services HealthOmics User Guide}.}

\item{workflowOwnerId}{The ID of the workflow owner.}

\item{workflowVersionName}{The name of the workflow version.}
}
\description{
Starts a new run or duplicates an existing run.

See \url{https://www.paws-r-sdk.com/docs/omics_start_run/} for full documentation.
}
\keyword{internal}
