% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_update_run_group}
\alias{omics_update_run_group}
\title{Updates the settings of a run group and returns a response with no body
if the operation is successful}
\usage{
omics_update_run_group(
  id,
  name = NULL,
  maxCpus = NULL,
  maxRuns = NULL,
  maxDuration = NULL,
  maxGpus = NULL
)
}
\arguments{
\item{id}{[required] The group's ID.}

\item{name}{A name for the group.}

\item{maxCpus}{The maximum number of CPUs to use.}

\item{maxRuns}{The maximum number of concurrent runs for the group.}

\item{maxDuration}{A maximum run time for the group in minutes.}

\item{maxGpus}{The maximum GPUs that can be used by a run group.}
}
\description{
Updates the settings of a run group and returns a response with no body if the operation is successful.

See \url{https://www.paws-r-sdk.com/docs/omics_update_run_group/} for full documentation.
}
\keyword{internal}
