% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_operations.R
\name{omics_upload_read_set_part}
\alias{omics_upload_read_set_part}
\title{Uploads a specific part of a read set into a sequence store}
\usage{
omics_upload_read_set_part(
  sequenceStoreId,
  uploadId,
  partSource,
  partNumber,
  payload
)
}
\arguments{
\item{sequenceStoreId}{[required] The Sequence Store ID used for the multipart upload.}

\item{uploadId}{[required] The ID for the initiated multipart upload.}

\item{partSource}{[required] The source file for an upload part.}

\item{partNumber}{[required] The number of the part being uploaded.}

\item{payload}{[required] The read set data to upload for a part.}
}
\description{
Uploads a specific part of a read set into a sequence store. When you a upload a read set part with a part number that already exists, the new part replaces the existing one. This operation returns a JSON formatted response containing a string identifier that is used to confirm that parts are being added to the intended upload.

See \url{https://www.paws-r-sdk.com/docs/omics_upload_read_set_part/} for full documentation.
}
\keyword{internal}
