% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_get_rule}
\alias{recyclebin_get_rule}
\title{Gets information about a Recycle Bin retention rule}
\usage{
recyclebin_get_rule(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique ID of the retention rule.}
}
\description{
Gets information about a Recycle Bin retention rule.

See \url{https://www.paws-r-sdk.com/docs/recyclebin_get_rule/} for full documentation.
}
\keyword{internal}
