% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_list_rules}
\alias{recyclebin_list_rules}
\title{Lists the Recycle Bin retention rules in the Region}
\usage{
recyclebin_list_rules(
  MaxResults = NULL,
  NextToken = NULL,
  ResourceType,
  ResourceTags = NULL,
  LockState = NULL,
  ExcludeResourceTags = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{ResourceType}{[required] The resource type retained by the retention rule. Only retention rules
that retain the specified resource type are listed. Currently, only EBS
volumes, EBS snapshots, and EBS-backed AMIs are supported.
\itemize{
\item To list retention rules that retain EBS volumes, specify
\code{EBS_VOLUME}.
\item To list retention rules that retain EBS snapshots, specify
\code{EBS_SNAPSHOT}.
\item To list retention rules that retain EBS-backed AMIs, specify
\code{EC2_IMAGE}.
}}

\item{ResourceTags}{[Tag-level retention rules only] Information about the resource tags
used to identify resources that are retained by the retention rule.}

\item{LockState}{The lock state of the retention rules to list. Only retention rules with
the specified lock state are returned.}

\item{ExcludeResourceTags}{[Region-level retention rules only] Information about the exclusion
tags used to identify resources that are to be excluded, or ignored, by
the retention rule.}
}
\description{
Lists the Recycle Bin retention rules in the Region.

See \url{https://www.paws-r-sdk.com/docs/recyclebin_list_rules/} for full documentation.
}
\keyword{internal}
