% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recyclebin_operations.R
\name{recyclebin_lock_rule}
\alias{recyclebin_lock_rule}
\title{Locks a Region-level retention rule}
\usage{
recyclebin_lock_rule(Identifier, LockConfiguration)
}
\arguments{
\item{Identifier}{[required] The unique ID of the retention rule.}

\item{LockConfiguration}{[required] Information about the retention rule lock configuration.}
}
\description{
Locks a Region-level retention rule. A locked retention rule can't be modified or deleted.

See \url{https://www.paws-r-sdk.com/docs/recyclebin_lock_rule/} for full documentation.
}
\keyword{internal}
