% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_create_bucket_metadata_table_configuration}
\alias{s3_create_bucket_metadata_table_configuration}
\title{We recommend that you create your S3 Metadata configurations by using
the V2 CreateBucketMetadataConfiguration API operation}
\usage{
s3_create_bucket_metadata_table_configuration(
  Bucket,
  ContentMD5 = NULL,
  ChecksumAlgorithm = NULL,
  MetadataTableConfiguration,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The general purpose bucket that you want to create the metadata table
configuration for.}

\item{ContentMD5}{The \code{Content-MD5} header for the metadata table configuration.}

\item{ChecksumAlgorithm}{The checksum algorithm to use with your metadata table configuration.}

\item{MetadataTableConfiguration}{[required] The contents of your metadata table configuration.}

\item{ExpectedBucketOwner}{The expected owner of the general purpose bucket that corresponds to
your metadata table configuration.}
}
\description{
We recommend that you create your S3 Metadata configurations by using the V2 \code{\link[=s3_create_bucket_metadata_configuration]{create_bucket_metadata_configuration}} API operation. We no longer recommend using the V1 \code{\link[=s3_create_bucket_metadata_table_configuration]{create_bucket_metadata_table_configuration}} API operation.

See \url{https://www.paws-r-sdk.com/docs/s3_create_bucket_metadata_table_configuration/} for full documentation.
}
\keyword{internal}
