% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_analytics_configuration}
\alias{s3_delete_bucket_analytics_configuration}
\title{This operation is not supported for directory buckets}
\usage{
s3_delete_bucket_analytics_configuration(
  Bucket,
  Id,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket from which an analytics configuration is deleted.}

\item{Id}{[required] The ID that identifies the analytics configuration.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide does not match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}
}
\description{
This operation is not supported for directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_delete_bucket_analytics_configuration/} for full documentation.
}
\keyword{internal}
