% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_cors}
\alias{s3_delete_bucket_cors}
\title{Deletes the cors configuration information set for the bucket}
\usage{
s3_delete_bucket_cors(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] Specifies the bucket whose \code{cors} configuration is being deleted.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Deletes the \code{cors} configuration information set for the bucket.

See \url{https://paws-r.github.io/docs/s3/delete_bucket_cors.html} for full documentation.
}
\keyword{internal}
