% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_inventory_configuration}
\alias{s3_delete_bucket_inventory_configuration}
\title{Deletes an inventory configuration (identified by the inventory ID) from
the bucket}
\usage{
s3_delete_bucket_inventory_configuration(Bucket, Id,
  ExpectedBucketOwner)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the inventory configuration to delete.}

\item{Id}{[required] The ID used to identify the inventory configuration.}

\item{ExpectedBucketOwner}{The account id of the expected bucket owner. If the bucket is owned by a
different account, the request will fail with an HTTP
\verb{403 (Access Denied)} error.}
}
\description{
Deletes an inventory configuration (identified by the inventory ID) from
the bucket.

To use this operation, you must have permissions to perform the
\code{s3:PutInventoryConfiguration} action. The bucket owner has this
permission by default. The bucket owner can grant this permission to
others. For more information about permissions, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/using-with-s3-actions.html#using-with-s3-actions-related-to-bucket-subresources}{Permissions Related to Bucket Subresource Operations}
and \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-access-control.html}{Managing Access Permissions to Your Amazon S3 Resources}.

For information about the Amazon S3 inventory feature, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage-inventory.html}{Amazon S3 Inventory}.

Operations related to
\code{\link[=s3_delete_bucket_inventory_configuration]{delete_bucket_inventory_configuration}}
include:
\itemize{
\item \code{\link[=s3_get_bucket_inventory_configuration]{get_bucket_inventory_configuration}}
\item \code{\link[=s3_put_bucket_inventory_configuration]{put_bucket_inventory_configuration}}
\item \code{\link[=s3_list_bucket_inventory_configurations]{list_bucket_inventory_configurations}}
}
}
\section{Request syntax}{
\preformatted{svc$delete_bucket_inventory_configuration(
  Bucket = "string",
  Id = "string",
  ExpectedBucketOwner = "string"
)
}
}

\keyword{internal}
