% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_lifecycle}
\alias{s3_delete_bucket_lifecycle}
\title{Deletes the lifecycle configuration from the specified bucket}
\usage{
s3_delete_bucket_lifecycle(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The bucket name of the lifecycle to delete.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Deletes the lifecycle configuration from the specified bucket. Amazon S3 removes all the lifecycle configuration rules in the lifecycle subresource associated with the bucket. Your objects never expire, and Amazon S3 no longer automatically deletes any objects on the basis of rules contained in the deleted lifecycle configuration.

See \url{https://paws-r.github.io/docs/s3/delete_bucket_lifecycle.html} for full documentation.
}
\keyword{internal}
