% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_tagging}
\alias{s3_delete_bucket_tagging}
\title{Deletes the tags from the bucket}
\usage{
s3_delete_bucket_tagging(Bucket)
}
\arguments{
\item{Bucket}{[required] The bucket that has the tag set to be removed.}
}
\description{
Deletes the tags from the bucket.
}
\details{
To use this operation, you must have permission to perform the
\code{s3:PutBucketTagging} action. By default, the bucket owner has this
permission and can grant this permission to others.

The following operations are related to \code{DeleteBucketTagging}:
\itemize{
\item GetBucketTagging
\item PutBucketTagging
}
}
\section{Request syntax}{
\preformatted{svc$delete_bucket_tagging(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes bucket tags.
\donttest{svc$delete_bucket_tagging(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
