% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_analytics_configuration}
\alias{s3_get_bucket_analytics_configuration}
\title{Gets an analytics configuration for the bucket (specified by the
analytics configuration ID)}
\usage{
s3_get_bucket_analytics_configuration(Bucket, Id)
}
\arguments{
\item{Bucket}{[required] The name of the bucket from which an analytics configuration is
retrieved.}

\item{Id}{[required] The identifier used to represent an analytics configuration.}
}
\description{
Gets an analytics configuration for the bucket (specified by the
analytics configuration ID).
}
\section{Request syntax}{
\preformatted{svc$get_bucket_analytics_configuration(
  Bucket = "string",
  Id = "string"
)
}
}

\keyword{internal}
