% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_cors}
\alias{s3_get_bucket_cors}
\title{Returns the Cross-Origin Resource Sharing (CORS) configuration
information set for the bucket}
\usage{
s3_get_bucket_cors(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The bucket name for which to get the cors configuration.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Returns the Cross-Origin Resource Sharing (CORS) configuration information set for the bucket.

See \url{https://paws-r.github.io/docs/s3/get_bucket_cors.html} for full documentation.
}
\keyword{internal}
