% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_location}
\alias{s3_get_bucket_location}
\title{Returns the Region the bucket resides in}
\usage{
s3_get_bucket_location(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The name of the bucket for which to get the location.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Returns the Region the bucket resides in. You set the bucket's Region using the \code{LocationConstraint} request parameter in a \code{\link[=s3_create_bucket]{create_bucket}} request. For more information, see \code{\link[=s3_create_bucket]{create_bucket}}.

See \url{https://paws-r.github.io/docs/s3/get_bucket_location.html} for full documentation.
}
\keyword{internal}
