% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_logging}
\alias{s3_get_bucket_logging}
\title{Returns the logging status of a bucket and the permissions users have to
view and modify that status}
\usage{
s3_get_bucket_logging(Bucket, ExpectedBucketOwner = NULL)
}
\arguments{
\item{Bucket}{[required] The bucket name for which to get the logging information.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Returns the logging status of a bucket and the permissions users have to view and modify that status.

See \url{https://paws-r.github.io/docs/s3/get_bucket_logging.html} for full documentation.
}
\keyword{internal}
