% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_object_torrent}
\alias{s3_get_object_torrent}
\title{Return torrent files from a bucket}
\usage{
s3_get_object_torrent(Bucket, Key, RequestPayer)
}
\arguments{
\item{Bucket}{[required]}

\item{Key}{[required]}

\item{RequestPayer}{}
}
\description{
Return torrent files from a bucket.
}
\section{Request syntax}{
\preformatted{svc$get_object_torrent(
  Bucket = "string",
  Key = "string",
  RequestPayer = "requester"
)
}
}

\examples{
# The following example retrieves torrent files of an object.
\donttest{svc$get_object_torrent(
  Bucket = "examplebucket",
  Key = "HappyFace.jpg"
)}

}
\keyword{internal}
