% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_head_bucket}
\alias{s3_head_bucket}
\title{This action is useful to determine if a bucket exists and you have
permission to access it}
\usage{
s3_head_bucket(Bucket, ExpectedBucketOwner)
}
\arguments{
\item{Bucket}{[required] The bucket name.

When using this action with an access point, you must direct requests to
the access point hostname. The access point hostname takes the form
\emph{AccessPointName}-\emph{AccountId}.s3-accesspoint.\emph{Region}.amazonaws.com.
When using this action with an access point through the AWS SDKs, you
provide the access point ARN in place of the bucket name. For more
information about access point ARNs, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/using-access-points.html}{Using Access Points}
in the \emph{Amazon S3 User Guide}.

When using this action with Amazon S3 on Outposts, you must direct
requests to the S3 on Outposts hostname. The S3 on Outposts hostname
takes the form
\emph{AccessPointName}-\emph{AccountId}.\emph{outpostID}.s3-outposts.\emph{Region}.amazonaws.com.
When using this action using S3 on Outposts through the AWS SDKs, you
provide the Outposts bucket ARN in place of the bucket name. For more
information about S3 on Outposts ARNs, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/S3onOutposts.html}{Using S3 on Outposts}
in the \emph{Amazon S3 User Guide}.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request will fail with an HTTP
\verb{403 (Access Denied)} error.}
}
\value{
An empty list.
}
\description{
This action is useful to determine if a bucket exists and you have
permission to access it. The action returns a \verb{200 OK} if the bucket
exists and you have permission to access it.

If the bucket does not exist or you do not have permission to access it,
the \code{HEAD} request returns a generic \verb{404 Not Found} or \verb{403 Forbidden}
code. A message body is not included, so you cannot determine the
exception beyond these error codes.

To use this operation, you must have permissions to perform the
\code{s3:ListBucket} action. The bucket owner has this permission by default
and can grant this permission to others. For more information about
permissions, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/using-with-s3-actions.html#using-with-s3-actions-related-to-bucket-subresources}{Permissions Related to Bucket Subresource Operations}
and \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-access-control.html}{Managing Access Permissions to Your Amazon S3 Resources}.
}
\section{Request syntax}{
\preformatted{svc$head_bucket(
  Bucket = "string",
  ExpectedBucketOwner = "string"
)
}
}

\examples{
\dontrun{
# This operation checks to see if a bucket exists.
svc$head_bucket(
  Bucket = "acl1"
)
}

}
\keyword{internal}
