% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_list_bucket_analytics_configurations}
\alias{s3_list_bucket_analytics_configurations}
\title{Lists the analytics configurations for the bucket}
\usage{
s3_list_bucket_analytics_configurations(
  Bucket,
  ContinuationToken = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket from which analytics configurations are
retrieved.}

\item{ContinuationToken}{The ContinuationToken that represents a placeholder from where this
request should begin.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Lists the analytics configurations for the bucket. You can have up to 1,000 analytics configurations per bucket.

See \url{https://paws-r.github.io/docs/s3/list_bucket_analytics_configurations.html} for full documentation.
}
\keyword{internal}
