% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_list_bucket_intelligent_tiering_configurations}
\alias{s3_list_bucket_intelligent_tiering_configurations}
\title{This operation is not supported for directory buckets}
\usage{
s3_list_bucket_intelligent_tiering_configurations(
  Bucket,
  ContinuationToken = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose configuration you want to modify
or retrieve.}

\item{ContinuationToken}{The \code{ContinuationToken} that represents a placeholder from where this
request should begin.}
}
\description{
This operation is not supported for directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_list_bucket_intelligent_tiering_configurations/} for full documentation.
}
\keyword{internal}
