% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_list_bucket_metrics_configurations}
\alias{s3_list_bucket_metrics_configurations}
\title{This operation is not supported by directory buckets}
\usage{
s3_list_bucket_metrics_configurations(
  Bucket,
  ContinuationToken = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the metrics configurations to
retrieve.}

\item{ContinuationToken}{The marker that is used to continue a metrics configuration listing that
has been truncated. Use the \code{NextContinuationToken} from a previously
truncated list response to continue the listing. The continuation token
is an opaque value that Amazon S3 understands.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide does not match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_list_bucket_metrics_configurations/} for full documentation.
}
\keyword{internal}
