% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_list_buckets}
\alias{s3_list_buckets}
\title{Returns a list of all buckets owned by the authenticated sender of the
request}
\usage{
s3_list_buckets()
}
\description{
Returns a list of all buckets owned by the authenticated sender of the request. To use this operation, you must have the \code{s3:ListAllMyBuckets} permission.

See \url{https://paws-r.github.io/docs/s3/list_buckets.html} for full documentation.
}
\keyword{internal}
