% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_list_buckets}
\alias{s3_list_buckets}
\title{This operation is not supported by directory buckets}
\usage{
s3_list_buckets()
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_list_buckets/} for full documentation.
}
\keyword{internal}
