% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_intelligent_tiering_configuration}
\alias{s3_put_bucket_intelligent_tiering_configuration}
\title{This operation is not supported for directory buckets}
\usage{
s3_put_bucket_intelligent_tiering_configuration(
  Bucket,
  Id,
  ExpectedBucketOwner = NULL,
  IntelligentTieringConfiguration
)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose configuration you want to modify
or retrieve.}

\item{Id}{[required] The ID used to identify the S3 Intelligent-Tiering configuration.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide does not match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}

\item{IntelligentTieringConfiguration}{[required] Container for S3 Intelligent-Tiering configuration.}
}
\description{
This operation is not supported for directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_put_bucket_intelligent_tiering_configuration/} for full documentation.
}
\keyword{internal}
