% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_metrics_configuration}
\alias{s3_put_bucket_metrics_configuration}
\title{This operation is not supported for directory buckets}
\usage{
s3_put_bucket_metrics_configuration(
  Bucket,
  Id,
  MetricsConfiguration,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket for which the metrics configuration is set.}

\item{Id}{[required] The ID used to identify the metrics configuration. The ID has a 64
character limit and can only contain letters, numbers, periods, dashes,
and underscores.}

\item{MetricsConfiguration}{[required] Specifies the metrics configuration.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide does not match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}
}
\description{
This operation is not supported for directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_put_bucket_metrics_configuration/} for full documentation.
}
\keyword{internal}
