% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_notification}
\alias{s3_put_bucket_notification}
\title{No longer used, see the PutBucketNotificationConfiguration operation}
\usage{
s3_put_bucket_notification(
  Bucket,
  ContentMD5 = NULL,
  ChecksumAlgorithm = NULL,
  NotificationConfiguration,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket.}

\item{ContentMD5}{The MD5 hash of the
\code{\link[=s3_put_public_access_block]{put_public_access_block}} request body.

For requests made using the Amazon Web Services Command Line Interface
(CLI) or Amazon Web Services SDKs, this field is calculated
automatically.}

\item{ChecksumAlgorithm}{Indicates the algorithm used to create the checksum for the object when
using the SDK. This header will not provide any additional functionality
if not using the SDK. When sending this header, there must be a
corresponding \code{x-amz-checksum} or \code{x-amz-trailer} header sent.
Otherwise, Amazon S3 fails the request with the HTTP status code
\verb{400 Bad Request}. For more information, see Checking object integrity
in the \emph{Amazon S3 User Guide}.

If you provide an individual checksum, Amazon S3 ignores any provided
\code{ChecksumAlgorithm} parameter.}

\item{NotificationConfiguration}{[required] The container for the configuration.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
No longer used, see the \code{\link[=s3_put_bucket_notification_configuration]{put_bucket_notification_configuration}} operation.

See \url{https://paws-r.github.io/docs/s3/put_bucket_notification.html} for full documentation.
}
\keyword{internal}
