% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_notification_configuration}
\alias{s3_put_bucket_notification_configuration}
\title{This operation is not supported for directory buckets}
\usage{
s3_put_bucket_notification_configuration(
  Bucket,
  NotificationConfiguration,
  ExpectedBucketOwner = NULL,
  SkipDestinationValidation = NULL
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket.}

\item{NotificationConfiguration}{[required]}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide does not match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}

\item{SkipDestinationValidation}{Skips validation of Amazon SQS, Amazon SNS, and Lambda destinations.
True or false value.}
}
\description{
This operation is not supported for directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_put_bucket_notification_configuration/} for full documentation.
}
\keyword{internal}
