% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_ownership_controls}
\alias{s3_put_bucket_ownership_controls}
\title{This operation is not supported for directory buckets}
\usage{
s3_put_bucket_ownership_controls(
  Bucket,
  ContentMD5 = NULL,
  ExpectedBucketOwner = NULL,
  OwnershipControls
)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose \code{OwnershipControls} you want to
set.}

\item{ContentMD5}{The MD5 hash of the \code{OwnershipControls} request body.

For requests made using the Amazon Web Services Command Line Interface
(CLI) or Amazon Web Services SDKs, this field is calculated
automatically.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the account ID that you
provide does not match the actual owner of the bucket, the request fails
with the HTTP status code \verb{403 Forbidden} (access denied).}

\item{OwnershipControls}{[required] The \code{OwnershipControls} (BucketOwnerEnforced, BucketOwnerPreferred, or
ObjectWriter) that you want to apply to this Amazon S3 bucket.}
}
\description{
This operation is not supported for directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3_put_bucket_ownership_controls/} for full documentation.
}
\keyword{internal}
