% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_object_lock_configuration}
\alias{s3_put_object_lock_configuration}
\title{Places an Object Lock configuration on the specified bucket}
\usage{
s3_put_object_lock_configuration(Bucket, ObjectLockConfiguration,
  RequestPayer, Token, ContentMD5, ExpectedBucketOwner)
}
\arguments{
\item{Bucket}{[required] The bucket whose Object Lock configuration you want to create or
replace.}

\item{ObjectLockConfiguration}{The Object Lock configuration that you want to apply to the specified
bucket.}

\item{RequestPayer}{}

\item{Token}{A token to allow Object Lock to be enabled for an existing bucket.}

\item{ContentMD5}{The MD5 hash for the request body.

For requests made using the AWS Command Line Interface (CLI) or AWS
SDKs, this field is calculated automatically.}

\item{ExpectedBucketOwner}{The account id of the expected bucket owner. If the bucket is owned by a
different account, the request will fail with an HTTP
\verb{403 (Access Denied)} error.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RequestCharged = "requester"
)
}\if{html}{\out{</div>}}
}
\description{
Places an Object Lock configuration on the specified bucket. The rule
specified in the Object Lock configuration will be applied by default to
every new object placed in the specified bucket.

\code{DefaultRetention} requires either Days or Years. You can't specify both
at the same time.

\strong{Related Resources}
\itemize{
\item \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-lock.html}{Locking Objects}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_object_lock_configuration(
  Bucket = "string",
  ObjectLockConfiguration = list(
    ObjectLockEnabled = "Enabled",
    Rule = list(
      DefaultRetention = list(
        Mode = "GOVERNANCE"|"COMPLIANCE",
        Days = 123,
        Years = 123
      )
    )
  ),
  RequestPayer = "requester",
  Token = "string",
  ContentMD5 = "string",
  ExpectedBucketOwner = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
