% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_object_lock_configuration}
\alias{s3_put_object_lock_configuration}
\title{Places an Object Lock configuration on the specified bucket}
\usage{
s3_put_object_lock_configuration(
  Bucket,
  ObjectLockConfiguration = NULL,
  RequestPayer = NULL,
  Token = NULL,
  ContentMD5 = NULL,
  ChecksumAlgorithm = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{Bucket}{[required] The bucket whose Object Lock configuration you want to create or
replace.}

\item{ObjectLockConfiguration}{The Object Lock configuration that you want to apply to the specified
bucket.}

\item{RequestPayer}{}

\item{Token}{A token to allow Object Lock to be enabled for an existing bucket.}

\item{ContentMD5}{The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface
(CLI) or Amazon Web Services SDKs, this field is calculated
automatically.}

\item{ChecksumAlgorithm}{Indicates the algorithm used to create the checksum for the object when
using the SDK. This header will not provide any additional functionality
if not using the SDK. When sending this header, there must be a
corresponding \code{x-amz-checksum} or \code{x-amz-trailer} header sent.
Otherwise, Amazon S3 fails the request with the HTTP status code
\verb{400 Bad Request}. For more information, see Checking object integrity
in the \emph{Amazon S3 User Guide}.

If you provide an individual checksum, Amazon S3 ignores any provided
\code{ChecksumAlgorithm} parameter.}

\item{ExpectedBucketOwner}{The account ID of the expected bucket owner. If the bucket is owned by a
different account, the request fails with the HTTP status code
\verb{403 Forbidden} (access denied).}
}
\description{
Places an Object Lock configuration on the specified bucket. The rule specified in the Object Lock configuration will be applied by default to every new object placed in the specified bucket. For more information, see Locking Objects.

See \url{https://paws-r.github.io/docs/s3/put_object_lock_configuration.html} for full documentation.
}
\keyword{internal}
