% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_create_access_point_for_object_lambda}
\alias{s3control_create_access_point_for_object_lambda}
\title{Creates an Object Lambda Access Point}
\usage{
s3control_create_access_point_for_object_lambda(AccountId, Name, Configuration)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for owner of the specified Object
Lambda Access Point.}

\item{Name}{[required] The name you want to assign to this Object Lambda Access Point.}

\item{Configuration}{[required] Object Lambda Access Point configuration as a JSON document.}
}
\description{
Creates an Object Lambda Access Point. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/transforming-objects.html}{Transforming objects with Object Lambda Access Points} in the \emph{Amazon S3 User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3control_create_access_point_for_object_lambda/} for full documentation.
}
\keyword{internal}
