% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_delete_storage_lens_configuration}
\alias{s3control_delete_storage_lens_configuration}
\title{Deletes the Amazon S3 Storage Lens configuration}
\usage{
s3control_delete_storage_lens_configuration(ConfigId, AccountId)
}
\arguments{
\item{ConfigId}{[required] The ID of the S3 Storage Lens configuration.}

\item{AccountId}{[required] The account ID of the requester.}
}
\description{
Deletes the Amazon S3 Storage Lens configuration. For more information about S3 Storage Lens, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage_lens.html}{Assessing your storage activity and usage with Amazon S3 Storage Lens} in the \emph{Amazon S3 User Guide}.

See \url{https://paws-r.github.io/docs/s3control/delete_storage_lens_configuration.html} for full documentation.
}
\keyword{internal}
