% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_describe_job}
\alias{s3control_describe_job}
\title{Retrieves the configuration parameters and status for a Batch Operations
job}
\usage{
s3control_describe_job(AccountId, JobId)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID associated with the S3 Batch
Operations job.}

\item{JobId}{[required] The ID for the job whose information you want to retrieve.}
}
\description{
Retrieves the configuration parameters and status for a Batch Operations job. For more information, see S3 Batch Operations in the \emph{Amazon S3 User Guide}.

See \url{https://paws-r.github.io/docs/s3control/describe_job.html} for full documentation.
}
\keyword{internal}
