% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_grant}
\alias{s3control_get_access_grant}
\title{Get the details of an access grant from your S3 Access Grants instance}
\usage{
s3control_get_access_grant(AccountId, AccessGrantId)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID of the S3 Access Grants instance.}

\item{AccessGrantId}{[required] The ID of the access grant. S3 Access Grants auto-generates this ID when
you create the access grant.}
}
\description{
Get the details of an access grant from your S3 Access Grants instance.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_access_grant/} for full documentation.
}
\keyword{internal}
