% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_point_policy}
\alias{s3control_get_access_point_policy}
\title{Returns the access point policy associated with the specified access
point}
\usage{
s3control_get_access_point_policy(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified access point.}

\item{Name}{[required] The name of the access point whose policy you want to retrieve.

For using this parameter with Amazon S3 on Outposts with the REST API,
you must specify the name and the x-amz-outpost-id as well.

For using this parameter with S3 on Outposts with the AWS SDK and CLI,
you must specify the ARN of the access point accessed in the format
\verb{arn:aws:s3-outposts:<Region>:<account-id>:outpost/<outpost-id>/accesspoint/<my-accesspoint-name>}.
For example, to access the access point \code{reports-ap} through outpost
\code{my-outpost} owned by account \code{123456789012} in Region \code{us-west-2}, use
the URL encoding of
\code{arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/accesspoint/reports-ap}.
The value must be URL encoded.}
}
\description{
Returns the access point policy associated with the specified access
point.

The following actions are related to
\code{\link[=s3control_get_access_point_policy]{get_access_point_policy}}:
\itemize{
\item \code{\link[=s3control_put_access_point_policy]{put_access_point_policy}}
\item \code{\link[=s3control_delete_access_point_policy]{delete_access_point_policy}}
}
}
\section{Request syntax}{
\preformatted{svc$get_access_point_policy(
  AccountId = "string",
  Name = "string"
)
}
}

\keyword{internal}
