% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_point_policy_for_object_lambda}
\alias{s3control_get_access_point_policy_for_object_lambda}
\title{Returns the resource policy for an Object Lambda Access Point}
\usage{
s3control_get_access_point_policy_for_object_lambda(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{Name}{[required] The name of the Object Lambda Access Point.}
}
\value{
A list with the following syntax:\preformatted{list(
  Policy = "string"
)
}
}
\description{
Returns the resource policy for an Object Lambda Access Point.

The following actions are related to
\code{\link[=s3control_get_access_point_policy_for_object_lambda]{get_access_point_policy_for_object_lambda}}:
\itemize{
\item \code{\link[=s3control_delete_access_point_policy_for_object_lambda]{delete_access_point_policy_for_object_lambda}}
\item \code{\link[=s3control_put_access_point_policy_for_object_lambda]{put_access_point_policy_for_object_lambda}}
}
}
\section{Request syntax}{
\preformatted{svc$get_access_point_policy_for_object_lambda(
  AccountId = "string",
  Name = "string"
)
}
}

\keyword{internal}
