% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_access_point_policy_status}
\alias{s3control_get_access_point_policy_status}
\title{This operation is not supported by directory buckets}
\usage{
s3control_get_access_point_policy_status(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified access point.}

\item{Name}{[required] The name of the access point whose policy status you want to retrieve.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_access_point_policy_status/} for full documentation.
}
\keyword{internal}
