% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_job_tagging}
\alias{s3control_get_job_tagging}
\title{Returns the tags on an S3 Batch Operations job}
\usage{
s3control_get_job_tagging(AccountId, JobId)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID associated with the S3 Batch
Operations job.}

\item{JobId}{[required] The ID for the S3 Batch Operations job whose tags you want to retrieve.}
}
\description{
Returns the tags on an S3 Batch Operations job. To use this operation, you must have permission to perform the \code{s3:GetJobTagging} action. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/batch-ops-managing-jobs.html#batch-ops-job-tags}{Controlling access and labeling jobs using tags} in the \emph{Amazon S3 User Guide}.

See \url{https://paws-r.github.io/docs/s3control/get_job_tagging.html} for full documentation.
}
\keyword{internal}
