% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_multi_region_access_point_policy}
\alias{s3control_get_multi_region_access_point_policy}
\title{This operation is not supported by directory buckets}
\usage{
s3control_get_multi_region_access_point_policy(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{Name}{[required] Specifies the Multi-Region Access Point. The name of the Multi-Region
Access Point is different from the alias. For more information about the
distinction between the name and the alias of an Multi-Region Access
Point, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/CreatingMultiRegionAccessPoints.html#multi-region-access-point-naming}{Managing Multi-Region Access Points}
in the \emph{Amazon S3 User Guide}.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_multi_region_access_point_policy/} for full documentation.
}
\keyword{internal}
