% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_multi_region_access_point_routes}
\alias{s3control_get_multi_region_access_point_routes}
\title{This operation is not supported by directory buckets}
\usage{
s3control_get_multi_region_access_point_routes(AccountId, Mrap)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{Mrap}{[required] The Multi-Region Access Point ARN.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_multi_region_access_point_routes/} for full documentation.
}
\keyword{internal}
