% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_get_public_access_block}
\alias{s3control_get_public_access_block}
\title{This operation is not supported by directory buckets}
\usage{
s3control_get_public_access_block(AccountId)
}
\arguments{
\item{AccountId}{[required] The account ID for the Amazon Web Services account whose
\code{PublicAccessBlock} configuration you want to retrieve.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_get_public_access_block/} for full documentation.
}
\keyword{internal}
