% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_multi_region_access_points}
\alias{s3control_list_multi_region_access_points}
\title{This operation is not supported by directory buckets}
\usage{
s3control_list_multi_region_access_points(
  AccountId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID for the owner of the Multi-Region
Access Point.}

\item{NextToken}{Not currently used. Do not use this parameter.}

\item{MaxResults}{Not currently used. Do not use this parameter.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_list_multi_region_access_points/} for full documentation.
}
\keyword{internal}
