% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_regional_buckets}
\alias{s3control_list_regional_buckets}
\title{This operation is not supported by directory buckets}
\usage{
s3control_list_regional_buckets(
  AccountId,
  NextToken = NULL,
  MaxResults = NULL,
  OutpostId = NULL
)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID of the Outposts bucket.}

\item{NextToken}{}

\item{MaxResults}{}

\item{OutpostId}{The ID of the Outposts resource.

This ID is required by Amazon S3 on Outposts buckets.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_list_regional_buckets/} for full documentation.
}
\keyword{internal}
