% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_list_storage_lens_configurations}
\alias{s3control_list_storage_lens_configurations}
\title{Gets a list of Amazon S3 Storage Lens configurations}
\usage{
s3control_list_storage_lens_configurations(AccountId, NextToken = NULL)
}
\arguments{
\item{AccountId}{[required] The account ID of the requester.}

\item{NextToken}{A pagination token to request the next page of results.}
}
\description{
Gets a list of Amazon S3 Storage Lens configurations. For more information about S3 Storage Lens, see Assessing your storage activity and usage with Amazon S3 Storage Lens in the \emph{Amazon S3 User Guide}.

See \url{https://paws-r.github.io/docs/s3control/list_storage_lens_configurations.html} for full documentation.
}
\keyword{internal}
