% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_access_point_configuration_for_object_lambda}
\alias{s3control_put_access_point_configuration_for_object_lambda}
\title{This operation is not supported by directory buckets}
\usage{
s3control_put_access_point_configuration_for_object_lambda(
  AccountId,
  Name,
  Configuration
)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{Name}{[required] The name of the Object Lambda Access Point.}

\item{Configuration}{[required] Object Lambda Access Point configuration document.}
}
\description{
This operation is not supported by directory buckets.

See \url{https://www.paws-r-sdk.com/docs/s3control_put_access_point_configuration_for_object_lambda/} for full documentation.
}
\keyword{internal}
