% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_access_point_policy_for_object_lambda}
\alias{s3control_put_access_point_policy_for_object_lambda}
\title{Creates or replaces resource policy for an Object Lambda Access Point}
\usage{
s3control_put_access_point_policy_for_object_lambda(AccountId, Name, Policy)
}
\arguments{
\item{AccountId}{[required] The account ID for the account that owns the specified Object Lambda
Access Point.}

\item{Name}{[required] The name of the Object Lambda Access Point.}

\item{Policy}{[required] Object Lambda Access Point resource policy document.}
}
\description{
Creates or replaces resource policy for an Object Lambda Access Point. For an example policy, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/olap-create.html#olap-create-cli}{Creating Object Lambda Access Points} in the \emph{Amazon S3 User Guide}.

See \url{https://paws-r.github.io/docs/s3control/put_access_point_policy_for_object_lambda.html} for full documentation.
}
\keyword{internal}
