% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_public_access_block}
\alias{s3control_put_public_access_block}
\title{Creates or modifies the PublicAccessBlock configuration for an Amazon
Web Services account}
\usage{
s3control_put_public_access_block(PublicAccessBlockConfiguration, AccountId)
}
\arguments{
\item{PublicAccessBlockConfiguration}{[required] The \code{PublicAccessBlock} configuration that you want to apply to the
specified Amazon Web Services account.}

\item{AccountId}{[required] The account ID for the Amazon Web Services account whose
\code{PublicAccessBlock} configuration you want to set.}
}
\description{
Creates or modifies the \code{PublicAccessBlock} configuration for an Amazon Web Services account. For this operation, users must have the \code{s3:PutAccountPublicAccessBlock} permission. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/access-control-block-public-access.html}{Using Amazon S3 block public access}.

See \url{https://paws-r.github.io/docs/s3control/put_public_access_block.html} for full documentation.
}
\keyword{internal}
