% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_operations.R
\name{s3control_put_storage_lens_configuration}
\alias{s3control_put_storage_lens_configuration}
\title{Puts an Amazon S3 Storage Lens configuration}
\usage{
s3control_put_storage_lens_configuration(
  ConfigId,
  AccountId,
  StorageLensConfiguration,
  Tags = NULL
)
}
\arguments{
\item{ConfigId}{[required] The ID of the S3 Storage Lens configuration.}

\item{AccountId}{[required] The account ID of the requester.}

\item{StorageLensConfiguration}{[required] The S3 Storage Lens configuration.}

\item{Tags}{The tag set of the S3 Storage Lens configuration.

You can set up to a maximum of 50 tags.}
}
\description{
Puts an Amazon S3 Storage Lens configuration. For more information about S3 Storage Lens, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage_lens.html}{Working with Amazon S3 Storage Lens} in the \emph{Amazon S3 User Guide}. For a complete list of S3 Storage Lens metrics, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage_lens_metrics_glossary.html}{S3 Storage Lens metrics glossary} in the \emph{Amazon S3 User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3control_put_storage_lens_configuration/} for full documentation.
}
\keyword{internal}
