% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3tables_operations.R
\name{s3tables_get_table_maintenance_job_status}
\alias{s3tables_get_table_maintenance_job_status}
\title{Gets the status of a maintenance job for a table}
\usage{
s3tables_get_table_maintenance_job_status(tableBucketARN, namespace, name)
}
\arguments{
\item{tableBucketARN}{[required] The Amazon Resource Name (ARN) of the table bucket.}

\item{namespace}{[required] The name of the namespace the table is associated with.}

\item{name}{[required] The name of the table containing the maintenance job status you want to
check.}
}
\description{
Gets the status of a maintenance job for a table. For more information, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-tables-maintenance.html}{S3 Tables maintenance} in the \emph{Amazon Simple Storage Service User Guide}.

See \url{https://www.paws-r-sdk.com/docs/s3tables_get_table_maintenance_job_status/} for full documentation.
}
\keyword{internal}
