% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_associate_file_system}
\alias{storagegateway_associate_file_system}
\title{Associate an Amazon FSx file system with the Amazon FSx file gateway}
\usage{
storagegateway_associate_file_system(UserName, Password, ClientToken,
  GatewayARN, LocationARN, Tags, AuditDestinationARN, CacheAttributes)
}
\arguments{
\item{UserName}{[required] The user name of the user credential that has permission to access the
root share D$ of the Amazon FSx file system. The user account must
belong to the Amazon FSx delegated admin user group.}

\item{Password}{[required] The password of the user credential.}

\item{ClientToken}{[required] A unique string value that you supply that is used by the file gateway
to ensure idempotent file system association creation.}

\item{GatewayARN}{[required]}

\item{LocationARN}{[required] The Amazon Resource Name (ARN) of the Amazon FSx file system to
associate with the Amazon FSx file gateway.}

\item{Tags}{A list of up to 50 tags that can be assigned to the file system
association. Each tag is a key-value pair.}

\item{AuditDestinationARN}{The Amazon Resource Name (ARN) of the storage used for the audit logs.}

\item{CacheAttributes}{}
}
\value{
A list with the following syntax:\preformatted{list(
  FileSystemAssociationARN = "string"
)
}
}
\description{
Associate an Amazon FSx file system with the Amazon FSx file gateway.
After the association process is complete, the file shares on the Amazon
FSx file system are available for access through the gateway. This
operation only supports the Amazon FSx file gateway type.
}
\section{Request syntax}{
\preformatted{svc$associate_file_system(
  UserName = "string",
  Password = "string",
  ClientToken = "string",
  GatewayARN = "string",
  LocationARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  AuditDestinationARN = "string",
  CacheAttributes = list(
    CacheStaleTimeoutInSeconds = 123
  )
)
}
}

\keyword{internal}
