% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_cancel_archival}
\alias{storagegateway_cancel_archival}
\title{Cancels archiving of a virtual tape to the virtual tape shelf (VTS)
after the archiving process is initiated}
\usage{
storagegateway_cancel_archival(GatewayARN, TapeARN)
}
\arguments{
\item{GatewayARN}{[required]}

\item{TapeARN}{[required] The Amazon Resource Name (ARN) of the virtual tape you want to cancel
archiving for.}
}
\description{
Cancels archiving of a virtual tape to the virtual tape shelf (VTS) after the archiving process is initiated. This operation is only supported in the tape gateway type.

See \url{https://paws-r.github.io/docs/storagegateway/cancel_archival.html} for full documentation.
}
\keyword{internal}
