% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_cancel_retrieval}
\alias{storagegateway_cancel_retrieval}
\title{Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to
a gateway after the retrieval process is initiated}
\usage{
storagegateway_cancel_retrieval(GatewayARN, TapeARN)
}
\arguments{
\item{GatewayARN}{[required]}

\item{TapeARN}{[required] The Amazon Resource Name (ARN) of the virtual tape you want to cancel
retrieval for.}
}
\description{
Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after the retrieval process is initiated. The virtual tape is returned to the VTS. This operation is only supported in the tape gateway type.

See \url{https://paws-r.github.io/docs/storagegateway/cancel_retrieval.html} for full documentation.
}
\keyword{internal}
