% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_delete_automatic_tape_creation_policy}
\alias{storagegateway_delete_automatic_tape_creation_policy}
\title{Deletes the automatic tape creation policy of a gateway}
\usage{
storagegateway_delete_automatic_tape_creation_policy(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GatewayARN = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the automatic tape creation policy of a gateway. If you delete
this policy, new virtual tapes must be created manually. Use the Amazon
Resource Name (ARN) of the gateway in your request to remove the policy.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_automatic_tape_creation_policy(
  GatewayARN = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
