% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_delete_gateway}
\alias{storagegateway_delete_gateway}
\title{Deletes a gateway}
\usage{
storagegateway_delete_gateway(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Deletes a gateway. To specify which gateway to delete, use the Amazon Resource Name (ARN) of the gateway in your request. The operation deletes the gateway; however, it does not delete the gateway virtual machine (VM) from your host computer.

See \url{https://paws-r.github.io/docs/storagegateway/delete_gateway.html} for full documentation.
}
\keyword{internal}
