% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_delete_tape_archive}
\alias{storagegateway_delete_tape_archive}
\title{Deletes the specified virtual tape from the virtual tape shelf (VTS)}
\usage{
storagegateway_delete_tape_archive(TapeARN, BypassGovernanceRetention)
}
\arguments{
\item{TapeARN}{[required] The Amazon Resource Name (ARN) of the virtual tape to delete from the
virtual tape shelf (VTS).}

\item{BypassGovernanceRetention}{Set to \code{TRUE} to delete an archived tape that belongs to a custom pool
with tape retention lock. Only archived tapes with tape retention lock
set to \code{governance} can be deleted. Archived tapes with tape retention
lock set to \code{compliance} can't be deleted.}
}
\description{
Deletes the specified virtual tape from the virtual tape shelf (VTS).
This operation is only supported in the tape gateway type.
}
\section{Request syntax}{
\preformatted{svc$delete_tape_archive(
  TapeARN = "string",
  BypassGovernanceRetention = TRUE|FALSE
)
}
}

\examples{
\dontrun{
# Deletes the specified virtual tape from the virtual tape shelf (VTS).
svc$delete_tape_archive(
  TapeARN = "arn:aws:storagegateway:us-east-1:204469490176:tape/TEST05A2A0"
)
}

}
\keyword{internal}
