% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_delete_tape_pool}
\alias{storagegateway_delete_tape_pool}
\title{Delete a custom tape pool}
\usage{
storagegateway_delete_tape_pool(PoolARN)
}
\arguments{
\item{PoolARN}{[required] The Amazon Resource Name (ARN) of the custom tape pool to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PoolARN = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Delete a custom tape pool. A custom tape pool can only be deleted if
there are no tapes in the pool and if there are no automatic tape
creation policies that reference the custom tape pool.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_tape_pool(
  PoolARN = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
