% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_availability_monitor_test}
\alias{storagegateway_describe_availability_monitor_test}
\title{Returns information about the most recent high availability monitoring
test that was performed on the host in a cluster}
\usage{
storagegateway_describe_availability_monitor_test(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns information about the most recent high availability monitoring test that was performed on the host in a cluster. If a test isn't performed, the status and start time in the response would be null.

See \url{https://paws-r.github.io/docs/storagegateway/describe_availability_monitor_test.html} for full documentation.
}
\keyword{internal}
