% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_cache}
\alias{storagegateway_describe_cache}
\title{Returns information about the cache of a gateway}
\usage{
storagegateway_describe_cache(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns information about the cache of a gateway. This operation is only supported in the cached volume, tape, and file gateway types.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_cache/} for full documentation.
}
\keyword{internal}
