% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_file_system_associations}
\alias{storagegateway_describe_file_system_associations}
\title{Gets the file system association information}
\usage{
storagegateway_describe_file_system_associations(FileSystemAssociationARNList)
}
\arguments{
\item{FileSystemAssociationARNList}{[required] An array containing the Amazon Resource Name (ARN) of each file system
association to be described.}
}
\description{
Gets the file system association information. This operation is only supported for FSx File Gateways.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_file_system_associations/} for full documentation.
}
\keyword{internal}
