% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_maintenance_start_time}
\alias{storagegateway_describe_maintenance_start_time}
\title{Returns your gateway's weekly maintenance start time including the day
and time of the week}
\usage{
storagegateway_describe_maintenance_start_time(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns your gateway's weekly maintenance start time including the day and time of the week. Note that values are in terms of the gateway's time zone.

See \url{https://paws-r.github.io/docs/storagegateway/describe_maintenance_start_time.html} for full documentation.
}
\keyword{internal}
