% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_maintenance_start_time}
\alias{storagegateway_describe_maintenance_start_time}
\title{Returns your gateway's weekly maintenance start time including the day
and time of the week}
\usage{
storagegateway_describe_maintenance_start_time(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GatewayARN = "string",
  HourOfDay = 123,
  MinuteOfHour = 123,
  DayOfWeek = 123,
  DayOfMonth = 123,
  Timezone = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns your gateway's weekly maintenance start time including the day
and time of the week. Note that values are in terms of the gateway's
time zone.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_maintenance_start_time(
  GatewayARN = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns your gateway's weekly maintenance start time including the day
# and time of the week.
svc$describe_maintenance_start_time(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)
}

}
\keyword{internal}
