% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_smb_settings}
\alias{storagegateway_describe_smb_settings}
\title{Gets a description of a Server Message Block (SMB) file share settings
from a file gateway}
\usage{
storagegateway_describe_smb_settings(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GatewayARN = "string",
  DomainName = "string",
  ActiveDirectoryStatus = "ACCESS_DENIED"|"DETACHED"|"JOINED"|"JOINING"|"NETWORK_ERROR"|"TIMEOUT"|"UNKNOWN_ERROR",
  SMBGuestPasswordSet = TRUE|FALSE,
  SMBSecurityStrategy = "ClientSpecified"|"MandatorySigning"|"MandatoryEncryption",
  FileSharesVisible = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Gets a description of a Server Message Block (SMB) file share settings
from a file gateway. This operation is only supported for file gateways.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_smb_settings(
  GatewayARN = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
