% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_smb_settings}
\alias{storagegateway_describe_smb_settings}
\title{Gets a description of a Server Message Block (SMB) file share settings
from a file gateway}
\usage{
storagegateway_describe_smb_settings(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Gets a description of a Server Message Block (SMB) file share settings from a file gateway. This operation is only supported for file gateways.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_smb_settings/} for full documentation.
}
\keyword{internal}
