% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_snapshot_schedule}
\alias{storagegateway_describe_snapshot_schedule}
\title{Describes the snapshot schedule for the specified gateway volume}
\usage{
storagegateway_describe_snapshot_schedule(VolumeARN)
}
\arguments{
\item{VolumeARN}{[required] The Amazon Resource Name (ARN) of the volume. Use the
\code{\link[=storagegateway_list_volumes]{list_volumes}} operation to return a list
of gateway volumes.}
}
\value{
A list with the following syntax:\preformatted{list(
  VolumeARN = "string",
  StartAt = 123,
  RecurrenceInHours = 123,
  Description = "string",
  Timezone = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Describes the snapshot schedule for the specified gateway volume. The
snapshot schedule information includes intervals at which snapshots are
automatically initiated on the volume. This operation is only supported
in the cached volume and stored volume types.
}
\section{Request syntax}{
\preformatted{svc$describe_snapshot_schedule(
  VolumeARN = "string"
)
}
}

\examples{
\dontrun{
# Describes the snapshot schedule for the specified gateway volume
# including intervals at which snapshots are automatically initiated.
svc$describe_snapshot_schedule(
  VolumeARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12..."
)
}

}
\keyword{internal}
