% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_tape_archives}
\alias{storagegateway_describe_tape_archives}
\title{Returns a description of specified virtual tapes in the virtual tape
shelf (VTS)}
\usage{
storagegateway_describe_tape_archives(
  TapeARNs = NULL,
  Marker = NULL,
  Limit = NULL
)
}
\arguments{
\item{TapeARNs}{Specifies one or more unique Amazon Resource Names (ARNs) that represent
the virtual tapes you want to describe.}

\item{Marker}{An opaque string that indicates the position at which to begin
describing virtual tapes.}

\item{Limit}{Specifies that the number of virtual tapes described be limited to the
specified number.}
}
\description{
Returns a description of specified virtual tapes in the virtual tape shelf (VTS). This operation is only supported in the tape gateway type.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_describe_tape_archives/} for full documentation.
}
\keyword{internal}
