% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_tape_recovery_points}
\alias{storagegateway_describe_tape_recovery_points}
\title{Returns a list of virtual tape recovery points that are available for
the specified tape gateway}
\usage{
storagegateway_describe_tape_recovery_points(
  GatewayARN,
  Marker = NULL,
  Limit = NULL
)
}
\arguments{
\item{GatewayARN}{[required]}

\item{Marker}{An opaque string that indicates the position at which to begin
describing the virtual tape recovery points.}

\item{Limit}{Specifies that the number of virtual tape recovery points that are
described be limited to the specified number.}
}
\description{
Returns a list of virtual tape recovery points that are available for the specified tape gateway.

See \url{https://paws-r.github.io/docs/storagegateway/describe_tape_recovery_points.html} for full documentation.
}
\keyword{internal}
