% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_disable_gateway}
\alias{storagegateway_disable_gateway}
\title{Disables a tape gateway when the gateway is no longer functioning}
\usage{
storagegateway_disable_gateway(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Disables a tape gateway when the gateway is no longer functioning. For example, if your gateway VM is damaged, you can disable the gateway so you can recover virtual tapes.

See \url{https://paws-r.github.io/docs/storagegateway/disable_gateway.html} for full documentation.
}
\keyword{internal}
