% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_automatic_tape_creation_policies}
\alias{storagegateway_list_automatic_tape_creation_policies}
\title{Lists the automatic tape creation policies for a gateway}
\usage{
storagegateway_list_automatic_tape_creation_policies(GatewayARN = NULL)
}
\arguments{
\item{GatewayARN}{}
}
\description{
Lists the automatic tape creation policies for a gateway. If there are no automatic tape creation policies for the gateway, it returns an empty list.

See \url{https://paws-r.github.io/docs/storagegateway/list_automatic_tape_creation_policies.html} for full documentation.
}
\keyword{internal}
