% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_local_disks}
\alias{storagegateway_list_local_disks}
\title{Returns a list of the gateway's local disks}
\usage{
storagegateway_list_local_disks(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns a list of the gateway's local disks. To specify which gateway to describe, you use the Amazon Resource Name (ARN) of the gateway in the body of the request.

See \url{https://paws-r.github.io/docs/storagegateway/list_local_disks.html} for full documentation.
}
\keyword{internal}
