% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_tapes}
\alias{storagegateway_list_tapes}
\title{Lists virtual tapes in your virtual tape library (VTL) and your virtual
tape shelf (VTS)}
\usage{
storagegateway_list_tapes(TapeARNs = NULL, Marker = NULL, Limit = NULL)
}
\arguments{
\item{TapeARNs}{}

\item{Marker}{A string that indicates the position at which to begin the returned list
of tapes.}

\item{Limit}{An optional number limit for the tapes in the list returned by this
call.}
}
\description{
Lists virtual tapes in your virtual tape library (VTL) and your virtual tape shelf (VTS). You specify the tapes to list by specifying one or more tape Amazon Resource Names (ARNs). If you don't specify a tape ARN, the operation lists all virtual tapes in both your VTL and VTS.

See \url{https://paws-r.github.io/docs/storagegateway/list_tapes.html} for full documentation.
}
\keyword{internal}
