% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_list_volume_recovery_points}
\alias{storagegateway_list_volume_recovery_points}
\title{Lists the recovery points for a specified gateway}
\usage{
storagegateway_list_volume_recovery_points(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Lists the recovery points for a specified gateway. This operation is only supported in the cached volume gateway type.

See \url{https://paws-r.github.io/docs/storagegateway/list_volume_recovery_points.html} for full documentation.
}
\keyword{internal}
