% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_notify_when_uploaded}
\alias{storagegateway_notify_when_uploaded}
\title{Sends you notification through CloudWatch Events when all files written
to your file share have been uploaded to S3}
\usage{
storagegateway_notify_when_uploaded(FileShareARN)
}
\arguments{
\item{FileShareARN}{[required]}
}
\description{
Sends you notification through CloudWatch Events when all files written to your file share have been uploaded to S3. Amazon S3.

See \url{https://paws-r.github.io/docs/storagegateway/notify_when_uploaded.html} for full documentation.
}
\keyword{internal}
