% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_shutdown_gateway}
\alias{storagegateway_shutdown_gateway}
\title{Shuts down a gateway}
\usage{
storagegateway_shutdown_gateway(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Shuts down a gateway. To specify which gateway to shut down, use the Amazon Resource Name (ARN) of the gateway in the body of your request.

See \url{https://paws-r.github.io/docs/storagegateway/shutdown_gateway.html} for full documentation.
}
\keyword{internal}
