% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_start_gateway}
\alias{storagegateway_start_gateway}
\title{Starts a gateway that you previously shut down (see ShutdownGateway)}
\usage{
storagegateway_start_gateway(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Starts a gateway that you previously shut down (see \code{\link[=storagegateway_shutdown_gateway]{shutdown_gateway}}). After the gateway starts, you can then make other API calls, your applications can read from or write to the gateway's storage volumes and you will be able to take snapshot backups.

See \url{https://paws-r.github.io/docs/storagegateway/start_gateway.html} for full documentation.
}
\keyword{internal}
