% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_file_system_association}
\alias{storagegateway_update_file_system_association}
\title{Updates a file system association}
\usage{
storagegateway_update_file_system_association(
  FileSystemAssociationARN,
  UserName = NULL,
  Password = NULL,
  AuditDestinationARN = NULL,
  CacheAttributes = NULL
)
}
\arguments{
\item{FileSystemAssociationARN}{[required] The Amazon Resource Name (ARN) of the file system association that you
want to update.}

\item{UserName}{The user name of the user credential that has permission to access the
root share D$ of the Amazon FSx file system. The user account must
belong to the Amazon FSx delegated admin user group.}

\item{Password}{The password of the user credential.}

\item{AuditDestinationARN}{The Amazon Resource Name (ARN) of the storage used for the audit logs.}

\item{CacheAttributes}{}
}
\description{
Updates a file system association. This operation is only supported in the FSx File Gateways.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_update_file_system_association/} for full documentation.
}
\keyword{internal}
