% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_gateway_software_now}
\alias{storagegateway_update_gateway_software_now}
\title{Updates the gateway virtual machine (VM) software}
\usage{
storagegateway_update_gateway_software_now(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Updates the gateway virtual machine (VM) software. The request immediately triggers the software update.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_update_gateway_software_now/} for full documentation.
}
\keyword{internal}
