% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_smb_file_share}
\alias{storagegateway_update_smb_file_share}
\title{Updates a Server Message Block (SMB) file share}
\usage{
storagegateway_update_smb_file_share(FileShareARN, KMSEncrypted, KMSKey,
  DefaultStorageClass, ObjectACL, ReadOnly, GuessMIMETypeEnabled,
  RequesterPays, SMBACLEnabled, AccessBasedEnumeration, AdminUserList,
  ValidUserList, InvalidUserList, AuditDestinationARN, CaseSensitivity,
  FileShareName, CacheAttributes, NotificationPolicy)
}
\arguments{
\item{FileShareARN}{[required] The Amazon Resource Name (ARN) of the SMB file share that you want to
update.}

\item{KMSEncrypted}{Set to \code{true} to use Amazon S3 server-side encryption with your own AWS
KMS key, or \code{false} to use a key managed by Amazon S3. Optional.

Valid Values: \code{true} | \code{false}}

\item{KMSKey}{The Amazon Resource Name (ARN) of a symmetric customer master key (CMK)
used for Amazon S3 server-side encryption. Storage Gateway does not
support asymmetric CMKs. This value can only be set when \code{KMSEncrypted}
is \code{true}. Optional.}

\item{DefaultStorageClass}{The default storage class for objects put into an Amazon S3 bucket by
the file gateway. The default value is \code{S3_INTELLIGENT_TIERING}.
Optional.

Valid Values: \code{S3_STANDARD} | \code{S3_INTELLIGENT_TIERING} |
\code{S3_STANDARD_IA} | \code{S3_ONEZONE_IA}}

\item{ObjectACL}{A value that sets the access control list (ACL) permission for objects
in the S3 bucket that a file gateway puts objects into. The default
value is \code{private}.}

\item{ReadOnly}{A value that sets the write status of a file share. Set this value to
\code{true} to set write status to read-only, otherwise set to \code{false}.

Valid Values: \code{true} | \code{false}}

\item{GuessMIMETypeEnabled}{A value that enables guessing of the MIME type for uploaded objects
based on file extensions. Set this value to \code{true} to enable MIME type
guessing, otherwise set to \code{false}. The default value is \code{true}.

Valid Values: \code{true} | \code{false}}

\item{RequesterPays}{A value that sets who pays the cost of the request and the cost
associated with data download from the S3 bucket. If this value is set
to \code{true}, the requester pays the costs; otherwise, the S3 bucket owner
pays. However, the S3 bucket owner always pays the cost of storing data.

\code{RequesterPays} is a configuration for the S3 bucket that backs the file
share, so make sure that the configuration on the file share is the same
as the S3 bucket configuration.

Valid Values: \code{true} | \code{false}}

\item{SMBACLEnabled}{Set this value to \code{true} to enable access control list (ACL) on the SMB
file share. Set it to \code{false} to map file and directory permissions to
the POSIX permissions.

For more information, see \href{https://docs.aws.amazon.com/storagegateway/latest/userguide/smb-acl.html}{Using Microsoft Windows ACLs to control access to an SMB file share}
in the \emph{AWS Storage Gateway User Guide}.

Valid Values: \code{true} | \code{false}}

\item{AccessBasedEnumeration}{The files and folders on this share will only be visible to users with
read access.}

\item{AdminUserList}{A list of users or groups in the Active Directory that have
administrator rights to the file share. A group must be prefixed with
the @ character. Acceptable formats include: \verb{DOMAIN\\User1}, \code{user1},
\verb{@group1}, and \verb{@DOMAIN\\group1}. Can only be set if Authentication is
set to \code{ActiveDirectory}.}

\item{ValidUserList}{A list of users or groups in the Active Directory that are allowed to
access the file share. A group must be prefixed with the @ character.
Acceptable formats include: \verb{DOMAIN\\User1}, \code{user1}, \verb{@group1}, and
\verb{@DOMAIN\\group1}. Can only be set if Authentication is set to
\code{ActiveDirectory}.}

\item{InvalidUserList}{A list of users or groups in the Active Directory that are not allowed
to access the file share. A group must be prefixed with the @ character.
Acceptable formats include: \verb{DOMAIN\\User1}, \code{user1}, \verb{@group1}, and
\verb{@DOMAIN\\group1}. Can only be set if Authentication is set to
\code{ActiveDirectory}.}

\item{AuditDestinationARN}{The Amazon Resource Name (ARN) of the storage used for the audit logs.}

\item{CaseSensitivity}{The case of an object name in an Amazon S3 bucket. For
\code{ClientSpecified}, the client determines the case sensitivity. For
\code{CaseSensitive}, the gateway determines the case sensitivity. The
default value is \code{ClientSpecified}.}

\item{FileShareName}{The name of the file share. Optional.

\code{FileShareName} must be set if an S3 prefix name is set in
\code{LocationARN}.}

\item{CacheAttributes}{Refresh cache information.}

\item{NotificationPolicy}{The notification policy of the file share.}
}
\value{
A list with the following syntax:\preformatted{list(
  FileShareARN = "string"
)
}
}
\description{
Updates a Server Message Block (SMB) file share. This operation is only
supported for file gateways.

To leave a file share field unchanged, set the corresponding input field
to null.

File gateways require AWS Security Token Service (AWS STS) to be
activated to enable you to create a file share. Make sure that AWS STS
is activated in the AWS Region you are creating your file gateway in. If
AWS STS is not activated in this AWS Region, activate it. For
information about how to activate AWS STS, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp_enable-regions.html}{Activating and deactivating AWS STS in an AWS Region}
in the \emph{AWS Identity and Access Management User Guide}.

File gateways don't support creating hard or symbolic links on a file
share.
}
\section{Request syntax}{
\preformatted{svc$update_smb_file_share(
  FileShareARN = "string",
  KMSEncrypted = TRUE|FALSE,
  KMSKey = "string",
  DefaultStorageClass = "string",
  ObjectACL = "private"|"public-read"|"public-read-write"|"authenticated-read"|"bucket-owner-read"|"bucket-owner-full-control"|"aws-exec-read",
  ReadOnly = TRUE|FALSE,
  GuessMIMETypeEnabled = TRUE|FALSE,
  RequesterPays = TRUE|FALSE,
  SMBACLEnabled = TRUE|FALSE,
  AccessBasedEnumeration = TRUE|FALSE,
  AdminUserList = list(
    "string"
  ),
  ValidUserList = list(
    "string"
  ),
  InvalidUserList = list(
    "string"
  ),
  AuditDestinationARN = "string",
  CaseSensitivity = "ClientSpecified"|"CaseSensitive",
  FileShareName = "string",
  CacheAttributes = list(
    CacheStaleTimeoutInSeconds = 123
  ),
  NotificationPolicy = "string"
)
}
}

\keyword{internal}
