% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_update_smb_file_share_visibility}
\alias{storagegateway_update_smb_file_share_visibility}
\title{Controls whether the shares on an S3 File Gateway are visible in a net
view or browse list}
\usage{
storagegateway_update_smb_file_share_visibility(GatewayARN, FileSharesVisible)
}
\arguments{
\item{GatewayARN}{[required]}

\item{FileSharesVisible}{[required] The shares on this gateway appear when listing shares.}
}
\description{
Controls whether the shares on an S3 File Gateway are visible in a net view or browse list. The operation is only supported for S3 File Gateways.

See \url{https://www.paws-r-sdk.com/docs/storagegateway_update_smb_file_share_visibility/} for full documentation.
}
\keyword{internal}
